fix: fixed delete-project request and added the delete modal
