+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_material_button.js
@@ -4,7 +4,7 @@ import {
   MatRippleLoader,
   MatRippleModule,
   Platform
-} from "./chunk-FLGVYJXI.js";
+} from "./chunk-7ZFYYFG2.js";
 import "./chunk-7GFUNZUG.js";
 import {
   ANIMATION_MODULE_TYPE,
