+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_material_dialog.js
@@ -1,0 +1,1779 @@
+import {
+  BasePortalOutlet,
+  CdkPortalOutlet,
+  ComponentPortal,
+  Overlay,
+  OverlayConfig,
+  OverlayContainer,
+  OverlayModule,
+  OverlayRef,
+  PortalModule,
+  TemplatePortal
+} from "./chunk-RM7V6FV7.js";
+import {
+  animate,
+  animateChild,
+  group,
+  query,
+  state,
+  style,
+  transition,
+  trigger
+} from "./chunk-VNW2MBA7.js";
+import "./chunk-5H5WHFDB.js";
+import {
+  A11yModule,
+  Directionality,
+  ESCAPE,
+  FocusMonitor,
+  FocusTrapFactory,
+  InteractivityChecker,
+  MatCommonModule,
+  Platform,
+  _getFocusedElementPierceShadowDom,
+  coerceNumberProperty,
+  hasModifierKey
+} from "./chunk-7ZFYYFG2.js";
+import {
+  DOCUMENT,
+  Location
+} from "./chunk-7GFUNZUG.js";
+import {
+  ANIMATION_MODULE_TYPE,
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  Directive,
+  ElementRef,
+  EventEmitter,
+  Inject,
+  Injectable,
+  InjectionToken,
+  Injector,
+  Input,
+  InputFlags,
+  NgModule,
+  NgZone,
+  Optional,
+  SkipSelf,
+  TemplateRef,
+  ViewChild,
+  ViewEncapsulation$1,
+  inject,
+  setClassMetadata,
+  ɵɵInheritDefinitionFeature,
+  ɵɵNgOnChangesFeature,
+  ɵɵStandaloneFeature,
+  ɵɵattribute,
+  ɵɵclassProp,
+  ɵɵdefineComponent,
+  ɵɵdefineDirective,
+  ɵɵdefineInjectable,
+  ɵɵdefineInjector,
+  ɵɵdefineNgModule,
+  ɵɵdirectiveInject,
+  ɵɵelementEnd,
+  ɵɵelementStart,
+  ɵɵgetInheritedFactory,
+  ɵɵhostProperty,
+  ɵɵinject,
+  ɵɵlistener,
+  ɵɵloadQuery,
+  ɵɵqueryRefresh,
+  ɵɵtemplate,
+  ɵɵviewQuery
+} from "./chunk-RUPN7I6S.js";
+import {
+  Subject,
+  defer,
+  filter,
+  merge,
+  of,
+  startWith,
+  take
+} from "./chunk-4J25ECOH.js";
+import {
+  __spreadProps,
+  __spreadValues
+} from "./chunk-WKYGNSYM.js";
+
+// node_modules/@angular/cdk/fesm2022/dialog.mjs
+function CdkDialogContainer_ng_template_0_Template(rf, ctx) {
+}
+var DialogConfig = class {
+  constructor() {
+    this.role = "dialog";
+    this.panelClass = "";
+    this.hasBackdrop = true;
+    this.backdropClass = "";
+    this.disableClose = false;
+    this.width = "";
+    this.height = "";
+    this.data = null;
+    this.ariaDescribedBy = null;
+    this.ariaLabelledBy = null;
+    this.ariaLabel = null;
+    this.ariaModal = true;
+    this.autoFocus = "first-tabbable";
+    this.restoreFocus = true;
+    this.closeOnNavigation = true;
+    this.closeOnDestroy = true;
+    this.closeOnOverlayDetachments = true;
+  }
+};
+function throwDialogContentAlreadyAttachedError() {
+  throw Error("Attempting to attach dialog content after content is already attached");
+}
+var _CdkDialogContainer = class _CdkDialogContainer extends BasePortalOutlet {
+  constructor(_elementRef, _focusTrapFactory, _document, _config, _interactivityChecker, _ngZone, _overlayRef, _focusMonitor) {
+    super();
+    this._elementRef = _elementRef;
+    this._focusTrapFactory = _focusTrapFactory;
+    this._config = _config;
+    this._interactivityChecker = _interactivityChecker;
+    this._ngZone = _ngZone;
+    this._overlayRef = _overlayRef;
+    this._focusMonitor = _focusMonitor;
+    this._platform = inject(Platform);
+    this._focusTrap = null;
+    this._elementFocusedBeforeDialogWasOpened = null;
+    this._closeInteractionType = null;
+    this._ariaLabelledByQueue = [];
+    this._changeDetectorRef = inject(ChangeDetectorRef);
+    this.attachDomPortal = (portal) => {
+      if (this._portalOutlet.hasAttached() && (typeof ngDevMode === "undefined" || ngDevMode)) {
+        throwDialogContentAlreadyAttachedError();
+      }
+      const result = this._portalOutlet.attachDomPortal(portal);
+      this._contentAttached();
+      return result;
+    };
+    this._document = _document;
+    if (this._config.ariaLabelledBy) {
+      this._ariaLabelledByQueue.push(this._config.ariaLabelledBy);
+    }
+  }
+  _addAriaLabelledBy(id) {
+    this._ariaLabelledByQueue.push(id);
+    this._changeDetectorRef.markForCheck();
+  }
+  _removeAriaLabelledBy(id) {
+    const index = this._ariaLabelledByQueue.indexOf(id);
+    if (index > -1) {
+      this._ariaLabelledByQueue.splice(index, 1);
+      this._changeDetectorRef.markForCheck();
+    }
+  }
+  _contentAttached() {
+    this._initializeFocusTrap();
+    this._handleBackdropClicks();
+    this._captureInitialFocus();
+  }
+  /**
+   * Can be used by child classes to customize the initial focus
+   * capturing behavior (e.g. if it's tied to an animation).
+   */
+  _captureInitialFocus() {
+    this._trapFocus();
+  }
+  ngOnDestroy() {
+    this._restoreFocus();
+  }
+  /**
+   * Attach a ComponentPortal as content to this dialog container.
+   * @param portal Portal to be attached as the dialog content.
+   */
+  attachComponentPortal(portal) {
+    if (this._portalOutlet.hasAttached() && (typeof ngDevMode === "undefined" || ngDevMode)) {
+      throwDialogContentAlreadyAttachedError();
+    }
+    const result = this._portalOutlet.attachComponentPortal(portal);
+    this._contentAttached();
+    return result;
+  }
+  /**
+   * Attach a TemplatePortal as content to this dialog container.
+   * @param portal Portal to be attached as the dialog content.
+   */
+  attachTemplatePortal(portal) {
+    if (this._portalOutlet.hasAttached() && (typeof ngDevMode === "undefined" || ngDevMode)) {
+      throwDialogContentAlreadyAttachedError();
+    }
+    const result = this._portalOutlet.attachTemplatePortal(portal);
+    this._contentAttached();
+    return result;
+  }
+  // TODO(crisbeto): this shouldn't be exposed, but there are internal references to it.
+  /** Captures focus if it isn't already inside the dialog. */
+  _recaptureFocus() {
+    if (!this._containsFocus()) {
+      this._trapFocus();
+    }
+  }
+  /**
+   * Focuses the provided element. If the element is not focusable, it will add a tabIndex
+   * attribute to forcefully focus it. The attribute is removed after focus is moved.
+   * @param element The element to focus.
+   */
+  _forceFocus(element, options) {
+    if (!this._interactivityChecker.isFocusable(element)) {
+      element.tabIndex = -1;
+      this._ngZone.runOutsideAngular(() => {
+        const callback = () => {
+          element.removeEventListener("blur", callback);
+          element.removeEventListener("mousedown", callback);
+          element.removeAttribute("tabindex");
+        };
+        element.addEventListener("blur", callback);
+        element.addEventListener("mousedown", callback);
+      });
+    }
+    element.focus(options);
+  }
+  /**
+   * Focuses the first element that matches the given selector within the focus trap.
+   * @param selector The CSS selector for the element to set focus to.
+   */
+  _focusByCssSelector(selector, options) {
+    let elementToFocus = this._elementRef.nativeElement.querySelector(selector);
+    if (elementToFocus) {
+      this._forceFocus(elementToFocus, options);
+    }
+  }
+  /**
+   * Moves the focus inside the focus trap. When autoFocus is not set to 'dialog', if focus
+   * cannot be moved then focus will go to the dialog container.
+   */
+  _trapFocus() {
+    const element = this._elementRef.nativeElement;
+    switch (this._config.autoFocus) {
+      case false:
+      case "dialog":
+        if (!this._containsFocus()) {
+          element.focus();
+        }
+        break;
+      case true:
+      case "first-tabbable":
+        this._focusTrap?.focusInitialElementWhenReady().then((focusedSuccessfully) => {
+          if (!focusedSuccessfully) {
+            this._focusDialogContainer();
+          }
+        });
+        break;
+      case "first-heading":
+        this._focusByCssSelector('h1, h2, h3, h4, h5, h6, [role="heading"]');
+        break;
+      default:
+        this._focusByCssSelector(this._config.autoFocus);
+        break;
+    }
+  }
+  /** Restores focus to the element that was focused before the dialog opened. */
+  _restoreFocus() {
+    const focusConfig = this._config.restoreFocus;
+    let focusTargetElement = null;
+    if (typeof focusConfig === "string") {
+      focusTargetElement = this._document.querySelector(focusConfig);
+    } else if (typeof focusConfig === "boolean") {
+      focusTargetElement = focusConfig ? this._elementFocusedBeforeDialogWasOpened : null;
+    } else if (focusConfig) {
+      focusTargetElement = focusConfig;
+    }
+    if (this._config.restoreFocus && focusTargetElement && typeof focusTargetElement.focus === "function") {
+      const activeElement = _getFocusedElementPierceShadowDom();
+      const element = this._elementRef.nativeElement;
+      if (!activeElement || activeElement === this._document.body || activeElement === element || element.contains(activeElement)) {
+        if (this._focusMonitor) {
+          this._focusMonitor.focusVia(focusTargetElement, this._closeInteractionType);
+          this._closeInteractionType = null;
+        } else {
+          focusTargetElement.focus();
+        }
+      }
+    }
+    if (this._focusTrap) {
+      this._focusTrap.destroy();
+    }
+  }
+  /** Focuses the dialog container. */
+  _focusDialogContainer() {
+    if (this._elementRef.nativeElement.focus) {
+      this._elementRef.nativeElement.focus();
+    }
+  }
+  /** Returns whether focus is inside the dialog. */
+  _containsFocus() {
+    const element = this._elementRef.nativeElement;
+    const activeElement = _getFocusedElementPierceShadowDom();
+    return element === activeElement || element.contains(activeElement);
+  }
+  /** Sets up the focus trap. */
+  _initializeFocusTrap() {
+    if (this._platform.isBrowser) {
+      this._focusTrap = this._focusTrapFactory.create(this._elementRef.nativeElement);
+      if (this._document) {
+        this._elementFocusedBeforeDialogWasOpened = _getFocusedElementPierceShadowDom();
+      }
+    }
+  }
+  /** Sets up the listener that handles clicks on the dialog backdrop. */
+  _handleBackdropClicks() {
+    this._overlayRef.backdropClick().subscribe(() => {
+      if (this._config.disableClose) {
+        this._recaptureFocus();
+      }
+    });
+  }
+};
+_CdkDialogContainer.ɵfac = function CdkDialogContainer_Factory(t) {
+  return new (t || _CdkDialogContainer)(ɵɵdirectiveInject(ElementRef), ɵɵdirectiveInject(FocusTrapFactory), ɵɵdirectiveInject(DOCUMENT, 8), ɵɵdirectiveInject(DialogConfig), ɵɵdirectiveInject(InteractivityChecker), ɵɵdirectiveInject(NgZone), ɵɵdirectiveInject(OverlayRef), ɵɵdirectiveInject(FocusMonitor));
+};
+_CdkDialogContainer.ɵcmp = ɵɵdefineComponent({
+  type: _CdkDialogContainer,
+  selectors: [["cdk-dialog-container"]],
+  viewQuery: function CdkDialogContainer_Query(rf, ctx) {
+    if (rf & 1) {
+      ɵɵviewQuery(CdkPortalOutlet, 7);
+    }
+    if (rf & 2) {
+      let _t;
+      ɵɵqueryRefresh(_t = ɵɵloadQuery()) && (ctx._portalOutlet = _t.first);
+    }
+  },
+  hostAttrs: ["tabindex", "-1", 1, "cdk-dialog-container"],
+  hostVars: 6,
+  hostBindings: function CdkDialogContainer_HostBindings(rf, ctx) {
+    if (rf & 2) {
+      ɵɵattribute("id", ctx._config.id || null)("role", ctx._config.role)("aria-modal", ctx._config.ariaModal)("aria-labelledby", ctx._config.ariaLabel ? null : ctx._ariaLabelledByQueue[0])("aria-label", ctx._config.ariaLabel)("aria-describedby", ctx._config.ariaDescribedBy || null);
+    }
+  },
+  standalone: true,
+  features: [ɵɵInheritDefinitionFeature, ɵɵStandaloneFeature],
+  decls: 1,
+  vars: 0,
+  consts: [["cdkPortalOutlet", ""]],
+  template: function CdkDialogContainer_Template(rf, ctx) {
+    if (rf & 1) {
+      ɵɵtemplate(0, CdkDialogContainer_ng_template_0_Template, 0, 0, "ng-template", 0);
+    }
+  },
+  dependencies: [CdkPortalOutlet],
+  styles: [".cdk-dialog-container{display:block;width:100%;height:100%;min-height:inherit;max-height:inherit}"],
+  encapsulation: 2
+});
+var CdkDialogContainer = _CdkDialogContainer;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(CdkDialogContainer, [{
+    type: Component,
+    args: [{
+      selector: "cdk-dialog-container",
+      encapsulation: ViewEncapsulation$1.None,
+      changeDetection: ChangeDetectionStrategy.Default,
+      standalone: true,
+      imports: [CdkPortalOutlet],
+      host: {
+        "class": "cdk-dialog-container",
+        "tabindex": "-1",
+        "[attr.id]": "_config.id || null",
+        "[attr.role]": "_config.role",
+        "[attr.aria-modal]": "_config.ariaModal",
+        "[attr.aria-labelledby]": "_config.ariaLabel ? null : _ariaLabelledByQueue[0]",
+        "[attr.aria-label]": "_config.ariaLabel",
+        "[attr.aria-describedby]": "_config.ariaDescribedBy || null"
+      },
+      template: "<ng-template cdkPortalOutlet />\n",
+      styles: [".cdk-dialog-container{display:block;width:100%;height:100%;min-height:inherit;max-height:inherit}"]
+    }]
+  }], () => [{
+    type: ElementRef
+  }, {
+    type: FocusTrapFactory
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DialogConfig]
+    }]
+  }, {
+    type: InteractivityChecker
+  }, {
+    type: NgZone
+  }, {
+    type: OverlayRef
+  }, {
+    type: FocusMonitor
+  }], {
+    _portalOutlet: [{
+      type: ViewChild,
+      args: [CdkPortalOutlet, {
+        static: true
+      }]
+    }]
+  });
+})();
+var DialogRef = class {
+  constructor(overlayRef, config) {
+    this.overlayRef = overlayRef;
+    this.config = config;
+    this.closed = new Subject();
+    this.disableClose = config.disableClose;
+    this.backdropClick = overlayRef.backdropClick();
+    this.keydownEvents = overlayRef.keydownEvents();
+    this.outsidePointerEvents = overlayRef.outsidePointerEvents();
+    this.id = config.id;
+    this.keydownEvents.subscribe((event) => {
+      if (event.keyCode === ESCAPE && !this.disableClose && !hasModifierKey(event)) {
+        event.preventDefault();
+        this.close(void 0, {
+          focusOrigin: "keyboard"
+        });
+      }
+    });
+    this.backdropClick.subscribe(() => {
+      if (!this.disableClose) {
+        this.close(void 0, {
+          focusOrigin: "mouse"
+        });
+      }
+    });
+    this._detachSubscription = overlayRef.detachments().subscribe(() => {
+      if (config.closeOnOverlayDetachments !== false) {
+        this.close();
+      }
+    });
+  }
+  /**
+   * Close the dialog.
+   * @param result Optional result to return to the dialog opener.
+   * @param options Additional options to customize the closing behavior.
+   */
+  close(result, options) {
+    if (this.containerInstance) {
+      const closedSubject = this.closed;
+      this.containerInstance._closeInteractionType = options?.focusOrigin || "program";
+      this._detachSubscription.unsubscribe();
+      this.overlayRef.dispose();
+      closedSubject.next(result);
+      closedSubject.complete();
+      this.componentInstance = this.containerInstance = null;
+    }
+  }
+  /** Updates the position of the dialog based on the current position strategy. */
+  updatePosition() {
+    this.overlayRef.updatePosition();
+    return this;
+  }
+  /**
+   * Updates the dialog's width and height.
+   * @param width New width of the dialog.
+   * @param height New height of the dialog.
+   */
+  updateSize(width = "", height = "") {
+    this.overlayRef.updateSize({
+      width,
+      height
+    });
+    return this;
+  }
+  /** Add a CSS class or an array of classes to the overlay pane. */
+  addPanelClass(classes) {
+    this.overlayRef.addPanelClass(classes);
+    return this;
+  }
+  /** Remove a CSS class or an array of classes from the overlay pane. */
+  removePanelClass(classes) {
+    this.overlayRef.removePanelClass(classes);
+    return this;
+  }
+};
+var DIALOG_SCROLL_STRATEGY = new InjectionToken("DialogScrollStrategy", {
+  providedIn: "root",
+  factory: () => {
+    const overlay = inject(Overlay);
+    return () => overlay.scrollStrategies.block();
+  }
+});
+var DIALOG_DATA = new InjectionToken("DialogData");
+var DEFAULT_DIALOG_CONFIG = new InjectionToken("DefaultDialogConfig");
+var uniqueId = 0;
+var _Dialog = class _Dialog {
+  /** Keeps track of the currently-open dialogs. */
+  get openDialogs() {
+    return this._parentDialog ? this._parentDialog.openDialogs : this._openDialogsAtThisLevel;
+  }
+  /** Stream that emits when a dialog has been opened. */
+  get afterOpened() {
+    return this._parentDialog ? this._parentDialog.afterOpened : this._afterOpenedAtThisLevel;
+  }
+  constructor(_overlay, _injector, _defaultOptions, _parentDialog, _overlayContainer, scrollStrategy) {
+    this._overlay = _overlay;
+    this._injector = _injector;
+    this._defaultOptions = _defaultOptions;
+    this._parentDialog = _parentDialog;
+    this._overlayContainer = _overlayContainer;
+    this._openDialogsAtThisLevel = [];
+    this._afterAllClosedAtThisLevel = new Subject();
+    this._afterOpenedAtThisLevel = new Subject();
+    this._ariaHiddenElements = /* @__PURE__ */ new Map();
+    this.afterAllClosed = defer(() => this.openDialogs.length ? this._getAfterAllClosed() : this._getAfterAllClosed().pipe(startWith(void 0)));
+    this._scrollStrategy = scrollStrategy;
+  }
+  open(componentOrTemplateRef, config) {
+    const defaults = this._defaultOptions || new DialogConfig();
+    config = __spreadValues(__spreadValues({}, defaults), config);
+    config.id = config.id || `cdk-dialog-${uniqueId++}`;
+    if (config.id && this.getDialogById(config.id) && (typeof ngDevMode === "undefined" || ngDevMode)) {
+      throw Error(`Dialog with id "${config.id}" exists already. The dialog id must be unique.`);
+    }
+    const overlayConfig = this._getOverlayConfig(config);
+    const overlayRef = this._overlay.create(overlayConfig);
+    const dialogRef = new DialogRef(overlayRef, config);
+    const dialogContainer = this._attachContainer(overlayRef, dialogRef, config);
+    dialogRef.containerInstance = dialogContainer;
+    this._attachDialogContent(componentOrTemplateRef, dialogRef, dialogContainer, config);
+    if (!this.openDialogs.length) {
+      this._hideNonDialogContentFromAssistiveTechnology();
+    }
+    this.openDialogs.push(dialogRef);
+    dialogRef.closed.subscribe(() => this._removeOpenDialog(dialogRef, true));
+    this.afterOpened.next(dialogRef);
+    return dialogRef;
+  }
+  /**
+   * Closes all of the currently-open dialogs.
+   */
+  closeAll() {
+    reverseForEach(this.openDialogs, (dialog) => dialog.close());
+  }
+  /**
+   * Finds an open dialog by its id.
+   * @param id ID to use when looking up the dialog.
+   */
+  getDialogById(id) {
+    return this.openDialogs.find((dialog) => dialog.id === id);
+  }
+  ngOnDestroy() {
+    reverseForEach(this._openDialogsAtThisLevel, (dialog) => {
+      if (dialog.config.closeOnDestroy === false) {
+        this._removeOpenDialog(dialog, false);
+      }
+    });
+    reverseForEach(this._openDialogsAtThisLevel, (dialog) => dialog.close());
+    this._afterAllClosedAtThisLevel.complete();
+    this._afterOpenedAtThisLevel.complete();
+    this._openDialogsAtThisLevel = [];
+  }
+  /**
+   * Creates an overlay config from a dialog config.
+   * @param config The dialog configuration.
+   * @returns The overlay configuration.
+   */
+  _getOverlayConfig(config) {
+    const state2 = new OverlayConfig({
+      positionStrategy: config.positionStrategy || this._overlay.position().global().centerHorizontally().centerVertically(),
+      scrollStrategy: config.scrollStrategy || this._scrollStrategy(),
+      panelClass: config.panelClass,
+      hasBackdrop: config.hasBackdrop,
+      direction: config.direction,
+      minWidth: config.minWidth,
+      minHeight: config.minHeight,
+      maxWidth: config.maxWidth,
+      maxHeight: config.maxHeight,
+      width: config.width,
+      height: config.height,
+      disposeOnNavigation: config.closeOnNavigation
+    });
+    if (config.backdropClass) {
+      state2.backdropClass = config.backdropClass;
+    }
+    return state2;
+  }
+  /**
+   * Attaches a dialog container to a dialog's already-created overlay.
+   * @param overlay Reference to the dialog's underlying overlay.
+   * @param config The dialog configuration.
+   * @returns A promise resolving to a ComponentRef for the attached container.
+   */
+  _attachContainer(overlay, dialogRef, config) {
+    const userInjector = config.injector || config.viewContainerRef?.injector;
+    const providers = [{
+      provide: DialogConfig,
+      useValue: config
+    }, {
+      provide: DialogRef,
+      useValue: dialogRef
+    }, {
+      provide: OverlayRef,
+      useValue: overlay
+    }];
+    let containerType;
+    if (config.container) {
+      if (typeof config.container === "function") {
+        containerType = config.container;
+      } else {
+        containerType = config.container.type;
+        providers.push(...config.container.providers(config));
+      }
+    } else {
+      containerType = CdkDialogContainer;
+    }
+    const containerPortal = new ComponentPortal(containerType, config.viewContainerRef, Injector.create({
+      parent: userInjector || this._injector,
+      providers
+    }), config.componentFactoryResolver);
+    const containerRef = overlay.attach(containerPortal);
+    return containerRef.instance;
+  }
+  /**
+   * Attaches the user-provided component to the already-created dialog container.
+   * @param componentOrTemplateRef The type of component being loaded into the dialog,
+   *     or a TemplateRef to instantiate as the content.
+   * @param dialogRef Reference to the dialog being opened.
+   * @param dialogContainer Component that is going to wrap the dialog content.
+   * @param config Configuration used to open the dialog.
+   */
+  _attachDialogContent(componentOrTemplateRef, dialogRef, dialogContainer, config) {
+    if (componentOrTemplateRef instanceof TemplateRef) {
+      const injector = this._createInjector(config, dialogRef, dialogContainer, void 0);
+      let context = {
+        $implicit: config.data,
+        dialogRef
+      };
+      if (config.templateContext) {
+        context = __spreadValues(__spreadValues({}, context), typeof config.templateContext === "function" ? config.templateContext() : config.templateContext);
+      }
+      dialogContainer.attachTemplatePortal(new TemplatePortal(componentOrTemplateRef, null, context, injector));
+    } else {
+      const injector = this._createInjector(config, dialogRef, dialogContainer, this._injector);
+      const contentRef = dialogContainer.attachComponentPortal(new ComponentPortal(componentOrTemplateRef, config.viewContainerRef, injector, config.componentFactoryResolver));
+      dialogRef.componentRef = contentRef;
+      dialogRef.componentInstance = contentRef.instance;
+    }
+  }
+  /**
+   * Creates a custom injector to be used inside the dialog. This allows a component loaded inside
+   * of a dialog to close itself and, optionally, to return a value.
+   * @param config Config object that is used to construct the dialog.
+   * @param dialogRef Reference to the dialog being opened.
+   * @param dialogContainer Component that is going to wrap the dialog content.
+   * @param fallbackInjector Injector to use as a fallback when a lookup fails in the custom
+   * dialog injector, if the user didn't provide a custom one.
+   * @returns The custom injector that can be used inside the dialog.
+   */
+  _createInjector(config, dialogRef, dialogContainer, fallbackInjector) {
+    const userInjector = config.injector || config.viewContainerRef?.injector;
+    const providers = [{
+      provide: DIALOG_DATA,
+      useValue: config.data
+    }, {
+      provide: DialogRef,
+      useValue: dialogRef
+    }];
+    if (config.providers) {
+      if (typeof config.providers === "function") {
+        providers.push(...config.providers(dialogRef, config, dialogContainer));
+      } else {
+        providers.push(...config.providers);
+      }
+    }
+    if (config.direction && (!userInjector || !userInjector.get(Directionality, null, {
+      optional: true
+    }))) {
+      providers.push({
+        provide: Directionality,
+        useValue: {
+          value: config.direction,
+          change: of()
+        }
+      });
+    }
+    return Injector.create({
+      parent: userInjector || fallbackInjector,
+      providers
+    });
+  }
+  /**
+   * Removes a dialog from the array of open dialogs.
+   * @param dialogRef Dialog to be removed.
+   * @param emitEvent Whether to emit an event if this is the last dialog.
+   */
+  _removeOpenDialog(dialogRef, emitEvent) {
+    const index = this.openDialogs.indexOf(dialogRef);
+    if (index > -1) {
+      this.openDialogs.splice(index, 1);
+      if (!this.openDialogs.length) {
+        this._ariaHiddenElements.forEach((previousValue, element) => {
+          if (previousValue) {
+            element.setAttribute("aria-hidden", previousValue);
+          } else {
+            element.removeAttribute("aria-hidden");
+          }
+        });
+        this._ariaHiddenElements.clear();
+        if (emitEvent) {
+          this._getAfterAllClosed().next();
+        }
+      }
+    }
+  }
+  /** Hides all of the content that isn't an overlay from assistive technology. */
+  _hideNonDialogContentFromAssistiveTechnology() {
+    const overlayContainer = this._overlayContainer.getContainerElement();
+    if (overlayContainer.parentElement) {
+      const siblings = overlayContainer.parentElement.children;
+      for (let i = siblings.length - 1; i > -1; i--) {
+        const sibling = siblings[i];
+        if (sibling !== overlayContainer && sibling.nodeName !== "SCRIPT" && sibling.nodeName !== "STYLE" && !sibling.hasAttribute("aria-live")) {
+          this._ariaHiddenElements.set(sibling, sibling.getAttribute("aria-hidden"));
+          sibling.setAttribute("aria-hidden", "true");
+        }
+      }
+    }
+  }
+  _getAfterAllClosed() {
+    const parent = this._parentDialog;
+    return parent ? parent._getAfterAllClosed() : this._afterAllClosedAtThisLevel;
+  }
+};
+_Dialog.ɵfac = function Dialog_Factory(t) {
+  return new (t || _Dialog)(ɵɵinject(Overlay), ɵɵinject(Injector), ɵɵinject(DEFAULT_DIALOG_CONFIG, 8), ɵɵinject(_Dialog, 12), ɵɵinject(OverlayContainer), ɵɵinject(DIALOG_SCROLL_STRATEGY));
+};
+_Dialog.ɵprov = ɵɵdefineInjectable({
+  token: _Dialog,
+  factory: _Dialog.ɵfac,
+  providedIn: "root"
+});
+var Dialog = _Dialog;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Dialog, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: Overlay
+  }, {
+    type: Injector
+  }, {
+    type: DialogConfig,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [DEFAULT_DIALOG_CONFIG]
+    }]
+  }, {
+    type: Dialog,
+    decorators: [{
+      type: Optional
+    }, {
+      type: SkipSelf
+    }]
+  }, {
+    type: OverlayContainer
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DIALOG_SCROLL_STRATEGY]
+    }]
+  }], null);
+})();
+function reverseForEach(items, callback) {
+  let i = items.length;
+  while (i--) {
+    callback(items[i]);
+  }
+}
+var _DialogModule = class _DialogModule {
+};
+_DialogModule.ɵfac = function DialogModule_Factory(t) {
+  return new (t || _DialogModule)();
+};
+_DialogModule.ɵmod = ɵɵdefineNgModule({
+  type: _DialogModule,
+  imports: [OverlayModule, PortalModule, A11yModule, CdkDialogContainer],
+  exports: [
+    // Re-export the PortalModule so that people extending the `CdkDialogContainer`
+    // don't have to remember to import it or be faced with an unhelpful error.
+    PortalModule,
+    CdkDialogContainer
+  ]
+});
+_DialogModule.ɵinj = ɵɵdefineInjector({
+  providers: [Dialog],
+  imports: [
+    OverlayModule,
+    PortalModule,
+    A11yModule,
+    // Re-export the PortalModule so that people extending the `CdkDialogContainer`
+    // don't have to remember to import it or be faced with an unhelpful error.
+    PortalModule
+  ]
+});
+var DialogModule = _DialogModule;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(DialogModule, [{
+    type: NgModule,
+    args: [{
+      imports: [OverlayModule, PortalModule, A11yModule, CdkDialogContainer],
+      exports: [
+        // Re-export the PortalModule so that people extending the `CdkDialogContainer`
+        // don't have to remember to import it or be faced with an unhelpful error.
+        PortalModule,
+        CdkDialogContainer
+      ],
+      providers: [Dialog]
+    }]
+  }], null, null);
+})();
+
+// node_modules/@angular/material/fesm2022/dialog.mjs
+function MatDialogContainer_ng_template_2_Template(rf, ctx) {
+}
+var MatDialogConfig = class {
+  constructor() {
+    this.role = "dialog";
+    this.panelClass = "";
+    this.hasBackdrop = true;
+    this.backdropClass = "";
+    this.disableClose = false;
+    this.width = "";
+    this.height = "";
+    this.data = null;
+    this.ariaDescribedBy = null;
+    this.ariaLabelledBy = null;
+    this.ariaLabel = null;
+    this.ariaModal = true;
+    this.autoFocus = "first-tabbable";
+    this.restoreFocus = true;
+    this.delayFocusTrap = true;
+    this.closeOnNavigation = true;
+  }
+};
+var OPEN_CLASS = "mdc-dialog--open";
+var OPENING_CLASS = "mdc-dialog--opening";
+var CLOSING_CLASS = "mdc-dialog--closing";
+var OPEN_ANIMATION_DURATION = 150;
+var CLOSE_ANIMATION_DURATION = 75;
+var _MatDialogContainer = class _MatDialogContainer extends CdkDialogContainer {
+  constructor(elementRef, focusTrapFactory, _document, dialogConfig, interactivityChecker, ngZone, overlayRef, _animationMode, focusMonitor) {
+    super(elementRef, focusTrapFactory, _document, dialogConfig, interactivityChecker, ngZone, overlayRef, focusMonitor);
+    this._animationMode = _animationMode;
+    this._animationStateChanged = new EventEmitter();
+    this._animationsEnabled = this._animationMode !== "NoopAnimations";
+    this._actionSectionCount = 0;
+    this._hostElement = this._elementRef.nativeElement;
+    this._enterAnimationDuration = this._animationsEnabled ? parseCssTime(this._config.enterAnimationDuration) ?? OPEN_ANIMATION_DURATION : 0;
+    this._exitAnimationDuration = this._animationsEnabled ? parseCssTime(this._config.exitAnimationDuration) ?? CLOSE_ANIMATION_DURATION : 0;
+    this._animationTimer = null;
+    this._finishDialogOpen = () => {
+      this._clearAnimationClasses();
+      this._openAnimationDone(this._enterAnimationDuration);
+    };
+    this._finishDialogClose = () => {
+      this._clearAnimationClasses();
+      this._animationStateChanged.emit({
+        state: "closed",
+        totalTime: this._exitAnimationDuration
+      });
+    };
+  }
+  _contentAttached() {
+    super._contentAttached();
+    this._startOpenAnimation();
+  }
+  /** Starts the dialog open animation if enabled. */
+  _startOpenAnimation() {
+    this._animationStateChanged.emit({
+      state: "opening",
+      totalTime: this._enterAnimationDuration
+    });
+    if (this._animationsEnabled) {
+      this._hostElement.style.setProperty(TRANSITION_DURATION_PROPERTY, `${this._enterAnimationDuration}ms`);
+      this._requestAnimationFrame(() => this._hostElement.classList.add(OPENING_CLASS, OPEN_CLASS));
+      this._waitForAnimationToComplete(this._enterAnimationDuration, this._finishDialogOpen);
+    } else {
+      this._hostElement.classList.add(OPEN_CLASS);
+      Promise.resolve().then(() => this._finishDialogOpen());
+    }
+  }
+  /**
+   * Starts the exit animation of the dialog if enabled. This method is
+   * called by the dialog ref.
+   */
+  _startExitAnimation() {
+    this._animationStateChanged.emit({
+      state: "closing",
+      totalTime: this._exitAnimationDuration
+    });
+    this._hostElement.classList.remove(OPEN_CLASS);
+    if (this._animationsEnabled) {
+      this._hostElement.style.setProperty(TRANSITION_DURATION_PROPERTY, `${this._exitAnimationDuration}ms`);
+      this._requestAnimationFrame(() => this._hostElement.classList.add(CLOSING_CLASS));
+      this._waitForAnimationToComplete(this._exitAnimationDuration, this._finishDialogClose);
+    } else {
+      Promise.resolve().then(() => this._finishDialogClose());
+    }
+  }
+  /**
+   * Updates the number action sections.
+   * @param delta Increase/decrease in the number of sections.
+   */
+  _updateActionSectionCount(delta) {
+    this._actionSectionCount += delta;
+    this._changeDetectorRef.markForCheck();
+  }
+  /** Clears all dialog animation classes. */
+  _clearAnimationClasses() {
+    this._hostElement.classList.remove(OPENING_CLASS, CLOSING_CLASS);
+  }
+  _waitForAnimationToComplete(duration, callback) {
+    if (this._animationTimer !== null) {
+      clearTimeout(this._animationTimer);
+    }
+    this._animationTimer = setTimeout(callback, duration);
+  }
+  /** Runs a callback in `requestAnimationFrame`, if available. */
+  _requestAnimationFrame(callback) {
+    this._ngZone.runOutsideAngular(() => {
+      if (typeof requestAnimationFrame === "function") {
+        requestAnimationFrame(callback);
+      } else {
+        callback();
+      }
+    });
+  }
+  _captureInitialFocus() {
+    if (!this._config.delayFocusTrap) {
+      this._trapFocus();
+    }
+  }
+  /**
+   * Callback for when the open dialog animation has finished. Intended to
+   * be called by sub-classes that use different animation implementations.
+   */
+  _openAnimationDone(totalTime) {
+    if (this._config.delayFocusTrap) {
+      this._trapFocus();
+    }
+    this._animationStateChanged.next({
+      state: "opened",
+      totalTime
+    });
+  }
+  ngOnDestroy() {
+    super.ngOnDestroy();
+    if (this._animationTimer !== null) {
+      clearTimeout(this._animationTimer);
+    }
+  }
+  attachComponentPortal(portal) {
+    const ref = super.attachComponentPortal(portal);
+    ref.location.nativeElement.classList.add("mat-mdc-dialog-component-host");
+    return ref;
+  }
+};
+_MatDialogContainer.ɵfac = function MatDialogContainer_Factory(t) {
+  return new (t || _MatDialogContainer)(ɵɵdirectiveInject(ElementRef), ɵɵdirectiveInject(FocusTrapFactory), ɵɵdirectiveInject(DOCUMENT, 8), ɵɵdirectiveInject(MatDialogConfig), ɵɵdirectiveInject(InteractivityChecker), ɵɵdirectiveInject(NgZone), ɵɵdirectiveInject(OverlayRef), ɵɵdirectiveInject(ANIMATION_MODULE_TYPE, 8), ɵɵdirectiveInject(FocusMonitor));
+};
+_MatDialogContainer.ɵcmp = ɵɵdefineComponent({
+  type: _MatDialogContainer,
+  selectors: [["mat-dialog-container"]],
+  hostAttrs: ["tabindex", "-1", 1, "mat-mdc-dialog-container", "mdc-dialog"],
+  hostVars: 10,
+  hostBindings: function MatDialogContainer_HostBindings(rf, ctx) {
+    if (rf & 2) {
+      ɵɵhostProperty("id", ctx._config.id);
+      ɵɵattribute("aria-modal", ctx._config.ariaModal)("role", ctx._config.role)("aria-labelledby", ctx._config.ariaLabel ? null : ctx._ariaLabelledByQueue[0])("aria-label", ctx._config.ariaLabel)("aria-describedby", ctx._config.ariaDescribedBy || null);
+      ɵɵclassProp("_mat-animation-noopable", !ctx._animationsEnabled)("mat-mdc-dialog-container-with-actions", ctx._actionSectionCount > 0);
+    }
+  },
+  standalone: true,
+  features: [ɵɵInheritDefinitionFeature, ɵɵStandaloneFeature],
+  decls: 3,
+  vars: 0,
+  consts: [[1, "mdc-dialog__container"], [1, "mat-mdc-dialog-surface", "mdc-dialog__surface"], ["cdkPortalOutlet", ""]],
+  template: function MatDialogContainer_Template(rf, ctx) {
+    if (rf & 1) {
+      ɵɵelementStart(0, "div", 0)(1, "div", 1);
+      ɵɵtemplate(2, MatDialogContainer_ng_template_2_Template, 0, 0, "ng-template", 2);
+      ɵɵelementEnd()();
+    }
+  },
+  dependencies: [CdkPortalOutlet],
+  styles: ['.mdc-elevation-overlay{position:absolute;border-radius:inherit;pointer-events:none;opacity:var(--mdc-elevation-overlay-opacity, 0);transition:opacity 280ms cubic-bezier(0.4, 0, 0.2, 1)}.mdc-dialog,.mdc-dialog__scrim{position:fixed;top:0;left:0;align-items:center;justify-content:center;box-sizing:border-box;width:100%;height:100%}.mdc-dialog{display:none;z-index:var(--mdc-dialog-z-index, 7)}.mdc-dialog .mdc-dialog__content{padding:20px 24px 20px 24px}.mdc-dialog .mdc-dialog__surface{min-width:280px}@media(max-width: 592px){.mdc-dialog .mdc-dialog__surface{max-width:calc(100vw - 32px)}}@media(min-width: 592px){.mdc-dialog .mdc-dialog__surface{max-width:560px}}.mdc-dialog .mdc-dialog__surface{max-height:calc(100% - 32px)}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-width:none}@media(max-width: 960px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-height:560px;width:560px}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{right:-12px}}@media(max-width: 720px)and (max-width: 672px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{width:calc(100vw - 112px)}}@media(max-width: 720px)and (min-width: 672px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{width:560px}}@media(max-width: 720px)and (max-height: 720px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-height:calc(100vh - 160px)}}@media(max-width: 720px)and (min-height: 720px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-height:560px}}@media(max-width: 720px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{right:-12px}}@media(max-width: 720px)and (max-height: 400px),(max-width: 600px),(min-width: 720px)and (max-height: 400px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{height:100%;max-height:100vh;max-width:100vw;width:100vw;border-radius:0}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{order:-1;left:-12px}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__header{padding:0 16px 9px;justify-content:flex-start}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__title{margin-left:calc(16px - 2 * 12px)}}@media(min-width: 960px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{width:calc(100vw - 400px)}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{right:-12px}}.mdc-dialog.mdc-dialog__scrim--hidden .mdc-dialog__scrim{opacity:0}.mdc-dialog__scrim{opacity:0;z-index:-1}.mdc-dialog__container{display:flex;flex-direction:row;align-items:center;justify-content:space-around;box-sizing:border-box;height:100%;opacity:0;pointer-events:none}.mdc-dialog__surface{position:relative;display:flex;flex-direction:column;flex-grow:0;flex-shrink:0;box-sizing:border-box;max-width:100%;max-height:100%;pointer-events:auto;overflow-y:auto;outline:0;transform:scale(0.8)}.mdc-dialog__surface .mdc-elevation-overlay{width:100%;height:100%;top:0;left:0}[dir=rtl] .mdc-dialog__surface,.mdc-dialog__surface[dir=rtl]{text-align:right}@media screen and (forced-colors: active),(-ms-high-contrast: active){.mdc-dialog__surface{outline:2px solid windowText}}.mdc-dialog__surface::before{position:absolute;box-sizing:border-box;width:100%;height:100%;top:0;left:0;border:2px solid rgba(0,0,0,0);border-radius:inherit;content:"";pointer-events:none}@media screen and (forced-colors: active){.mdc-dialog__surface::before{border-color:CanvasText}}@media screen and (-ms-high-contrast: active),screen and (-ms-high-contrast: none){.mdc-dialog__surface::before{content:none}}.mdc-dialog__title{display:block;margin-top:0;position:relative;flex-shrink:0;box-sizing:border-box;margin:0 0 1px;padding:0 24px 9px}.mdc-dialog__title::before{display:inline-block;width:0;height:40px;content:"";vertical-align:0}[dir=rtl] .mdc-dialog__title,.mdc-dialog__title[dir=rtl]{text-align:right}.mdc-dialog--scrollable .mdc-dialog__title{margin-bottom:1px;padding-bottom:15px}.mdc-dialog--fullscreen .mdc-dialog__header{align-items:baseline;border-bottom:1px solid rgba(0,0,0,0);display:inline-flex;justify-content:space-between;padding:0 24px 9px;z-index:1}@media screen and (forced-colors: active){.mdc-dialog--fullscreen .mdc-dialog__header{border-bottom-color:CanvasText}}.mdc-dialog--fullscreen .mdc-dialog__header .mdc-dialog__close{right:-12px}.mdc-dialog--fullscreen .mdc-dialog__title{margin-bottom:0;padding:0;border-bottom:0}.mdc-dialog--fullscreen.mdc-dialog--scrollable .mdc-dialog__title{border-bottom:0;margin-bottom:0}.mdc-dialog--fullscreen .mdc-dialog__close{top:5px}.mdc-dialog--fullscreen.mdc-dialog--scrollable .mdc-dialog__actions{border-top:1px solid rgba(0,0,0,0)}@media screen and (forced-colors: active){.mdc-dialog--fullscreen.mdc-dialog--scrollable .mdc-dialog__actions{border-top-color:CanvasText}}.mdc-dialog--fullscreen--titleless .mdc-dialog__close{margin-top:4px}.mdc-dialog--fullscreen--titleless.mdc-dialog--scrollable .mdc-dialog__close{margin-top:0}.mdc-dialog__content{flex-grow:1;box-sizing:border-box;margin:0;overflow:auto}.mdc-dialog__content>:first-child{margin-top:0}.mdc-dialog__content>:last-child{margin-bottom:0}.mdc-dialog__title+.mdc-dialog__content,.mdc-dialog__header+.mdc-dialog__content{padding-top:0}.mdc-dialog--scrollable .mdc-dialog__title+.mdc-dialog__content{padding-top:8px;padding-bottom:8px}.mdc-dialog__content .mdc-deprecated-list:first-child:last-child{padding:6px 0 0}.mdc-dialog--scrollable .mdc-dialog__content .mdc-deprecated-list:first-child:last-child{padding:0}.mdc-dialog__actions{display:flex;position:relative;flex-shrink:0;flex-wrap:wrap;align-items:center;justify-content:flex-end;box-sizing:border-box;min-height:52px;margin:0;padding:8px;border-top:1px solid rgba(0,0,0,0)}@media screen and (forced-colors: active){.mdc-dialog__actions{border-top-color:CanvasText}}.mdc-dialog--stacked .mdc-dialog__actions{flex-direction:column;align-items:flex-end}.mdc-dialog__button{margin-left:8px;margin-right:0;max-width:100%;text-align:right}[dir=rtl] .mdc-dialog__button,.mdc-dialog__button[dir=rtl]{margin-left:0;margin-right:8px}.mdc-dialog__button:first-child{margin-left:0;margin-right:0}[dir=rtl] .mdc-dialog__button:first-child,.mdc-dialog__button:first-child[dir=rtl]{margin-left:0;margin-right:0}[dir=rtl] .mdc-dialog__button,.mdc-dialog__button[dir=rtl]{text-align:left}.mdc-dialog--stacked .mdc-dialog__button:not(:first-child){margin-top:12px}.mdc-dialog--open,.mdc-dialog--opening,.mdc-dialog--closing{display:flex}.mdc-dialog--opening .mdc-dialog__scrim{transition:opacity 150ms linear}.mdc-dialog--opening .mdc-dialog__container{transition:opacity 75ms linear,transform 150ms 0ms cubic-bezier(0, 0, 0.2, 1)}.mdc-dialog--closing .mdc-dialog__scrim,.mdc-dialog--closing .mdc-dialog__container{transition:opacity 75ms linear}.mdc-dialog--closing .mdc-dialog__container{transform:none}.mdc-dialog--closing .mdc-dialog__surface{transform:none}.mdc-dialog--open .mdc-dialog__scrim{opacity:1}.mdc-dialog--open .mdc-dialog__container{opacity:1}.mdc-dialog--open .mdc-dialog__surface{transform:none}.mdc-dialog--open.mdc-dialog__surface-scrim--shown .mdc-dialog__surface-scrim{opacity:1}.mdc-dialog--open.mdc-dialog__surface-scrim--hiding .mdc-dialog__surface-scrim{transition:opacity 75ms linear}.mdc-dialog--open.mdc-dialog__surface-scrim--showing .mdc-dialog__surface-scrim{transition:opacity 150ms linear}.mdc-dialog__surface-scrim{display:none;opacity:0;position:absolute;width:100%;height:100%;z-index:1}.mdc-dialog__surface-scrim--shown .mdc-dialog__surface-scrim,.mdc-dialog__surface-scrim--showing .mdc-dialog__surface-scrim,.mdc-dialog__surface-scrim--hiding .mdc-dialog__surface-scrim{display:block}.mdc-dialog-scroll-lock{overflow:hidden}.mdc-dialog--no-content-padding .mdc-dialog__content{padding:0}.mdc-dialog--sheet .mdc-dialog__container .mdc-dialog__close{right:12px;top:9px;position:absolute;z-index:1}.mdc-dialog__scrim--removed{pointer-events:none}.mdc-dialog__scrim--removed .mdc-dialog__scrim,.mdc-dialog__scrim--removed .mdc-dialog__surface-scrim{display:none}.mat-mdc-dialog-content{max-height:65vh}.mat-mdc-dialog-container{position:static;display:block}.mat-mdc-dialog-container,.mat-mdc-dialog-container .mdc-dialog__container,.mat-mdc-dialog-container .mdc-dialog__surface{max-height:inherit;min-height:inherit;min-width:inherit;max-width:inherit}.mat-mdc-dialog-container .mdc-dialog__surface{width:100%;height:100%}.mat-mdc-dialog-component-host{display:contents}.mat-mdc-dialog-container{--mdc-dialog-container-elevation: var(--mdc-dialog-container-elevation-shadow);outline:0}.mat-mdc-dialog-container .mdc-dialog__surface{background-color:var(--mdc-dialog-container-color, white)}.mat-mdc-dialog-container .mdc-dialog__surface{box-shadow:var(--mdc-dialog-container-elevation, 0px 11px 15px -7px rgba(0, 0, 0, 0.2), 0px 24px 38px 3px rgba(0, 0, 0, 0.14), 0px 9px 46px 8px rgba(0, 0, 0, 0.12))}.mat-mdc-dialog-container .mdc-dialog__surface{border-radius:var(--mdc-dialog-container-shape, 4px)}.mat-mdc-dialog-container .mdc-dialog__title{font-family:var(--mdc-dialog-subhead-font, Roboto, sans-serif);line-height:var(--mdc-dialog-subhead-line-height, 1.5rem);font-size:var(--mdc-dialog-subhead-size, 1rem);font-weight:var(--mdc-dialog-subhead-weight, 400);letter-spacing:var(--mdc-dialog-subhead-tracking, 0.03125em)}.mat-mdc-dialog-container .mdc-dialog__title{color:var(--mdc-dialog-subhead-color, rgba(0, 0, 0, 0.87))}.mat-mdc-dialog-container .mdc-dialog__content{font-family:var(--mdc-dialog-supporting-text-font, Roboto, sans-serif);line-height:var(--mdc-dialog-supporting-text-line-height, 1.5rem);font-size:var(--mdc-dialog-supporting-text-size, 1rem);font-weight:var(--mdc-dialog-supporting-text-weight, 400);letter-spacing:var(--mdc-dialog-supporting-text-tracking, 0.03125em)}.mat-mdc-dialog-container .mdc-dialog__content{color:var(--mdc-dialog-supporting-text-color, rgba(0, 0, 0, 0.6))}.mat-mdc-dialog-container .mdc-dialog__container{transition:opacity linear var(--mat-dialog-transition-duration, 0ms)}.mat-mdc-dialog-container .mdc-dialog__surface{transition:transform var(--mat-dialog-transition-duration, 0ms) 0ms cubic-bezier(0, 0, 0.2, 1)}.mat-mdc-dialog-container._mat-animation-noopable .mdc-dialog__container,.mat-mdc-dialog-container._mat-animation-noopable .mdc-dialog__surface{transition:none}.cdk-overlay-pane.mat-mdc-dialog-panel{max-width:var(--mat-dialog-container-max-width, 80vw);min-width:var(--mat-dialog-container-min-width, 0)}@media(max-width: 599px){.cdk-overlay-pane.mat-mdc-dialog-panel{max-width:var(--mat-dialog-container-small-max-width, 80vw)}}.mat-mdc-dialog-title{padding:var(--mat-dialog-headline-padding, 0 24px 9px)}.mat-mdc-dialog-content{display:block}.mat-mdc-dialog-container .mat-mdc-dialog-content{padding:var(--mat-dialog-content-padding, 20px 24px)}.mat-mdc-dialog-container-with-actions .mat-mdc-dialog-content{padding:var(--mat-dialog-with-actions-content-padding, 20px 24px)}.mat-mdc-dialog-container .mat-mdc-dialog-title+.mat-mdc-dialog-content{padding-top:0}.mat-mdc-dialog-actions{padding:var(--mat-dialog-actions-padding, 8px);justify-content:var(--mat-dialog-actions-alignment, start)}.mat-mdc-dialog-actions.mat-mdc-dialog-actions-align-start,.mat-mdc-dialog-actions[align=start]{justify-content:start}.mat-mdc-dialog-actions.mat-mdc-dialog-actions-align-center,.mat-mdc-dialog-actions[align=center]{justify-content:center}.mat-mdc-dialog-actions.mat-mdc-dialog-actions-align-end,.mat-mdc-dialog-actions[align=end]{justify-content:flex-end}.mat-mdc-dialog-actions .mat-button-base+.mat-button-base,.mat-mdc-dialog-actions .mat-mdc-button-base+.mat-mdc-button-base{margin-left:8px}[dir=rtl] .mat-mdc-dialog-actions .mat-button-base+.mat-button-base,[dir=rtl] .mat-mdc-dialog-actions .mat-mdc-button-base+.mat-mdc-button-base{margin-left:0;margin-right:8px}'],
+  encapsulation: 2
+});
+var MatDialogContainer = _MatDialogContainer;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogContainer, [{
+    type: Component,
+    args: [{
+      selector: "mat-dialog-container",
+      encapsulation: ViewEncapsulation$1.None,
+      changeDetection: ChangeDetectionStrategy.Default,
+      standalone: true,
+      imports: [CdkPortalOutlet],
+      host: {
+        "class": "mat-mdc-dialog-container mdc-dialog",
+        "tabindex": "-1",
+        "[attr.aria-modal]": "_config.ariaModal",
+        "[id]": "_config.id",
+        "[attr.role]": "_config.role",
+        "[attr.aria-labelledby]": "_config.ariaLabel ? null : _ariaLabelledByQueue[0]",
+        "[attr.aria-label]": "_config.ariaLabel",
+        "[attr.aria-describedby]": "_config.ariaDescribedBy || null",
+        "[class._mat-animation-noopable]": "!_animationsEnabled",
+        "[class.mat-mdc-dialog-container-with-actions]": "_actionSectionCount > 0"
+      },
+      template: '<div class="mdc-dialog__container">\n  <div class="mat-mdc-dialog-surface mdc-dialog__surface">\n    <ng-template cdkPortalOutlet />\n  </div>\n</div>\n',
+      styles: ['.mdc-elevation-overlay{position:absolute;border-radius:inherit;pointer-events:none;opacity:var(--mdc-elevation-overlay-opacity, 0);transition:opacity 280ms cubic-bezier(0.4, 0, 0.2, 1)}.mdc-dialog,.mdc-dialog__scrim{position:fixed;top:0;left:0;align-items:center;justify-content:center;box-sizing:border-box;width:100%;height:100%}.mdc-dialog{display:none;z-index:var(--mdc-dialog-z-index, 7)}.mdc-dialog .mdc-dialog__content{padding:20px 24px 20px 24px}.mdc-dialog .mdc-dialog__surface{min-width:280px}@media(max-width: 592px){.mdc-dialog .mdc-dialog__surface{max-width:calc(100vw - 32px)}}@media(min-width: 592px){.mdc-dialog .mdc-dialog__surface{max-width:560px}}.mdc-dialog .mdc-dialog__surface{max-height:calc(100% - 32px)}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-width:none}@media(max-width: 960px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-height:560px;width:560px}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{right:-12px}}@media(max-width: 720px)and (max-width: 672px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{width:calc(100vw - 112px)}}@media(max-width: 720px)and (min-width: 672px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{width:560px}}@media(max-width: 720px)and (max-height: 720px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-height:calc(100vh - 160px)}}@media(max-width: 720px)and (min-height: 720px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{max-height:560px}}@media(max-width: 720px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{right:-12px}}@media(max-width: 720px)and (max-height: 400px),(max-width: 600px),(min-width: 720px)and (max-height: 400px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{height:100%;max-height:100vh;max-width:100vw;width:100vw;border-radius:0}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{order:-1;left:-12px}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__header{padding:0 16px 9px;justify-content:flex-start}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__title{margin-left:calc(16px - 2 * 12px)}}@media(min-width: 960px){.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface{width:calc(100vw - 400px)}.mdc-dialog.mdc-dialog--fullscreen .mdc-dialog__surface .mdc-dialog__close{right:-12px}}.mdc-dialog.mdc-dialog__scrim--hidden .mdc-dialog__scrim{opacity:0}.mdc-dialog__scrim{opacity:0;z-index:-1}.mdc-dialog__container{display:flex;flex-direction:row;align-items:center;justify-content:space-around;box-sizing:border-box;height:100%;opacity:0;pointer-events:none}.mdc-dialog__surface{position:relative;display:flex;flex-direction:column;flex-grow:0;flex-shrink:0;box-sizing:border-box;max-width:100%;max-height:100%;pointer-events:auto;overflow-y:auto;outline:0;transform:scale(0.8)}.mdc-dialog__surface .mdc-elevation-overlay{width:100%;height:100%;top:0;left:0}[dir=rtl] .mdc-dialog__surface,.mdc-dialog__surface[dir=rtl]{text-align:right}@media screen and (forced-colors: active),(-ms-high-contrast: active){.mdc-dialog__surface{outline:2px solid windowText}}.mdc-dialog__surface::before{position:absolute;box-sizing:border-box;width:100%;height:100%;top:0;left:0;border:2px solid rgba(0,0,0,0);border-radius:inherit;content:"";pointer-events:none}@media screen and (forced-colors: active){.mdc-dialog__surface::before{border-color:CanvasText}}@media screen and (-ms-high-contrast: active),screen and (-ms-high-contrast: none){.mdc-dialog__surface::before{content:none}}.mdc-dialog__title{display:block;margin-top:0;position:relative;flex-shrink:0;box-sizing:border-box;margin:0 0 1px;padding:0 24px 9px}.mdc-dialog__title::before{display:inline-block;width:0;height:40px;content:"";vertical-align:0}[dir=rtl] .mdc-dialog__title,.mdc-dialog__title[dir=rtl]{text-align:right}.mdc-dialog--scrollable .mdc-dialog__title{margin-bottom:1px;padding-bottom:15px}.mdc-dialog--fullscreen .mdc-dialog__header{align-items:baseline;border-bottom:1px solid rgba(0,0,0,0);display:inline-flex;justify-content:space-between;padding:0 24px 9px;z-index:1}@media screen and (forced-colors: active){.mdc-dialog--fullscreen .mdc-dialog__header{border-bottom-color:CanvasText}}.mdc-dialog--fullscreen .mdc-dialog__header .mdc-dialog__close{right:-12px}.mdc-dialog--fullscreen .mdc-dialog__title{margin-bottom:0;padding:0;border-bottom:0}.mdc-dialog--fullscreen.mdc-dialog--scrollable .mdc-dialog__title{border-bottom:0;margin-bottom:0}.mdc-dialog--fullscreen .mdc-dialog__close{top:5px}.mdc-dialog--fullscreen.mdc-dialog--scrollable .mdc-dialog__actions{border-top:1px solid rgba(0,0,0,0)}@media screen and (forced-colors: active){.mdc-dialog--fullscreen.mdc-dialog--scrollable .mdc-dialog__actions{border-top-color:CanvasText}}.mdc-dialog--fullscreen--titleless .mdc-dialog__close{margin-top:4px}.mdc-dialog--fullscreen--titleless.mdc-dialog--scrollable .mdc-dialog__close{margin-top:0}.mdc-dialog__content{flex-grow:1;box-sizing:border-box;margin:0;overflow:auto}.mdc-dialog__content>:first-child{margin-top:0}.mdc-dialog__content>:last-child{margin-bottom:0}.mdc-dialog__title+.mdc-dialog__content,.mdc-dialog__header+.mdc-dialog__content{padding-top:0}.mdc-dialog--scrollable .mdc-dialog__title+.mdc-dialog__content{padding-top:8px;padding-bottom:8px}.mdc-dialog__content .mdc-deprecated-list:first-child:last-child{padding:6px 0 0}.mdc-dialog--scrollable .mdc-dialog__content .mdc-deprecated-list:first-child:last-child{padding:0}.mdc-dialog__actions{display:flex;position:relative;flex-shrink:0;flex-wrap:wrap;align-items:center;justify-content:flex-end;box-sizing:border-box;min-height:52px;margin:0;padding:8px;border-top:1px solid rgba(0,0,0,0)}@media screen and (forced-colors: active){.mdc-dialog__actions{border-top-color:CanvasText}}.mdc-dialog--stacked .mdc-dialog__actions{flex-direction:column;align-items:flex-end}.mdc-dialog__button{margin-left:8px;margin-right:0;max-width:100%;text-align:right}[dir=rtl] .mdc-dialog__button,.mdc-dialog__button[dir=rtl]{margin-left:0;margin-right:8px}.mdc-dialog__button:first-child{margin-left:0;margin-right:0}[dir=rtl] .mdc-dialog__button:first-child,.mdc-dialog__button:first-child[dir=rtl]{margin-left:0;margin-right:0}[dir=rtl] .mdc-dialog__button,.mdc-dialog__button[dir=rtl]{text-align:left}.mdc-dialog--stacked .mdc-dialog__button:not(:first-child){margin-top:12px}.mdc-dialog--open,.mdc-dialog--opening,.mdc-dialog--closing{display:flex}.mdc-dialog--opening .mdc-dialog__scrim{transition:opacity 150ms linear}.mdc-dialog--opening .mdc-dialog__container{transition:opacity 75ms linear,transform 150ms 0ms cubic-bezier(0, 0, 0.2, 1)}.mdc-dialog--closing .mdc-dialog__scrim,.mdc-dialog--closing .mdc-dialog__container{transition:opacity 75ms linear}.mdc-dialog--closing .mdc-dialog__container{transform:none}.mdc-dialog--closing .mdc-dialog__surface{transform:none}.mdc-dialog--open .mdc-dialog__scrim{opacity:1}.mdc-dialog--open .mdc-dialog__container{opacity:1}.mdc-dialog--open .mdc-dialog__surface{transform:none}.mdc-dialog--open.mdc-dialog__surface-scrim--shown .mdc-dialog__surface-scrim{opacity:1}.mdc-dialog--open.mdc-dialog__surface-scrim--hiding .mdc-dialog__surface-scrim{transition:opacity 75ms linear}.mdc-dialog--open.mdc-dialog__surface-scrim--showing .mdc-dialog__surface-scrim{transition:opacity 150ms linear}.mdc-dialog__surface-scrim{display:none;opacity:0;position:absolute;width:100%;height:100%;z-index:1}.mdc-dialog__surface-scrim--shown .mdc-dialog__surface-scrim,.mdc-dialog__surface-scrim--showing .mdc-dialog__surface-scrim,.mdc-dialog__surface-scrim--hiding .mdc-dialog__surface-scrim{display:block}.mdc-dialog-scroll-lock{overflow:hidden}.mdc-dialog--no-content-padding .mdc-dialog__content{padding:0}.mdc-dialog--sheet .mdc-dialog__container .mdc-dialog__close{right:12px;top:9px;position:absolute;z-index:1}.mdc-dialog__scrim--removed{pointer-events:none}.mdc-dialog__scrim--removed .mdc-dialog__scrim,.mdc-dialog__scrim--removed .mdc-dialog__surface-scrim{display:none}.mat-mdc-dialog-content{max-height:65vh}.mat-mdc-dialog-container{position:static;display:block}.mat-mdc-dialog-container,.mat-mdc-dialog-container .mdc-dialog__container,.mat-mdc-dialog-container .mdc-dialog__surface{max-height:inherit;min-height:inherit;min-width:inherit;max-width:inherit}.mat-mdc-dialog-container .mdc-dialog__surface{width:100%;height:100%}.mat-mdc-dialog-component-host{display:contents}.mat-mdc-dialog-container{--mdc-dialog-container-elevation: var(--mdc-dialog-container-elevation-shadow);outline:0}.mat-mdc-dialog-container .mdc-dialog__surface{background-color:var(--mdc-dialog-container-color, white)}.mat-mdc-dialog-container .mdc-dialog__surface{box-shadow:var(--mdc-dialog-container-elevation, 0px 11px 15px -7px rgba(0, 0, 0, 0.2), 0px 24px 38px 3px rgba(0, 0, 0, 0.14), 0px 9px 46px 8px rgba(0, 0, 0, 0.12))}.mat-mdc-dialog-container .mdc-dialog__surface{border-radius:var(--mdc-dialog-container-shape, 4px)}.mat-mdc-dialog-container .mdc-dialog__title{font-family:var(--mdc-dialog-subhead-font, Roboto, sans-serif);line-height:var(--mdc-dialog-subhead-line-height, 1.5rem);font-size:var(--mdc-dialog-subhead-size, 1rem);font-weight:var(--mdc-dialog-subhead-weight, 400);letter-spacing:var(--mdc-dialog-subhead-tracking, 0.03125em)}.mat-mdc-dialog-container .mdc-dialog__title{color:var(--mdc-dialog-subhead-color, rgba(0, 0, 0, 0.87))}.mat-mdc-dialog-container .mdc-dialog__content{font-family:var(--mdc-dialog-supporting-text-font, Roboto, sans-serif);line-height:var(--mdc-dialog-supporting-text-line-height, 1.5rem);font-size:var(--mdc-dialog-supporting-text-size, 1rem);font-weight:var(--mdc-dialog-supporting-text-weight, 400);letter-spacing:var(--mdc-dialog-supporting-text-tracking, 0.03125em)}.mat-mdc-dialog-container .mdc-dialog__content{color:var(--mdc-dialog-supporting-text-color, rgba(0, 0, 0, 0.6))}.mat-mdc-dialog-container .mdc-dialog__container{transition:opacity linear var(--mat-dialog-transition-duration, 0ms)}.mat-mdc-dialog-container .mdc-dialog__surface{transition:transform var(--mat-dialog-transition-duration, 0ms) 0ms cubic-bezier(0, 0, 0.2, 1)}.mat-mdc-dialog-container._mat-animation-noopable .mdc-dialog__container,.mat-mdc-dialog-container._mat-animation-noopable .mdc-dialog__surface{transition:none}.cdk-overlay-pane.mat-mdc-dialog-panel{max-width:var(--mat-dialog-container-max-width, 80vw);min-width:var(--mat-dialog-container-min-width, 0)}@media(max-width: 599px){.cdk-overlay-pane.mat-mdc-dialog-panel{max-width:var(--mat-dialog-container-small-max-width, 80vw)}}.mat-mdc-dialog-title{padding:var(--mat-dialog-headline-padding, 0 24px 9px)}.mat-mdc-dialog-content{display:block}.mat-mdc-dialog-container .mat-mdc-dialog-content{padding:var(--mat-dialog-content-padding, 20px 24px)}.mat-mdc-dialog-container-with-actions .mat-mdc-dialog-content{padding:var(--mat-dialog-with-actions-content-padding, 20px 24px)}.mat-mdc-dialog-container .mat-mdc-dialog-title+.mat-mdc-dialog-content{padding-top:0}.mat-mdc-dialog-actions{padding:var(--mat-dialog-actions-padding, 8px);justify-content:var(--mat-dialog-actions-alignment, start)}.mat-mdc-dialog-actions.mat-mdc-dialog-actions-align-start,.mat-mdc-dialog-actions[align=start]{justify-content:start}.mat-mdc-dialog-actions.mat-mdc-dialog-actions-align-center,.mat-mdc-dialog-actions[align=center]{justify-content:center}.mat-mdc-dialog-actions.mat-mdc-dialog-actions-align-end,.mat-mdc-dialog-actions[align=end]{justify-content:flex-end}.mat-mdc-dialog-actions .mat-button-base+.mat-button-base,.mat-mdc-dialog-actions .mat-mdc-button-base+.mat-mdc-button-base{margin-left:8px}[dir=rtl] .mat-mdc-dialog-actions .mat-button-base+.mat-button-base,[dir=rtl] .mat-mdc-dialog-actions .mat-mdc-button-base+.mat-mdc-button-base{margin-left:0;margin-right:8px}']
+    }]
+  }], () => [{
+    type: ElementRef
+  }, {
+    type: FocusTrapFactory
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: MatDialogConfig
+  }, {
+    type: InteractivityChecker
+  }, {
+    type: NgZone
+  }, {
+    type: OverlayRef
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [ANIMATION_MODULE_TYPE]
+    }]
+  }, {
+    type: FocusMonitor
+  }], null);
+})();
+var TRANSITION_DURATION_PROPERTY = "--mat-dialog-transition-duration";
+function parseCssTime(time) {
+  if (time == null) {
+    return null;
+  }
+  if (typeof time === "number") {
+    return time;
+  }
+  if (time.endsWith("ms")) {
+    return coerceNumberProperty(time.substring(0, time.length - 2));
+  }
+  if (time.endsWith("s")) {
+    return coerceNumberProperty(time.substring(0, time.length - 1)) * 1e3;
+  }
+  if (time === "0") {
+    return 0;
+  }
+  return null;
+}
+var MatDialogState;
+(function(MatDialogState2) {
+  MatDialogState2[MatDialogState2["OPEN"] = 0] = "OPEN";
+  MatDialogState2[MatDialogState2["CLOSING"] = 1] = "CLOSING";
+  MatDialogState2[MatDialogState2["CLOSED"] = 2] = "CLOSED";
+})(MatDialogState || (MatDialogState = {}));
+var MatDialogRef = class {
+  constructor(_ref, config, _containerInstance) {
+    this._ref = _ref;
+    this._containerInstance = _containerInstance;
+    this._afterOpened = new Subject();
+    this._beforeClosed = new Subject();
+    this._state = MatDialogState.OPEN;
+    this.disableClose = config.disableClose;
+    this.id = _ref.id;
+    _ref.addPanelClass("mat-mdc-dialog-panel");
+    _containerInstance._animationStateChanged.pipe(filter((event) => event.state === "opened"), take(1)).subscribe(() => {
+      this._afterOpened.next();
+      this._afterOpened.complete();
+    });
+    _containerInstance._animationStateChanged.pipe(filter((event) => event.state === "closed"), take(1)).subscribe(() => {
+      clearTimeout(this._closeFallbackTimeout);
+      this._finishDialogClose();
+    });
+    _ref.overlayRef.detachments().subscribe(() => {
+      this._beforeClosed.next(this._result);
+      this._beforeClosed.complete();
+      this._finishDialogClose();
+    });
+    merge(this.backdropClick(), this.keydownEvents().pipe(filter((event) => event.keyCode === ESCAPE && !this.disableClose && !hasModifierKey(event)))).subscribe((event) => {
+      if (!this.disableClose) {
+        event.preventDefault();
+        _closeDialogVia(this, event.type === "keydown" ? "keyboard" : "mouse");
+      }
+    });
+  }
+  /**
+   * Close the dialog.
+   * @param dialogResult Optional result to return to the dialog opener.
+   */
+  close(dialogResult) {
+    this._result = dialogResult;
+    this._containerInstance._animationStateChanged.pipe(filter((event) => event.state === "closing"), take(1)).subscribe((event) => {
+      this._beforeClosed.next(dialogResult);
+      this._beforeClosed.complete();
+      this._ref.overlayRef.detachBackdrop();
+      this._closeFallbackTimeout = setTimeout(() => this._finishDialogClose(), event.totalTime + 100);
+    });
+    this._state = MatDialogState.CLOSING;
+    this._containerInstance._startExitAnimation();
+  }
+  /**
+   * Gets an observable that is notified when the dialog is finished opening.
+   */
+  afterOpened() {
+    return this._afterOpened;
+  }
+  /**
+   * Gets an observable that is notified when the dialog is finished closing.
+   */
+  afterClosed() {
+    return this._ref.closed;
+  }
+  /**
+   * Gets an observable that is notified when the dialog has started closing.
+   */
+  beforeClosed() {
+    return this._beforeClosed;
+  }
+  /**
+   * Gets an observable that emits when the overlay's backdrop has been clicked.
+   */
+  backdropClick() {
+    return this._ref.backdropClick;
+  }
+  /**
+   * Gets an observable that emits when keydown events are targeted on the overlay.
+   */
+  keydownEvents() {
+    return this._ref.keydownEvents;
+  }
+  /**
+   * Updates the dialog's position.
+   * @param position New dialog position.
+   */
+  updatePosition(position) {
+    let strategy = this._ref.config.positionStrategy;
+    if (position && (position.left || position.right)) {
+      position.left ? strategy.left(position.left) : strategy.right(position.right);
+    } else {
+      strategy.centerHorizontally();
+    }
+    if (position && (position.top || position.bottom)) {
+      position.top ? strategy.top(position.top) : strategy.bottom(position.bottom);
+    } else {
+      strategy.centerVertically();
+    }
+    this._ref.updatePosition();
+    return this;
+  }
+  /**
+   * Updates the dialog's width and height.
+   * @param width New width of the dialog.
+   * @param height New height of the dialog.
+   */
+  updateSize(width = "", height = "") {
+    this._ref.updateSize(width, height);
+    return this;
+  }
+  /** Add a CSS class or an array of classes to the overlay pane. */
+  addPanelClass(classes) {
+    this._ref.addPanelClass(classes);
+    return this;
+  }
+  /** Remove a CSS class or an array of classes from the overlay pane. */
+  removePanelClass(classes) {
+    this._ref.removePanelClass(classes);
+    return this;
+  }
+  /** Gets the current state of the dialog's lifecycle. */
+  getState() {
+    return this._state;
+  }
+  /**
+   * Finishes the dialog close by updating the state of the dialog
+   * and disposing the overlay.
+   */
+  _finishDialogClose() {
+    this._state = MatDialogState.CLOSED;
+    this._ref.close(this._result, {
+      focusOrigin: this._closeInteractionType
+    });
+    this.componentInstance = null;
+  }
+};
+function _closeDialogVia(ref, interactionType, result) {
+  ref._closeInteractionType = interactionType;
+  return ref.close(result);
+}
+var MAT_DIALOG_DATA = new InjectionToken("MatMdcDialogData");
+var MAT_DIALOG_DEFAULT_OPTIONS = new InjectionToken("mat-mdc-dialog-default-options");
+var MAT_DIALOG_SCROLL_STRATEGY = new InjectionToken("mat-mdc-dialog-scroll-strategy", {
+  providedIn: "root",
+  factory: () => {
+    const overlay = inject(Overlay);
+    return () => overlay.scrollStrategies.block();
+  }
+});
+function MAT_DIALOG_SCROLL_STRATEGY_PROVIDER_FACTORY(overlay) {
+  return () => overlay.scrollStrategies.block();
+}
+var MAT_DIALOG_SCROLL_STRATEGY_PROVIDER = {
+  provide: MAT_DIALOG_SCROLL_STRATEGY,
+  deps: [Overlay],
+  useFactory: MAT_DIALOG_SCROLL_STRATEGY_PROVIDER_FACTORY
+};
+var uniqueId2 = 0;
+var _MatDialog = class _MatDialog {
+  /** Keeps track of the currently-open dialogs. */
+  get openDialogs() {
+    return this._parentDialog ? this._parentDialog.openDialogs : this._openDialogsAtThisLevel;
+  }
+  /** Stream that emits when a dialog has been opened. */
+  get afterOpened() {
+    return this._parentDialog ? this._parentDialog.afterOpened : this._afterOpenedAtThisLevel;
+  }
+  _getAfterAllClosed() {
+    const parent = this._parentDialog;
+    return parent ? parent._getAfterAllClosed() : this._afterAllClosedAtThisLevel;
+  }
+  constructor(_overlay, injector, location, _defaultOptions, _scrollStrategy, _parentDialog, _overlayContainer, _animationMode) {
+    this._overlay = _overlay;
+    this._defaultOptions = _defaultOptions;
+    this._scrollStrategy = _scrollStrategy;
+    this._parentDialog = _parentDialog;
+    this._openDialogsAtThisLevel = [];
+    this._afterAllClosedAtThisLevel = new Subject();
+    this._afterOpenedAtThisLevel = new Subject();
+    this.dialogConfigClass = MatDialogConfig;
+    this.afterAllClosed = defer(() => this.openDialogs.length ? this._getAfterAllClosed() : this._getAfterAllClosed().pipe(startWith(void 0)));
+    this._dialog = injector.get(Dialog);
+    this._dialogRefConstructor = MatDialogRef;
+    this._dialogContainerType = MatDialogContainer;
+    this._dialogDataToken = MAT_DIALOG_DATA;
+  }
+  open(componentOrTemplateRef, config) {
+    let dialogRef;
+    config = __spreadValues(__spreadValues({}, this._defaultOptions || new MatDialogConfig()), config);
+    config.id = config.id || `mat-mdc-dialog-${uniqueId2++}`;
+    config.scrollStrategy = config.scrollStrategy || this._scrollStrategy();
+    const cdkRef = this._dialog.open(componentOrTemplateRef, __spreadProps(__spreadValues({}, config), {
+      positionStrategy: this._overlay.position().global().centerHorizontally().centerVertically(),
+      // Disable closing since we need to sync it up to the animation ourselves.
+      disableClose: true,
+      // Disable closing on destroy, because this service cleans up its open dialogs as well.
+      // We want to do the cleanup here, rather than the CDK service, because the CDK destroys
+      // the dialogs immediately whereas we want it to wait for the animations to finish.
+      closeOnDestroy: false,
+      // Disable closing on detachments so that we can sync up the animation.
+      // The Material dialog ref handles this manually.
+      closeOnOverlayDetachments: false,
+      container: {
+        type: this._dialogContainerType,
+        providers: () => [
+          // Provide our config as the CDK config as well since it has the same interface as the
+          // CDK one, but it contains the actual values passed in by the user for things like
+          // `disableClose` which we disable for the CDK dialog since we handle it ourselves.
+          {
+            provide: this.dialogConfigClass,
+            useValue: config
+          },
+          {
+            provide: DialogConfig,
+            useValue: config
+          }
+        ]
+      },
+      templateContext: () => ({
+        dialogRef
+      }),
+      providers: (ref, cdkConfig, dialogContainer) => {
+        dialogRef = new this._dialogRefConstructor(ref, config, dialogContainer);
+        dialogRef.updatePosition(config?.position);
+        return [{
+          provide: this._dialogContainerType,
+          useValue: dialogContainer
+        }, {
+          provide: this._dialogDataToken,
+          useValue: cdkConfig.data
+        }, {
+          provide: this._dialogRefConstructor,
+          useValue: dialogRef
+        }];
+      }
+    }));
+    dialogRef.componentRef = cdkRef.componentRef;
+    dialogRef.componentInstance = cdkRef.componentInstance;
+    this.openDialogs.push(dialogRef);
+    this.afterOpened.next(dialogRef);
+    dialogRef.afterClosed().subscribe(() => {
+      const index = this.openDialogs.indexOf(dialogRef);
+      if (index > -1) {
+        this.openDialogs.splice(index, 1);
+        if (!this.openDialogs.length) {
+          this._getAfterAllClosed().next();
+        }
+      }
+    });
+    return dialogRef;
+  }
+  /**
+   * Closes all of the currently-open dialogs.
+   */
+  closeAll() {
+    this._closeDialogs(this.openDialogs);
+  }
+  /**
+   * Finds an open dialog by its id.
+   * @param id ID to use when looking up the dialog.
+   */
+  getDialogById(id) {
+    return this.openDialogs.find((dialog) => dialog.id === id);
+  }
+  ngOnDestroy() {
+    this._closeDialogs(this._openDialogsAtThisLevel);
+    this._afterAllClosedAtThisLevel.complete();
+    this._afterOpenedAtThisLevel.complete();
+  }
+  _closeDialogs(dialogs) {
+    let i = dialogs.length;
+    while (i--) {
+      dialogs[i].close();
+    }
+  }
+};
+_MatDialog.ɵfac = function MatDialog_Factory(t) {
+  return new (t || _MatDialog)(ɵɵinject(Overlay), ɵɵinject(Injector), ɵɵinject(Location, 8), ɵɵinject(MAT_DIALOG_DEFAULT_OPTIONS, 8), ɵɵinject(MAT_DIALOG_SCROLL_STRATEGY), ɵɵinject(_MatDialog, 12), ɵɵinject(OverlayContainer), ɵɵinject(ANIMATION_MODULE_TYPE, 8));
+};
+_MatDialog.ɵprov = ɵɵdefineInjectable({
+  token: _MatDialog,
+  factory: _MatDialog.ɵfac,
+  providedIn: "root"
+});
+var MatDialog = _MatDialog;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialog, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: Overlay
+  }, {
+    type: Injector
+  }, {
+    type: Location,
+    decorators: [{
+      type: Optional
+    }]
+  }, {
+    type: MatDialogConfig,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [MAT_DIALOG_DEFAULT_OPTIONS]
+    }]
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [MAT_DIALOG_SCROLL_STRATEGY]
+    }]
+  }, {
+    type: MatDialog,
+    decorators: [{
+      type: Optional
+    }, {
+      type: SkipSelf
+    }]
+  }, {
+    type: OverlayContainer
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [ANIMATION_MODULE_TYPE]
+    }]
+  }], null);
+})();
+var dialogElementUid = 0;
+var _MatDialogClose = class _MatDialogClose {
+  constructor(dialogRef, _elementRef, _dialog) {
+    this.dialogRef = dialogRef;
+    this._elementRef = _elementRef;
+    this._dialog = _dialog;
+    this.type = "button";
+  }
+  ngOnInit() {
+    if (!this.dialogRef) {
+      this.dialogRef = getClosestDialog(this._elementRef, this._dialog.openDialogs);
+    }
+  }
+  ngOnChanges(changes) {
+    const proxiedChange = changes["_matDialogClose"] || changes["_matDialogCloseResult"];
+    if (proxiedChange) {
+      this.dialogResult = proxiedChange.currentValue;
+    }
+  }
+  _onButtonClick(event) {
+    _closeDialogVia(this.dialogRef, event.screenX === 0 && event.screenY === 0 ? "keyboard" : "mouse", this.dialogResult);
+  }
+};
+_MatDialogClose.ɵfac = function MatDialogClose_Factory(t) {
+  return new (t || _MatDialogClose)(ɵɵdirectiveInject(MatDialogRef, 8), ɵɵdirectiveInject(ElementRef), ɵɵdirectiveInject(MatDialog));
+};
+_MatDialogClose.ɵdir = ɵɵdefineDirective({
+  type: _MatDialogClose,
+  selectors: [["", "mat-dialog-close", ""], ["", "matDialogClose", ""]],
+  hostVars: 2,
+  hostBindings: function MatDialogClose_HostBindings(rf, ctx) {
+    if (rf & 1) {
+      ɵɵlistener("click", function MatDialogClose_click_HostBindingHandler($event) {
+        return ctx._onButtonClick($event);
+      });
+    }
+    if (rf & 2) {
+      ɵɵattribute("aria-label", ctx.ariaLabel || null)("type", ctx.type);
+    }
+  },
+  inputs: {
+    ariaLabel: [InputFlags.None, "aria-label", "ariaLabel"],
+    type: "type",
+    dialogResult: [InputFlags.None, "mat-dialog-close", "dialogResult"],
+    _matDialogClose: [InputFlags.None, "matDialogClose", "_matDialogClose"]
+  },
+  exportAs: ["matDialogClose"],
+  standalone: true,
+  features: [ɵɵNgOnChangesFeature]
+});
+var MatDialogClose = _MatDialogClose;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogClose, [{
+    type: Directive,
+    args: [{
+      selector: "[mat-dialog-close], [matDialogClose]",
+      exportAs: "matDialogClose",
+      standalone: true,
+      host: {
+        "(click)": "_onButtonClick($event)",
+        "[attr.aria-label]": "ariaLabel || null",
+        "[attr.type]": "type"
+      }
+    }]
+  }], () => [{
+    type: MatDialogRef,
+    decorators: [{
+      type: Optional
+    }]
+  }, {
+    type: ElementRef
+  }, {
+    type: MatDialog
+  }], {
+    ariaLabel: [{
+      type: Input,
+      args: ["aria-label"]
+    }],
+    type: [{
+      type: Input
+    }],
+    dialogResult: [{
+      type: Input,
+      args: ["mat-dialog-close"]
+    }],
+    _matDialogClose: [{
+      type: Input,
+      args: ["matDialogClose"]
+    }]
+  });
+})();
+var _MatDialogLayoutSection = class _MatDialogLayoutSection {
+  constructor(_dialogRef, _elementRef, _dialog) {
+    this._dialogRef = _dialogRef;
+    this._elementRef = _elementRef;
+    this._dialog = _dialog;
+  }
+  ngOnInit() {
+    if (!this._dialogRef) {
+      this._dialogRef = getClosestDialog(this._elementRef, this._dialog.openDialogs);
+    }
+    if (this._dialogRef) {
+      Promise.resolve().then(() => {
+        this._onAdd();
+      });
+    }
+  }
+  ngOnDestroy() {
+    const instance = this._dialogRef?._containerInstance;
+    if (instance) {
+      Promise.resolve().then(() => {
+        this._onRemove();
+      });
+    }
+  }
+};
+_MatDialogLayoutSection.ɵfac = function MatDialogLayoutSection_Factory(t) {
+  return new (t || _MatDialogLayoutSection)(ɵɵdirectiveInject(MatDialogRef, 8), ɵɵdirectiveInject(ElementRef), ɵɵdirectiveInject(MatDialog));
+};
+_MatDialogLayoutSection.ɵdir = ɵɵdefineDirective({
+  type: _MatDialogLayoutSection,
+  standalone: true
+});
+var MatDialogLayoutSection = _MatDialogLayoutSection;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogLayoutSection, [{
+    type: Directive,
+    args: [{
+      standalone: true
+    }]
+  }], () => [{
+    type: MatDialogRef,
+    decorators: [{
+      type: Optional
+    }]
+  }, {
+    type: ElementRef
+  }, {
+    type: MatDialog
+  }], null);
+})();
+var _MatDialogTitle = class _MatDialogTitle extends MatDialogLayoutSection {
+  constructor() {
+    super(...arguments);
+    this.id = `mat-mdc-dialog-title-${dialogElementUid++}`;
+  }
+  _onAdd() {
+    this._dialogRef._containerInstance?._addAriaLabelledBy?.(this.id);
+  }
+  _onRemove() {
+    this._dialogRef?._containerInstance?._removeAriaLabelledBy?.(this.id);
+  }
+};
+_MatDialogTitle.ɵfac = /* @__PURE__ */ (() => {
+  let ɵMatDialogTitle_BaseFactory;
+  return function MatDialogTitle_Factory(t) {
+    return (ɵMatDialogTitle_BaseFactory || (ɵMatDialogTitle_BaseFactory = ɵɵgetInheritedFactory(_MatDialogTitle)))(t || _MatDialogTitle);
+  };
+})();
+_MatDialogTitle.ɵdir = ɵɵdefineDirective({
+  type: _MatDialogTitle,
+  selectors: [["", "mat-dialog-title", ""], ["", "matDialogTitle", ""]],
+  hostAttrs: [1, "mat-mdc-dialog-title", "mdc-dialog__title"],
+  hostVars: 1,
+  hostBindings: function MatDialogTitle_HostBindings(rf, ctx) {
+    if (rf & 2) {
+      ɵɵhostProperty("id", ctx.id);
+    }
+  },
+  inputs: {
+    id: "id"
+  },
+  exportAs: ["matDialogTitle"],
+  standalone: true,
+  features: [ɵɵInheritDefinitionFeature]
+});
+var MatDialogTitle = _MatDialogTitle;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogTitle, [{
+    type: Directive,
+    args: [{
+      selector: "[mat-dialog-title], [matDialogTitle]",
+      exportAs: "matDialogTitle",
+      standalone: true,
+      host: {
+        "class": "mat-mdc-dialog-title mdc-dialog__title",
+        "[id]": "id"
+      }
+    }]
+  }], null, {
+    id: [{
+      type: Input
+    }]
+  });
+})();
+var _MatDialogContent = class _MatDialogContent {
+};
+_MatDialogContent.ɵfac = function MatDialogContent_Factory(t) {
+  return new (t || _MatDialogContent)();
+};
+_MatDialogContent.ɵdir = ɵɵdefineDirective({
+  type: _MatDialogContent,
+  selectors: [["", "mat-dialog-content", ""], ["mat-dialog-content"], ["", "matDialogContent", ""]],
+  hostAttrs: [1, "mat-mdc-dialog-content", "mdc-dialog__content"],
+  standalone: true
+});
+var MatDialogContent = _MatDialogContent;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogContent, [{
+    type: Directive,
+    args: [{
+      selector: `[mat-dialog-content], mat-dialog-content, [matDialogContent]`,
+      host: {
+        "class": "mat-mdc-dialog-content mdc-dialog__content"
+      },
+      standalone: true
+    }]
+  }], null, null);
+})();
+var _MatDialogActions = class _MatDialogActions extends MatDialogLayoutSection {
+  _onAdd() {
+    this._dialogRef._containerInstance?._updateActionSectionCount?.(1);
+  }
+  _onRemove() {
+    this._dialogRef._containerInstance?._updateActionSectionCount?.(-1);
+  }
+};
+_MatDialogActions.ɵfac = /* @__PURE__ */ (() => {
+  let ɵMatDialogActions_BaseFactory;
+  return function MatDialogActions_Factory(t) {
+    return (ɵMatDialogActions_BaseFactory || (ɵMatDialogActions_BaseFactory = ɵɵgetInheritedFactory(_MatDialogActions)))(t || _MatDialogActions);
+  };
+})();
+_MatDialogActions.ɵdir = ɵɵdefineDirective({
+  type: _MatDialogActions,
+  selectors: [["", "mat-dialog-actions", ""], ["mat-dialog-actions"], ["", "matDialogActions", ""]],
+  hostAttrs: [1, "mat-mdc-dialog-actions", "mdc-dialog__actions"],
+  hostVars: 6,
+  hostBindings: function MatDialogActions_HostBindings(rf, ctx) {
+    if (rf & 2) {
+      ɵɵclassProp("mat-mdc-dialog-actions-align-start", ctx.align === "start")("mat-mdc-dialog-actions-align-center", ctx.align === "center")("mat-mdc-dialog-actions-align-end", ctx.align === "end");
+    }
+  },
+  inputs: {
+    align: "align"
+  },
+  standalone: true,
+  features: [ɵɵInheritDefinitionFeature]
+});
+var MatDialogActions = _MatDialogActions;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogActions, [{
+    type: Directive,
+    args: [{
+      selector: `[mat-dialog-actions], mat-dialog-actions, [matDialogActions]`,
+      standalone: true,
+      host: {
+        "class": "mat-mdc-dialog-actions mdc-dialog__actions",
+        "[class.mat-mdc-dialog-actions-align-start]": 'align === "start"',
+        "[class.mat-mdc-dialog-actions-align-center]": 'align === "center"',
+        "[class.mat-mdc-dialog-actions-align-end]": 'align === "end"'
+      }
+    }]
+  }], null, {
+    align: [{
+      type: Input
+    }]
+  });
+})();
+function getClosestDialog(element, openDialogs) {
+  let parent = element.nativeElement.parentElement;
+  while (parent && !parent.classList.contains("mat-mdc-dialog-container")) {
+    parent = parent.parentElement;
+  }
+  return parent ? openDialogs.find((dialog) => dialog.id === parent.id) : null;
+}
+var DIRECTIVES = [MatDialogContainer, MatDialogClose, MatDialogTitle, MatDialogActions, MatDialogContent];
+var _MatDialogModule = class _MatDialogModule {
+};
+_MatDialogModule.ɵfac = function MatDialogModule_Factory(t) {
+  return new (t || _MatDialogModule)();
+};
+_MatDialogModule.ɵmod = ɵɵdefineNgModule({
+  type: _MatDialogModule,
+  imports: [DialogModule, OverlayModule, PortalModule, MatCommonModule, MatDialogContainer, MatDialogClose, MatDialogTitle, MatDialogActions, MatDialogContent],
+  exports: [MatCommonModule, MatDialogContainer, MatDialogClose, MatDialogTitle, MatDialogActions, MatDialogContent]
+});
+_MatDialogModule.ɵinj = ɵɵdefineInjector({
+  providers: [MatDialog],
+  imports: [DialogModule, OverlayModule, PortalModule, MatCommonModule, MatCommonModule]
+});
+var MatDialogModule = _MatDialogModule;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(MatDialogModule, [{
+    type: NgModule,
+    args: [{
+      imports: [DialogModule, OverlayModule, PortalModule, MatCommonModule, ...DIRECTIVES],
+      exports: [MatCommonModule, ...DIRECTIVES],
+      providers: [MatDialog]
+    }]
+  }], null, null);
+})();
+var _defaultParams = {
+  params: {
+    enterAnimationDuration: "150ms",
+    exitAnimationDuration: "75ms"
+  }
+};
+var matDialogAnimations = {
+  /** Animation that is applied on the dialog container by default. */
+  dialogContainer: trigger("dialogContainer", [
+    // Note: The `enter` animation transitions to `transform: none`, because for some reason
+    // specifying the transform explicitly, causes IE both to blur the dialog content and
+    // decimate the animation performance. Leaving it as `none` solves both issues.
+    state("void, exit", style({
+      opacity: 0,
+      transform: "scale(0.7)"
+    })),
+    state("enter", style({
+      transform: "none"
+    })),
+    transition("* => enter", group([animate("{{enterAnimationDuration}} cubic-bezier(0, 0, 0.2, 1)", style({
+      transform: "none",
+      opacity: 1
+    })), query("@*", animateChild(), {
+      optional: true
+    })]), _defaultParams),
+    transition("* => void, * => exit", group([animate("{{exitAnimationDuration}} cubic-bezier(0.4, 0.0, 0.2, 1)", style({
+      opacity: 0
+    })), query("@*", animateChild(), {
+      optional: true
+    })]), _defaultParams)
+  ])
+};
+export {
+  MAT_DIALOG_DATA,
+  MAT_DIALOG_DEFAULT_OPTIONS,
+  MAT_DIALOG_SCROLL_STRATEGY,
+  MAT_DIALOG_SCROLL_STRATEGY_PROVIDER,
+  MAT_DIALOG_SCROLL_STRATEGY_PROVIDER_FACTORY,
+  MatDialog,
+  MatDialogActions,
+  MatDialogClose,
+  MatDialogConfig,
+  MatDialogContainer,
+  MatDialogContent,
+  MatDialogModule,
+  MatDialogRef,
+  MatDialogState,
+  MatDialogTitle,
+  _closeDialogVia,
+  _defaultParams,
+  matDialogAnimations
+};
+//# sourceMappingURL=@angular_material_dialog.js.map
