+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_material_input.js
@@ -14,8 +14,8 @@ import {
   MatLabel,
   MatPrefix,
   MatSuffix
-} from "./chunk-C2Z32BJ6.js";
-import "./chunk-2AXA5NS2.js";
+} from "./chunk-SXOQ3UHC.js";
+import "./chunk-VNW2MBA7.js";
 import {
   ErrorStateMatcher,
   MatCommonModule,
@@ -26,7 +26,7 @@ import {
   coerceNumberProperty,
   getSupportedInputTypes,
   normalizePassiveListenerOptions
-} from "./chunk-FLGVYJXI.js";
+} from "./chunk-7ZFYYFG2.js";
 import {
   DOCUMENT
 } from "./chunk-7GFUNZUG.js";
