+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_material_table.js
@@ -7,13 +7,13 @@ import {
   _VIEW_REPEATER_STRATEGY,
   _ViewRepeaterOperation,
   isDataSource
-} from "./chunk-ATNDQTJA.js";
+} from "./chunk-5H5WHFDB.js";
 import {
   Directionality,
   MatCommonModule,
   Platform,
   _isNumberValue
-} from "./chunk-FLGVYJXI.js";
+} from "./chunk-7ZFYYFG2.js";
 import {
   DOCUMENT
 } from "./chunk-7GFUNZUG.js";
