+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/@angular_platform-browser_animations_async.js
@@ -48,7 +48,7 @@ var _AsyncAnimationRendererFactory = class _AsyncAnimationRendererFactory {
    * @internal
    */
   loadImpl() {
-    const moduleImpl = this.moduleImpl ?? import("./browser-BVPPLFFG.js");
+    const moduleImpl = this.moduleImpl ?? import("./browser-7WVVOWIQ.js");
     return moduleImpl.catch((e) => {
       throw new RuntimeError(5300, (typeof ngDevMode === "undefined" || ngDevMode) && "Async loading for animations package was enabled, but loading failed. Angular falls back to using regular rendering. No animations will be displayed and their styles won't be applied.");
     }).then(({
