+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/_metadata.json
@@ -1,124 +1,139 @@
 {
-  "hash": "aac4111b",
-  "configHash": "38c8622a",
-  "lockfileHash": "2be94f71",
-  "browserHash": "4dd5c888",
+  "hash": "1264eee2",
+  "configHash": "5dd32aa1",
+  "lockfileHash": "0e11fd84",
+  "browserHash": "f93aa3e3",
   "optimized": {
     "@angular/common": {
       "src": "../../../../../node_modules/@angular/common/fesm2022/common.mjs",
       "file": "@angular_common.js",
-      "fileHash": "fd633fe5",
+      "fileHash": "17a04359",
       "needsInterop": false
     },
     "@angular/common/http": {
       "src": "../../../../../node_modules/@angular/common/fesm2022/http.mjs",
       "file": "@angular_common_http.js",
-      "fileHash": "20ce3c8b",
+      "fileHash": "85bd3786",
       "needsInterop": false
     },
     "@angular/core": {
       "src": "../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "76ef5756",
+      "fileHash": "ab058838",
       "needsInterop": false
     },
     "@angular/forms": {
       "src": "../../../../../node_modules/@angular/forms/fesm2022/forms.mjs",
       "file": "@angular_forms.js",
-      "fileHash": "27408ba2",
+      "fileHash": "255be621",
       "needsInterop": false
     },
     "@angular/material/button": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/button.mjs",
       "file": "@angular_material_button.js",
-      "fileHash": "84ea3131",
+      "fileHash": "8a3cb349",
+      "needsInterop": false
+    },
+    "@angular/material/dialog": {
+      "src": "../../../../../node_modules/@angular/material/fesm2022/dialog.mjs",
+      "file": "@angular_material_dialog.js",
+      "fileHash": "2e979262",
+      "needsInterop": false
+    },
+    "@angular/material/divider": {
+      "src": "../../../../../node_modules/@angular/material/fesm2022/divider.mjs",
+      "file": "@angular_material_divider.js",
+      "fileHash": "a3a1c0f8",
       "needsInterop": false
     },
     "@angular/material/form-field": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/form-field.mjs",
       "file": "@angular_material_form-field.js",
-      "fileHash": "8559d044",
+      "fileHash": "c0ab74b9",
       "needsInterop": false
     },
     "@angular/material/input": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/input.mjs",
       "file": "@angular_material_input.js",
-      "fileHash": "48008810",
+      "fileHash": "40a1852c",
       "needsInterop": false
     },
     "@angular/material/select": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/select.mjs",
       "file": "@angular_material_select.js",
-      "fileHash": "e122e30e",
+      "fileHash": "6042822b",
       "needsInterop": false
     },
     "@angular/material/table": {
       "src": "../../../../../node_modules/@angular/material/fesm2022/table.mjs",
       "file": "@angular_material_table.js",
-      "fileHash": "734ccf15",
+      "fileHash": "c78a17e5",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "6f1b893b",
+      "fileHash": "07bf2a45",
       "needsInterop": false
     },
     "@angular/platform-browser/animations/async": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/animations/async.mjs",
       "file": "@angular_platform-browser_animations_async.js",
-      "fileHash": "e8613cdc",
+      "fileHash": "e2f882b1",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "ef8ce38e",
+      "fileHash": "2287bc6c",
       "needsInterop": false
     },
     "@auth0/angular-jwt": {
       "src": "../../../../../node_modules/@auth0/angular-jwt/fesm2020/auth0-angular-jwt.mjs",
       "file": "@auth0_angular-jwt.js",
-      "fileHash": "7828d9a1",
+      "fileHash": "57c1b405",
       "needsInterop": false
     },
     "rxjs": {
       "src": "../../../../../node_modules/rxjs/dist/esm5/index.js",
       "file": "rxjs.js",
-      "fileHash": "4bac04a9",
+      "fileHash": "533c92ad",
       "needsInterop": false
     },
     "sweetalert": {
       "src": "../../../../../node_modules/sweetalert/dist/sweetalert.min.js",
       "file": "sweetalert.js",
-      "fileHash": "7aa0d3c9",
+      "fileHash": "9bfe2490",
       "needsInterop": true
     }
   },
   "chunks": {
+    "browser-7WVVOWIQ": {
+      "file": "browser-7WVVOWIQ.js"
+    },
     "chunk-ARW74JFL": {
       "file": "chunk-ARW74JFL.js"
-    },
-    "browser-BVPPLFFG": {
-      "file": "browser-BVPPLFFG.js"
     },
     "chunk-VWEJHRNV": {
       "file": "chunk-VWEJHRNV.js"
     },
+    "chunk-RM7V6FV7": {
+      "file": "chunk-RM7V6FV7.js"
+    },
     "chunk-A5OTKC33": {
       "file": "chunk-A5OTKC33.js"
     },
-    "chunk-C2Z32BJ6": {
-      "file": "chunk-C2Z32BJ6.js"
+    "chunk-SXOQ3UHC": {
+      "file": "chunk-SXOQ3UHC.js"
     },
-    "chunk-2AXA5NS2": {
-      "file": "chunk-2AXA5NS2.js"
+    "chunk-VNW2MBA7": {
+      "file": "chunk-VNW2MBA7.js"
     },
-    "chunk-ATNDQTJA": {
-      "file": "chunk-ATNDQTJA.js"
+    "chunk-5H5WHFDB": {
+      "file": "chunk-5H5WHFDB.js"
     },
-    "chunk-FLGVYJXI": {
-      "file": "chunk-FLGVYJXI.js"
+    "chunk-7ZFYYFG2": {
+      "file": "chunk-7ZFYYFG2.js"
     },
     "chunk-7GFUNZUG": {
       "file": "chunk-7GFUNZUG.js"
