+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/browser-7WVVOWIQ.js
@@ -6,7 +6,7 @@ import {
   sequence,
   style,
   ɵPRE_STYLE
-} from "./chunk-2AXA5NS2.js";
+} from "./chunk-VNW2MBA7.js";
 import "./chunk-7GFUNZUG.js";
 import {
   Injectable,
@@ -4061,4 +4061,4 @@ export {
    * License: MIT
    *)
 */
-//# sourceMappingURL=browser-BVPPLFFG.js.map
+//# sourceMappingURL=browser-7WVVOWIQ.js.map
