+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/chunk-7ZFYYFG2.js
@@ -86,164 +86,6 @@ import {
   __spreadValues
 } from "./chunk-WKYGNSYM.js";
 
-// node_modules/@angular/cdk/fesm2022/bidi.mjs
-var DIR_DOCUMENT = new InjectionToken("cdk-dir-doc", {
-  providedIn: "root",
-  factory: DIR_DOCUMENT_FACTORY
-});
-function DIR_DOCUMENT_FACTORY() {
-  return inject(DOCUMENT);
-}
-var RTL_LOCALE_PATTERN = /^(ar|ckb|dv|he|iw|fa|nqo|ps|sd|ug|ur|yi|.*[-_](Adlm|Arab|Hebr|Nkoo|Rohg|Thaa))(?!.*[-_](Latn|Cyrl)($|-|_))($|-|_)/i;
-function _resolveDirectionality(rawValue) {
-  const value = rawValue?.toLowerCase() || "";
-  if (value === "auto" && typeof navigator !== "undefined" && navigator?.language) {
-    return RTL_LOCALE_PATTERN.test(navigator.language) ? "rtl" : "ltr";
-  }
-  return value === "rtl" ? "rtl" : "ltr";
-}
-var _Directionality = class _Directionality {
-  constructor(_document) {
-    this.value = "ltr";
-    this.change = new EventEmitter();
-    if (_document) {
-      const bodyDir = _document.body ? _document.body.dir : null;
-      const htmlDir = _document.documentElement ? _document.documentElement.dir : null;
-      this.value = _resolveDirectionality(bodyDir || htmlDir || "ltr");
-    }
-  }
-  ngOnDestroy() {
-    this.change.complete();
-  }
-};
-_Directionality.ɵfac = function Directionality_Factory(t) {
-  return new (t || _Directionality)(ɵɵinject(DIR_DOCUMENT, 8));
-};
-_Directionality.ɵprov = ɵɵdefineInjectable({
-  token: _Directionality,
-  factory: _Directionality.ɵfac,
-  providedIn: "root"
-});
-var Directionality = _Directionality;
-(() => {
-  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Directionality, [{
-    type: Injectable,
-    args: [{
-      providedIn: "root"
-    }]
-  }], () => [{
-    type: void 0,
-    decorators: [{
-      type: Optional
-    }, {
-      type: Inject,
-      args: [DIR_DOCUMENT]
-    }]
-  }], null);
-})();
-var _Dir = class _Dir {
-  constructor() {
-    this._dir = "ltr";
-    this._isInitialized = false;
-    this.change = new EventEmitter();
-  }
-  /** @docs-private */
-  get dir() {
-    return this._dir;
-  }
-  set dir(value) {
-    const previousValue = this._dir;
-    this._dir = _resolveDirectionality(value);
-    this._rawDir = value;
-    if (previousValue !== this._dir && this._isInitialized) {
-      this.change.emit(this._dir);
-    }
-  }
-  /** Current layout direction of the element. */
-  get value() {
-    return this.dir;
-  }
-  /** Initialize once default value has been set. */
-  ngAfterContentInit() {
-    this._isInitialized = true;
-  }
-  ngOnDestroy() {
-    this.change.complete();
-  }
-};
-_Dir.ɵfac = function Dir_Factory(t) {
-  return new (t || _Dir)();
-};
-_Dir.ɵdir = ɵɵdefineDirective({
-  type: _Dir,
-  selectors: [["", "dir", ""]],
-  hostVars: 1,
-  hostBindings: function Dir_HostBindings(rf, ctx) {
-    if (rf & 2) {
-      ɵɵattribute("dir", ctx._rawDir);
-    }
-  },
-  inputs: {
-    dir: "dir"
-  },
-  outputs: {
-    change: "dirChange"
-  },
-  exportAs: ["dir"],
-  standalone: true,
-  features: [ɵɵProvidersFeature([{
-    provide: Directionality,
-    useExisting: _Dir
-  }])]
-});
-var Dir = _Dir;
-(() => {
-  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Dir, [{
-    type: Directive,
-    args: [{
-      selector: "[dir]",
-      providers: [{
-        provide: Directionality,
-        useExisting: Dir
-      }],
-      host: {
-        "[attr.dir]": "_rawDir"
-      },
-      exportAs: "dir",
-      standalone: true
-    }]
-  }], null, {
-    change: [{
-      type: Output,
-      args: ["dirChange"]
-    }],
-    dir: [{
-      type: Input
-    }]
-  });
-})();
-var _BidiModule = class _BidiModule {
-};
-_BidiModule.ɵfac = function BidiModule_Factory(t) {
-  return new (t || _BidiModule)();
-};
-_BidiModule.ɵmod = ɵɵdefineNgModule({
-  type: _BidiModule,
-  imports: [Dir],
-  exports: [Dir]
-});
-_BidiModule.ɵinj = ɵɵdefineInjector({});
-var BidiModule = _BidiModule;
-(() => {
-  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(BidiModule, [{
-    type: NgModule,
-    args: [{
-      imports: [Dir],
-      exports: [Dir]
-    }]
-  }], null, null);
-})();
-
 // node_modules/@angular/cdk/fesm2022/platform.mjs
 var hasV8BreakIterator;
 try {
@@ -459,6 +301,35 @@ function _isTestEnvironment() {
     typeof jest !== "undefined" && !!jest || // @ts-ignore
     typeof Mocha !== "undefined" && !!Mocha
   );
+}
+
+// node_modules/@angular/cdk/fesm2022/keycodes.mjs
+var TAB = 9;
+var ENTER = 13;
+var SHIFT = 16;
+var CONTROL = 17;
+var ALT = 18;
+var ESCAPE = 27;
+var SPACE = 32;
+var PAGE_UP = 33;
+var PAGE_DOWN = 34;
+var END = 35;
+var HOME = 36;
+var LEFT_ARROW = 37;
+var UP_ARROW = 38;
+var RIGHT_ARROW = 39;
+var DOWN_ARROW = 40;
+var ZERO = 48;
+var NINE = 57;
+var A = 65;
+var Z = 90;
+var META = 91;
+var MAC_META = 224;
+function hasModifierKey(event, ...modifiers) {
+  if (modifiers.length) {
+    return modifiers.some((modifier) => event[modifier]);
+  }
+  return event.altKey || event.shiftKey || event.ctrlKey || event.metaKey;
 }
 
 // node_modules/@angular/cdk/fesm2022/coercion.mjs
@@ -736,35 +607,6 @@ var ObserversModule = _ObserversModule;
     }]
   }], null, null);
 })();
-
-// node_modules/@angular/cdk/fesm2022/keycodes.mjs
-var TAB = 9;
-var ENTER = 13;
-var SHIFT = 16;
-var CONTROL = 17;
-var ALT = 18;
-var ESCAPE = 27;
-var SPACE = 32;
-var PAGE_UP = 33;
-var PAGE_DOWN = 34;
-var END = 35;
-var HOME = 36;
-var LEFT_ARROW = 37;
-var UP_ARROW = 38;
-var RIGHT_ARROW = 39;
-var DOWN_ARROW = 40;
-var ZERO = 48;
-var NINE = 57;
-var A = 65;
-var Z = 90;
-var META = 91;
-var MAC_META = 224;
-function hasModifierKey(event, ...modifiers) {
-  if (modifiers.length) {
-    return modifiers.some((modifier) => event[modifier]);
-  }
-  return event.altKey || event.shiftKey || event.ctrlKey || event.metaKey;
-}
 
 // node_modules/@angular/cdk/fesm2022/layout.mjs
 var _LayoutModule = class _LayoutModule {
@@ -3045,6 +2887,164 @@ var A11yModule = _A11yModule;
   }], null);
 })();
 
+// node_modules/@angular/cdk/fesm2022/bidi.mjs
+var DIR_DOCUMENT = new InjectionToken("cdk-dir-doc", {
+  providedIn: "root",
+  factory: DIR_DOCUMENT_FACTORY
+});
+function DIR_DOCUMENT_FACTORY() {
+  return inject(DOCUMENT);
+}
+var RTL_LOCALE_PATTERN = /^(ar|ckb|dv|he|iw|fa|nqo|ps|sd|ug|ur|yi|.*[-_](Adlm|Arab|Hebr|Nkoo|Rohg|Thaa))(?!.*[-_](Latn|Cyrl)($|-|_))($|-|_)/i;
+function _resolveDirectionality(rawValue) {
+  const value = rawValue?.toLowerCase() || "";
+  if (value === "auto" && typeof navigator !== "undefined" && navigator?.language) {
+    return RTL_LOCALE_PATTERN.test(navigator.language) ? "rtl" : "ltr";
+  }
+  return value === "rtl" ? "rtl" : "ltr";
+}
+var _Directionality = class _Directionality {
+  constructor(_document) {
+    this.value = "ltr";
+    this.change = new EventEmitter();
+    if (_document) {
+      const bodyDir = _document.body ? _document.body.dir : null;
+      const htmlDir = _document.documentElement ? _document.documentElement.dir : null;
+      this.value = _resolveDirectionality(bodyDir || htmlDir || "ltr");
+    }
+  }
+  ngOnDestroy() {
+    this.change.complete();
+  }
+};
+_Directionality.ɵfac = function Directionality_Factory(t) {
+  return new (t || _Directionality)(ɵɵinject(DIR_DOCUMENT, 8));
+};
+_Directionality.ɵprov = ɵɵdefineInjectable({
+  token: _Directionality,
+  factory: _Directionality.ɵfac,
+  providedIn: "root"
+});
+var Directionality = _Directionality;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Directionality, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Optional
+    }, {
+      type: Inject,
+      args: [DIR_DOCUMENT]
+    }]
+  }], null);
+})();
+var _Dir = class _Dir {
+  constructor() {
+    this._dir = "ltr";
+    this._isInitialized = false;
+    this.change = new EventEmitter();
+  }
+  /** @docs-private */
+  get dir() {
+    return this._dir;
+  }
+  set dir(value) {
+    const previousValue = this._dir;
+    this._dir = _resolveDirectionality(value);
+    this._rawDir = value;
+    if (previousValue !== this._dir && this._isInitialized) {
+      this.change.emit(this._dir);
+    }
+  }
+  /** Current layout direction of the element. */
+  get value() {
+    return this.dir;
+  }
+  /** Initialize once default value has been set. */
+  ngAfterContentInit() {
+    this._isInitialized = true;
+  }
+  ngOnDestroy() {
+    this.change.complete();
+  }
+};
+_Dir.ɵfac = function Dir_Factory(t) {
+  return new (t || _Dir)();
+};
+_Dir.ɵdir = ɵɵdefineDirective({
+  type: _Dir,
+  selectors: [["", "dir", ""]],
+  hostVars: 1,
+  hostBindings: function Dir_HostBindings(rf, ctx) {
+    if (rf & 2) {
+      ɵɵattribute("dir", ctx._rawDir);
+    }
+  },
+  inputs: {
+    dir: "dir"
+  },
+  outputs: {
+    change: "dirChange"
+  },
+  exportAs: ["dir"],
+  standalone: true,
+  features: [ɵɵProvidersFeature([{
+    provide: Directionality,
+    useExisting: _Dir
+  }])]
+});
+var Dir = _Dir;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Dir, [{
+    type: Directive,
+    args: [{
+      selector: "[dir]",
+      providers: [{
+        provide: Directionality,
+        useExisting: Dir
+      }],
+      host: {
+        "[attr.dir]": "_rawDir"
+      },
+      exportAs: "dir",
+      standalone: true
+    }]
+  }], null, {
+    change: [{
+      type: Output,
+      args: ["dirChange"]
+    }],
+    dir: [{
+      type: Input
+    }]
+  });
+})();
+var _BidiModule = class _BidiModule {
+};
+_BidiModule.ɵfac = function BidiModule_Factory(t) {
+  return new (t || _BidiModule)();
+};
+_BidiModule.ɵmod = ɵɵdefineNgModule({
+  type: _BidiModule,
+  imports: [Dir],
+  exports: [Dir]
+});
+_BidiModule.ɵinj = ɵɵdefineInjector({});
+var BidiModule = _BidiModule;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(BidiModule, [{
+    type: NgModule,
+    args: [{
+      imports: [Dir],
+      exports: [Dir]
+    }]
+  }], null, null);
+})();
+
 // node_modules/@angular/cdk/fesm2022/cdk.mjs
 var VERSION = new Version("17.3.10");
 
@@ -4988,23 +4988,15 @@ var _MatInternalFormField = __MatInternalFormField;
 })();
 
 export {
-  Directionality,
-  BidiModule,
   Platform,
   getSupportedInputTypes,
   normalizePassiveListenerOptions,
   RtlScrollAxisType,
   supportsScrollBehavior,
   getRtlScrollAxisType,
+  _getFocusedElementPierceShadowDom,
   _getEventTarget,
   _isTestEnvironment,
-  coerceBooleanProperty,
-  coerceNumberProperty,
-  _isNumberValue,
-  coerceArray,
-  coerceCssPixelValue,
-  coerceElement,
-  ObserversModule,
   ENTER,
   ESCAPE,
   SPACE,
@@ -5014,11 +5006,23 @@ export {
   DOWN_ARROW,
   A,
   hasModifierKey,
+  coerceBooleanProperty,
+  coerceNumberProperty,
+  _isNumberValue,
+  coerceArray,
+  coerceCssPixelValue,
+  coerceElement,
+  ObserversModule,
   addAriaReferencedId,
   removeAriaReferencedId,
   ActiveDescendantKeyManager,
+  InteractivityChecker,
+  FocusTrapFactory,
   LiveAnnouncer,
   FocusMonitor,
+  A11yModule,
+  Directionality,
+  BidiModule,
   MatCommonModule,
   _ErrorStateTracker,
   ErrorStateMatcher,
@@ -5032,4 +5036,4 @@ export {
   MatOptionModule,
   MatRippleLoader
 };
-//# sourceMappingURL=chunk-FLGVYJXI.js.map
+//# sourceMappingURL=chunk-7ZFYYFG2.js.map
