+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/chunk-RM7V6FV7.js
@@ -1,0 +1,3316 @@
+import {
+  ScrollDispatcher,
+  ScrollingModule,
+  ViewportRuler
+} from "./chunk-5H5WHFDB.js";
+import {
+  BidiModule,
+  Directionality,
+  ESCAPE,
+  Platform,
+  _getEventTarget,
+  _isTestEnvironment,
+  coerceArray,
+  coerceCssPixelValue,
+  hasModifierKey,
+  supportsScrollBehavior
+} from "./chunk-7ZFYYFG2.js";
+import {
+  DOCUMENT,
+  Location
+} from "./chunk-7GFUNZUG.js";
+import {
+  ANIMATION_MODULE_TYPE,
+  ApplicationRef,
+  ComponentFactoryResolver$1,
+  Directive,
+  ElementRef,
+  EventEmitter,
+  Inject,
+  Injectable,
+  InjectionToken,
+  Injector,
+  Input,
+  InputFlags,
+  NgModule,
+  NgZone,
+  Optional,
+  Output,
+  TemplateRef,
+  ViewContainerRef,
+  booleanAttribute,
+  inject,
+  setClassMetadata,
+  ɵɵInheritDefinitionFeature,
+  ɵɵInputTransformsFeature,
+  ɵɵNgOnChangesFeature,
+  ɵɵProvidersFeature,
+  ɵɵdefineDirective,
+  ɵɵdefineInjectable,
+  ɵɵdefineInjector,
+  ɵɵdefineNgModule,
+  ɵɵdirectiveInject,
+  ɵɵgetInheritedFactory,
+  ɵɵinject
+} from "./chunk-RUPN7I6S.js";
+import {
+  Subject,
+  Subscription,
+  filter,
+  merge,
+  take,
+  takeUntil,
+  takeWhile
+} from "./chunk-4J25ECOH.js";
+import {
+  __spreadProps,
+  __spreadValues
+} from "./chunk-WKYGNSYM.js";
+
+// node_modules/@angular/cdk/fesm2022/portal.mjs
+function throwNullPortalError() {
+  throw Error("Must provide a portal to attach");
+}
+function throwPortalAlreadyAttachedError() {
+  throw Error("Host already has a portal attached");
+}
+function throwPortalOutletAlreadyDisposedError() {
+  throw Error("This PortalOutlet has already been disposed");
+}
+function throwUnknownPortalTypeError() {
+  throw Error("Attempting to attach an unknown Portal type. BasePortalOutlet accepts either a ComponentPortal or a TemplatePortal.");
+}
+function throwNullPortalOutletError() {
+  throw Error("Attempting to attach a portal to a null PortalOutlet");
+}
+function throwNoPortalAttachedError() {
+  throw Error("Attempting to detach a portal that is not attached to a host");
+}
+var Portal = class {
+  /** Attach this portal to a host. */
+  attach(host) {
+    if (typeof ngDevMode === "undefined" || ngDevMode) {
+      if (host == null) {
+        throwNullPortalOutletError();
+      }
+      if (host.hasAttached()) {
+        throwPortalAlreadyAttachedError();
+      }
+    }
+    this._attachedHost = host;
+    return host.attach(this);
+  }
+  /** Detach this portal from its host */
+  detach() {
+    let host = this._attachedHost;
+    if (host != null) {
+      this._attachedHost = null;
+      host.detach();
+    } else if (typeof ngDevMode === "undefined" || ngDevMode) {
+      throwNoPortalAttachedError();
+    }
+  }
+  /** Whether this portal is attached to a host. */
+  get isAttached() {
+    return this._attachedHost != null;
+  }
+  /**
+   * Sets the PortalOutlet reference without performing `attach()`. This is used directly by
+   * the PortalOutlet when it is performing an `attach()` or `detach()`.
+   */
+  setAttachedHost(host) {
+    this._attachedHost = host;
+  }
+};
+var ComponentPortal = class extends Portal {
+  constructor(component, viewContainerRef, injector, componentFactoryResolver, projectableNodes) {
+    super();
+    this.component = component;
+    this.viewContainerRef = viewContainerRef;
+    this.injector = injector;
+    this.componentFactoryResolver = componentFactoryResolver;
+    this.projectableNodes = projectableNodes;
+  }
+};
+var TemplatePortal = class extends Portal {
+  constructor(templateRef, viewContainerRef, context, injector) {
+    super();
+    this.templateRef = templateRef;
+    this.viewContainerRef = viewContainerRef;
+    this.context = context;
+    this.injector = injector;
+  }
+  get origin() {
+    return this.templateRef.elementRef;
+  }
+  /**
+   * Attach the portal to the provided `PortalOutlet`.
+   * When a context is provided it will override the `context` property of the `TemplatePortal`
+   * instance.
+   */
+  attach(host, context = this.context) {
+    this.context = context;
+    return super.attach(host);
+  }
+  detach() {
+    this.context = void 0;
+    return super.detach();
+  }
+};
+var DomPortal = class extends Portal {
+  constructor(element) {
+    super();
+    this.element = element instanceof ElementRef ? element.nativeElement : element;
+  }
+};
+var BasePortalOutlet = class {
+  constructor() {
+    this._isDisposed = false;
+    this.attachDomPortal = null;
+  }
+  /** Whether this host has an attached portal. */
+  hasAttached() {
+    return !!this._attachedPortal;
+  }
+  /** Attaches a portal. */
+  attach(portal) {
+    if (typeof ngDevMode === "undefined" || ngDevMode) {
+      if (!portal) {
+        throwNullPortalError();
+      }
+      if (this.hasAttached()) {
+        throwPortalAlreadyAttachedError();
+      }
+      if (this._isDisposed) {
+        throwPortalOutletAlreadyDisposedError();
+      }
+    }
+    if (portal instanceof ComponentPortal) {
+      this._attachedPortal = portal;
+      return this.attachComponentPortal(portal);
+    } else if (portal instanceof TemplatePortal) {
+      this._attachedPortal = portal;
+      return this.attachTemplatePortal(portal);
+    } else if (this.attachDomPortal && portal instanceof DomPortal) {
+      this._attachedPortal = portal;
+      return this.attachDomPortal(portal);
+    }
+    if (typeof ngDevMode === "undefined" || ngDevMode) {
+      throwUnknownPortalTypeError();
+    }
+  }
+  /** Detaches a previously attached portal. */
+  detach() {
+    if (this._attachedPortal) {
+      this._attachedPortal.setAttachedHost(null);
+      this._attachedPortal = null;
+    }
+    this._invokeDisposeFn();
+  }
+  /** Permanently dispose of this portal host. */
+  dispose() {
+    if (this.hasAttached()) {
+      this.detach();
+    }
+    this._invokeDisposeFn();
+    this._isDisposed = true;
+  }
+  /** @docs-private */
+  setDisposeFn(fn) {
+    this._disposeFn = fn;
+  }
+  _invokeDisposeFn() {
+    if (this._disposeFn) {
+      this._disposeFn();
+      this._disposeFn = null;
+    }
+  }
+};
+var DomPortalOutlet = class extends BasePortalOutlet {
+  /**
+   * @param outletElement Element into which the content is projected.
+   * @param _componentFactoryResolver Used to resolve the component factory.
+   *   Only required when attaching component portals.
+   * @param _appRef Reference to the application. Only used in component portals when there
+   *   is no `ViewContainerRef` available.
+   * @param _defaultInjector Injector to use as a fallback when the portal being attached doesn't
+   *   have one. Only used for component portals.
+   * @param _document Reference to the document. Used when attaching a DOM portal. Will eventually
+   *   become a required parameter.
+   */
+  constructor(outletElement, _componentFactoryResolver, _appRef, _defaultInjector, _document) {
+    super();
+    this.outletElement = outletElement;
+    this._componentFactoryResolver = _componentFactoryResolver;
+    this._appRef = _appRef;
+    this._defaultInjector = _defaultInjector;
+    this.attachDomPortal = (portal) => {
+      if (!this._document && (typeof ngDevMode === "undefined" || ngDevMode)) {
+        throw Error("Cannot attach DOM portal without _document constructor parameter");
+      }
+      const element = portal.element;
+      if (!element.parentNode && (typeof ngDevMode === "undefined" || ngDevMode)) {
+        throw Error("DOM portal content must be attached to a parent node.");
+      }
+      const anchorNode = this._document.createComment("dom-portal");
+      element.parentNode.insertBefore(anchorNode, element);
+      this.outletElement.appendChild(element);
+      this._attachedPortal = portal;
+      super.setDisposeFn(() => {
+        if (anchorNode.parentNode) {
+          anchorNode.parentNode.replaceChild(element, anchorNode);
+        }
+      });
+    };
+    this._document = _document;
+  }
+  /**
+   * Attach the given ComponentPortal to DOM element using the ComponentFactoryResolver.
+   * @param portal Portal to be attached
+   * @returns Reference to the created component.
+   */
+  attachComponentPortal(portal) {
+    const resolver = portal.componentFactoryResolver || this._componentFactoryResolver;
+    if ((typeof ngDevMode === "undefined" || ngDevMode) && !resolver) {
+      throw Error("Cannot attach component portal to outlet without a ComponentFactoryResolver.");
+    }
+    const componentFactory = resolver.resolveComponentFactory(portal.component);
+    let componentRef;
+    if (portal.viewContainerRef) {
+      componentRef = portal.viewContainerRef.createComponent(componentFactory, portal.viewContainerRef.length, portal.injector || portal.viewContainerRef.injector, portal.projectableNodes || void 0);
+      this.setDisposeFn(() => componentRef.destroy());
+    } else {
+      if ((typeof ngDevMode === "undefined" || ngDevMode) && !this._appRef) {
+        throw Error("Cannot attach component portal to outlet without an ApplicationRef.");
+      }
+      componentRef = componentFactory.create(portal.injector || this._defaultInjector || Injector.NULL);
+      this._appRef.attachView(componentRef.hostView);
+      this.setDisposeFn(() => {
+        if (this._appRef.viewCount > 0) {
+          this._appRef.detachView(componentRef.hostView);
+        }
+        componentRef.destroy();
+      });
+    }
+    this.outletElement.appendChild(this._getComponentRootNode(componentRef));
+    this._attachedPortal = portal;
+    return componentRef;
+  }
+  /**
+   * Attaches a template portal to the DOM as an embedded view.
+   * @param portal Portal to be attached.
+   * @returns Reference to the created embedded view.
+   */
+  attachTemplatePortal(portal) {
+    let viewContainer = portal.viewContainerRef;
+    let viewRef = viewContainer.createEmbeddedView(portal.templateRef, portal.context, {
+      injector: portal.injector
+    });
+    viewRef.rootNodes.forEach((rootNode) => this.outletElement.appendChild(rootNode));
+    viewRef.detectChanges();
+    this.setDisposeFn(() => {
+      let index = viewContainer.indexOf(viewRef);
+      if (index !== -1) {
+        viewContainer.remove(index);
+      }
+    });
+    this._attachedPortal = portal;
+    return viewRef;
+  }
+  /**
+   * Clears out a portal from the DOM.
+   */
+  dispose() {
+    super.dispose();
+    this.outletElement.remove();
+  }
+  /** Gets the root HTMLElement for an instantiated component. */
+  _getComponentRootNode(componentRef) {
+    return componentRef.hostView.rootNodes[0];
+  }
+};
+var _CdkPortal = class _CdkPortal extends TemplatePortal {
+  constructor(templateRef, viewContainerRef) {
+    super(templateRef, viewContainerRef);
+  }
+};
+_CdkPortal.ɵfac = function CdkPortal_Factory(t) {
+  return new (t || _CdkPortal)(ɵɵdirectiveInject(TemplateRef), ɵɵdirectiveInject(ViewContainerRef));
+};
+_CdkPortal.ɵdir = ɵɵdefineDirective({
+  type: _CdkPortal,
+  selectors: [["", "cdkPortal", ""]],
+  exportAs: ["cdkPortal"],
+  standalone: true,
+  features: [ɵɵInheritDefinitionFeature]
+});
+var CdkPortal = _CdkPortal;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(CdkPortal, [{
+    type: Directive,
+    args: [{
+      selector: "[cdkPortal]",
+      exportAs: "cdkPortal",
+      standalone: true
+    }]
+  }], () => [{
+    type: TemplateRef
+  }, {
+    type: ViewContainerRef
+  }], null);
+})();
+var _TemplatePortalDirective = class _TemplatePortalDirective extends CdkPortal {
+};
+_TemplatePortalDirective.ɵfac = /* @__PURE__ */ (() => {
+  let ɵTemplatePortalDirective_BaseFactory;
+  return function TemplatePortalDirective_Factory(t) {
+    return (ɵTemplatePortalDirective_BaseFactory || (ɵTemplatePortalDirective_BaseFactory = ɵɵgetInheritedFactory(_TemplatePortalDirective)))(t || _TemplatePortalDirective);
+  };
+})();
+_TemplatePortalDirective.ɵdir = ɵɵdefineDirective({
+  type: _TemplatePortalDirective,
+  selectors: [["", "cdk-portal", ""], ["", "portal", ""]],
+  exportAs: ["cdkPortal"],
+  standalone: true,
+  features: [ɵɵProvidersFeature([{
+    provide: CdkPortal,
+    useExisting: _TemplatePortalDirective
+  }]), ɵɵInheritDefinitionFeature]
+});
+var TemplatePortalDirective = _TemplatePortalDirective;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(TemplatePortalDirective, [{
+    type: Directive,
+    args: [{
+      selector: "[cdk-portal], [portal]",
+      exportAs: "cdkPortal",
+      providers: [{
+        provide: CdkPortal,
+        useExisting: TemplatePortalDirective
+      }],
+      standalone: true
+    }]
+  }], null, null);
+})();
+var _CdkPortalOutlet = class _CdkPortalOutlet extends BasePortalOutlet {
+  constructor(_componentFactoryResolver, _viewContainerRef, _document) {
+    super();
+    this._componentFactoryResolver = _componentFactoryResolver;
+    this._viewContainerRef = _viewContainerRef;
+    this._isInitialized = false;
+    this.attached = new EventEmitter();
+    this.attachDomPortal = (portal) => {
+      if (!this._document && (typeof ngDevMode === "undefined" || ngDevMode)) {
+        throw Error("Cannot attach DOM portal without _document constructor parameter");
+      }
+      const element = portal.element;
+      if (!element.parentNode && (typeof ngDevMode === "undefined" || ngDevMode)) {
+        throw Error("DOM portal content must be attached to a parent node.");
+      }
+      const anchorNode = this._document.createComment("dom-portal");
+      portal.setAttachedHost(this);
+      element.parentNode.insertBefore(anchorNode, element);
+      this._getRootNode().appendChild(element);
+      this._attachedPortal = portal;
+      super.setDisposeFn(() => {
+        if (anchorNode.parentNode) {
+          anchorNode.parentNode.replaceChild(element, anchorNode);
+        }
+      });
+    };
+    this._document = _document;
+  }
+  /** Portal associated with the Portal outlet. */
+  get portal() {
+    return this._attachedPortal;
+  }
+  set portal(portal) {
+    if (this.hasAttached() && !portal && !this._isInitialized) {
+      return;
+    }
+    if (this.hasAttached()) {
+      super.detach();
+    }
+    if (portal) {
+      super.attach(portal);
+    }
+    this._attachedPortal = portal || null;
+  }
+  /** Component or view reference that is attached to the portal. */
+  get attachedRef() {
+    return this._attachedRef;
+  }
+  ngOnInit() {
+    this._isInitialized = true;
+  }
+  ngOnDestroy() {
+    super.dispose();
+    this._attachedRef = this._attachedPortal = null;
+  }
+  /**
+   * Attach the given ComponentPortal to this PortalOutlet using the ComponentFactoryResolver.
+   *
+   * @param portal Portal to be attached to the portal outlet.
+   * @returns Reference to the created component.
+   */
+  attachComponentPortal(portal) {
+    portal.setAttachedHost(this);
+    const viewContainerRef = portal.viewContainerRef != null ? portal.viewContainerRef : this._viewContainerRef;
+    const resolver = portal.componentFactoryResolver || this._componentFactoryResolver;
+    const componentFactory = resolver.resolveComponentFactory(portal.component);
+    const ref = viewContainerRef.createComponent(componentFactory, viewContainerRef.length, portal.injector || viewContainerRef.injector, portal.projectableNodes || void 0);
+    if (viewContainerRef !== this._viewContainerRef) {
+      this._getRootNode().appendChild(ref.hostView.rootNodes[0]);
+    }
+    super.setDisposeFn(() => ref.destroy());
+    this._attachedPortal = portal;
+    this._attachedRef = ref;
+    this.attached.emit(ref);
+    return ref;
+  }
+  /**
+   * Attach the given TemplatePortal to this PortalHost as an embedded View.
+   * @param portal Portal to be attached.
+   * @returns Reference to the created embedded view.
+   */
+  attachTemplatePortal(portal) {
+    portal.setAttachedHost(this);
+    const viewRef = this._viewContainerRef.createEmbeddedView(portal.templateRef, portal.context, {
+      injector: portal.injector
+    });
+    super.setDisposeFn(() => this._viewContainerRef.clear());
+    this._attachedPortal = portal;
+    this._attachedRef = viewRef;
+    this.attached.emit(viewRef);
+    return viewRef;
+  }
+  /** Gets the root node of the portal outlet. */
+  _getRootNode() {
+    const nativeElement = this._viewContainerRef.element.nativeElement;
+    return nativeElement.nodeType === nativeElement.ELEMENT_NODE ? nativeElement : nativeElement.parentNode;
+  }
+};
+_CdkPortalOutlet.ɵfac = function CdkPortalOutlet_Factory(t) {
+  return new (t || _CdkPortalOutlet)(ɵɵdirectiveInject(ComponentFactoryResolver$1), ɵɵdirectiveInject(ViewContainerRef), ɵɵdirectiveInject(DOCUMENT));
+};
+_CdkPortalOutlet.ɵdir = ɵɵdefineDirective({
+  type: _CdkPortalOutlet,
+  selectors: [["", "cdkPortalOutlet", ""]],
+  inputs: {
+    portal: [InputFlags.None, "cdkPortalOutlet", "portal"]
+  },
+  outputs: {
+    attached: "attached"
+  },
+  exportAs: ["cdkPortalOutlet"],
+  standalone: true,
+  features: [ɵɵInheritDefinitionFeature]
+});
+var CdkPortalOutlet = _CdkPortalOutlet;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(CdkPortalOutlet, [{
+    type: Directive,
+    args: [{
+      selector: "[cdkPortalOutlet]",
+      exportAs: "cdkPortalOutlet",
+      standalone: true
+    }]
+  }], () => [{
+    type: ComponentFactoryResolver$1
+  }, {
+    type: ViewContainerRef
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }], {
+    portal: [{
+      type: Input,
+      args: ["cdkPortalOutlet"]
+    }],
+    attached: [{
+      type: Output
+    }]
+  });
+})();
+var _PortalHostDirective = class _PortalHostDirective extends CdkPortalOutlet {
+};
+_PortalHostDirective.ɵfac = /* @__PURE__ */ (() => {
+  let ɵPortalHostDirective_BaseFactory;
+  return function PortalHostDirective_Factory(t) {
+    return (ɵPortalHostDirective_BaseFactory || (ɵPortalHostDirective_BaseFactory = ɵɵgetInheritedFactory(_PortalHostDirective)))(t || _PortalHostDirective);
+  };
+})();
+_PortalHostDirective.ɵdir = ɵɵdefineDirective({
+  type: _PortalHostDirective,
+  selectors: [["", "cdkPortalHost", ""], ["", "portalHost", ""]],
+  inputs: {
+    portal: [InputFlags.None, "cdkPortalHost", "portal"]
+  },
+  exportAs: ["cdkPortalHost"],
+  standalone: true,
+  features: [ɵɵProvidersFeature([{
+    provide: CdkPortalOutlet,
+    useExisting: _PortalHostDirective
+  }]), ɵɵInheritDefinitionFeature]
+});
+var PortalHostDirective = _PortalHostDirective;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(PortalHostDirective, [{
+    type: Directive,
+    args: [{
+      selector: "[cdkPortalHost], [portalHost]",
+      exportAs: "cdkPortalHost",
+      inputs: [{
+        name: "portal",
+        alias: "cdkPortalHost"
+      }],
+      providers: [{
+        provide: CdkPortalOutlet,
+        useExisting: PortalHostDirective
+      }],
+      standalone: true
+    }]
+  }], null, null);
+})();
+var _PortalModule = class _PortalModule {
+};
+_PortalModule.ɵfac = function PortalModule_Factory(t) {
+  return new (t || _PortalModule)();
+};
+_PortalModule.ɵmod = ɵɵdefineNgModule({
+  type: _PortalModule,
+  imports: [CdkPortal, CdkPortalOutlet, TemplatePortalDirective, PortalHostDirective],
+  exports: [CdkPortal, CdkPortalOutlet, TemplatePortalDirective, PortalHostDirective]
+});
+_PortalModule.ɵinj = ɵɵdefineInjector({});
+var PortalModule = _PortalModule;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(PortalModule, [{
+    type: NgModule,
+    args: [{
+      imports: [CdkPortal, CdkPortalOutlet, TemplatePortalDirective, PortalHostDirective],
+      exports: [CdkPortal, CdkPortalOutlet, TemplatePortalDirective, PortalHostDirective]
+    }]
+  }], null, null);
+})();
+
+// node_modules/@angular/cdk/fesm2022/overlay.mjs
+var scrollBehaviorSupported = supportsScrollBehavior();
+var BlockScrollStrategy = class {
+  constructor(_viewportRuler, document) {
+    this._viewportRuler = _viewportRuler;
+    this._previousHTMLStyles = {
+      top: "",
+      left: ""
+    };
+    this._isEnabled = false;
+    this._document = document;
+  }
+  /** Attaches this scroll strategy to an overlay. */
+  attach() {
+  }
+  /** Blocks page-level scroll while the attached overlay is open. */
+  enable() {
+    if (this._canBeEnabled()) {
+      const root = this._document.documentElement;
+      this._previousScrollPosition = this._viewportRuler.getViewportScrollPosition();
+      this._previousHTMLStyles.left = root.style.left || "";
+      this._previousHTMLStyles.top = root.style.top || "";
+      root.style.left = coerceCssPixelValue(-this._previousScrollPosition.left);
+      root.style.top = coerceCssPixelValue(-this._previousScrollPosition.top);
+      root.classList.add("cdk-global-scrollblock");
+      this._isEnabled = true;
+    }
+  }
+  /** Unblocks page-level scroll while the attached overlay is open. */
+  disable() {
+    if (this._isEnabled) {
+      const html = this._document.documentElement;
+      const body = this._document.body;
+      const htmlStyle = html.style;
+      const bodyStyle = body.style;
+      const previousHtmlScrollBehavior = htmlStyle.scrollBehavior || "";
+      const previousBodyScrollBehavior = bodyStyle.scrollBehavior || "";
+      this._isEnabled = false;
+      htmlStyle.left = this._previousHTMLStyles.left;
+      htmlStyle.top = this._previousHTMLStyles.top;
+      html.classList.remove("cdk-global-scrollblock");
+      if (scrollBehaviorSupported) {
+        htmlStyle.scrollBehavior = bodyStyle.scrollBehavior = "auto";
+      }
+      window.scroll(this._previousScrollPosition.left, this._previousScrollPosition.top);
+      if (scrollBehaviorSupported) {
+        htmlStyle.scrollBehavior = previousHtmlScrollBehavior;
+        bodyStyle.scrollBehavior = previousBodyScrollBehavior;
+      }
+    }
+  }
+  _canBeEnabled() {
+    const html = this._document.documentElement;
+    if (html.classList.contains("cdk-global-scrollblock") || this._isEnabled) {
+      return false;
+    }
+    const body = this._document.body;
+    const viewport = this._viewportRuler.getViewportSize();
+    return body.scrollHeight > viewport.height || body.scrollWidth > viewport.width;
+  }
+};
+function getMatScrollStrategyAlreadyAttachedError() {
+  return Error(`Scroll strategy has already been attached.`);
+}
+var CloseScrollStrategy = class {
+  constructor(_scrollDispatcher, _ngZone, _viewportRuler, _config) {
+    this._scrollDispatcher = _scrollDispatcher;
+    this._ngZone = _ngZone;
+    this._viewportRuler = _viewportRuler;
+    this._config = _config;
+    this._scrollSubscription = null;
+    this._detach = () => {
+      this.disable();
+      if (this._overlayRef.hasAttached()) {
+        this._ngZone.run(() => this._overlayRef.detach());
+      }
+    };
+  }
+  /** Attaches this scroll strategy to an overlay. */
+  attach(overlayRef) {
+    if (this._overlayRef && (typeof ngDevMode === "undefined" || ngDevMode)) {
+      throw getMatScrollStrategyAlreadyAttachedError();
+    }
+    this._overlayRef = overlayRef;
+  }
+  /** Enables the closing of the attached overlay on scroll. */
+  enable() {
+    if (this._scrollSubscription) {
+      return;
+    }
+    const stream = this._scrollDispatcher.scrolled(0).pipe(filter((scrollable) => {
+      return !scrollable || !this._overlayRef.overlayElement.contains(scrollable.getElementRef().nativeElement);
+    }));
+    if (this._config && this._config.threshold && this._config.threshold > 1) {
+      this._initialScrollPosition = this._viewportRuler.getViewportScrollPosition().top;
+      this._scrollSubscription = stream.subscribe(() => {
+        const scrollPosition = this._viewportRuler.getViewportScrollPosition().top;
+        if (Math.abs(scrollPosition - this._initialScrollPosition) > this._config.threshold) {
+          this._detach();
+        } else {
+          this._overlayRef.updatePosition();
+        }
+      });
+    } else {
+      this._scrollSubscription = stream.subscribe(this._detach);
+    }
+  }
+  /** Disables the closing the attached overlay on scroll. */
+  disable() {
+    if (this._scrollSubscription) {
+      this._scrollSubscription.unsubscribe();
+      this._scrollSubscription = null;
+    }
+  }
+  detach() {
+    this.disable();
+    this._overlayRef = null;
+  }
+};
+var NoopScrollStrategy = class {
+  /** Does nothing, as this scroll strategy is a no-op. */
+  enable() {
+  }
+  /** Does nothing, as this scroll strategy is a no-op. */
+  disable() {
+  }
+  /** Does nothing, as this scroll strategy is a no-op. */
+  attach() {
+  }
+};
+function isElementScrolledOutsideView(element, scrollContainers) {
+  return scrollContainers.some((containerBounds) => {
+    const outsideAbove = element.bottom < containerBounds.top;
+    const outsideBelow = element.top > containerBounds.bottom;
+    const outsideLeft = element.right < containerBounds.left;
+    const outsideRight = element.left > containerBounds.right;
+    return outsideAbove || outsideBelow || outsideLeft || outsideRight;
+  });
+}
+function isElementClippedByScrolling(element, scrollContainers) {
+  return scrollContainers.some((scrollContainerRect) => {
+    const clippedAbove = element.top < scrollContainerRect.top;
+    const clippedBelow = element.bottom > scrollContainerRect.bottom;
+    const clippedLeft = element.left < scrollContainerRect.left;
+    const clippedRight = element.right > scrollContainerRect.right;
+    return clippedAbove || clippedBelow || clippedLeft || clippedRight;
+  });
+}
+var RepositionScrollStrategy = class {
+  constructor(_scrollDispatcher, _viewportRuler, _ngZone, _config) {
+    this._scrollDispatcher = _scrollDispatcher;
+    this._viewportRuler = _viewportRuler;
+    this._ngZone = _ngZone;
+    this._config = _config;
+    this._scrollSubscription = null;
+  }
+  /** Attaches this scroll strategy to an overlay. */
+  attach(overlayRef) {
+    if (this._overlayRef && (typeof ngDevMode === "undefined" || ngDevMode)) {
+      throw getMatScrollStrategyAlreadyAttachedError();
+    }
+    this._overlayRef = overlayRef;
+  }
+  /** Enables repositioning of the attached overlay on scroll. */
+  enable() {
+    if (!this._scrollSubscription) {
+      const throttle = this._config ? this._config.scrollThrottle : 0;
+      this._scrollSubscription = this._scrollDispatcher.scrolled(throttle).subscribe(() => {
+        this._overlayRef.updatePosition();
+        if (this._config && this._config.autoClose) {
+          const overlayRect = this._overlayRef.overlayElement.getBoundingClientRect();
+          const {
+            width,
+            height
+          } = this._viewportRuler.getViewportSize();
+          const parentRects = [{
+            width,
+            height,
+            bottom: height,
+            right: width,
+            top: 0,
+            left: 0
+          }];
+          if (isElementScrolledOutsideView(overlayRect, parentRects)) {
+            this.disable();
+            this._ngZone.run(() => this._overlayRef.detach());
+          }
+        }
+      });
+    }
+  }
+  /** Disables repositioning of the attached overlay on scroll. */
+  disable() {
+    if (this._scrollSubscription) {
+      this._scrollSubscription.unsubscribe();
+      this._scrollSubscription = null;
+    }
+  }
+  detach() {
+    this.disable();
+    this._overlayRef = null;
+  }
+};
+var _ScrollStrategyOptions = class _ScrollStrategyOptions {
+  constructor(_scrollDispatcher, _viewportRuler, _ngZone, document) {
+    this._scrollDispatcher = _scrollDispatcher;
+    this._viewportRuler = _viewportRuler;
+    this._ngZone = _ngZone;
+    this.noop = () => new NoopScrollStrategy();
+    this.close = (config) => new CloseScrollStrategy(this._scrollDispatcher, this._ngZone, this._viewportRuler, config);
+    this.block = () => new BlockScrollStrategy(this._viewportRuler, this._document);
+    this.reposition = (config) => new RepositionScrollStrategy(this._scrollDispatcher, this._viewportRuler, this._ngZone, config);
+    this._document = document;
+  }
+};
+_ScrollStrategyOptions.ɵfac = function ScrollStrategyOptions_Factory(t) {
+  return new (t || _ScrollStrategyOptions)(ɵɵinject(ScrollDispatcher), ɵɵinject(ViewportRuler), ɵɵinject(NgZone), ɵɵinject(DOCUMENT));
+};
+_ScrollStrategyOptions.ɵprov = ɵɵdefineInjectable({
+  token: _ScrollStrategyOptions,
+  factory: _ScrollStrategyOptions.ɵfac,
+  providedIn: "root"
+});
+var ScrollStrategyOptions = _ScrollStrategyOptions;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(ScrollStrategyOptions, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: ScrollDispatcher
+  }, {
+    type: ViewportRuler
+  }, {
+    type: NgZone
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }], null);
+})();
+var OverlayConfig = class {
+  constructor(config) {
+    this.scrollStrategy = new NoopScrollStrategy();
+    this.panelClass = "";
+    this.hasBackdrop = false;
+    this.backdropClass = "cdk-overlay-dark-backdrop";
+    this.disposeOnNavigation = false;
+    if (config) {
+      const configKeys = Object.keys(config);
+      for (const key of configKeys) {
+        if (config[key] !== void 0) {
+          this[key] = config[key];
+        }
+      }
+    }
+  }
+};
+var ConnectedOverlayPositionChange = class {
+  constructor(connectionPair, scrollableViewProperties) {
+    this.connectionPair = connectionPair;
+    this.scrollableViewProperties = scrollableViewProperties;
+  }
+};
+function validateVerticalPosition(property, value) {
+  if (value !== "top" && value !== "bottom" && value !== "center") {
+    throw Error(`ConnectedPosition: Invalid ${property} "${value}". Expected "top", "bottom" or "center".`);
+  }
+}
+function validateHorizontalPosition(property, value) {
+  if (value !== "start" && value !== "end" && value !== "center") {
+    throw Error(`ConnectedPosition: Invalid ${property} "${value}". Expected "start", "end" or "center".`);
+  }
+}
+var _BaseOverlayDispatcher = class _BaseOverlayDispatcher {
+  constructor(document) {
+    this._attachedOverlays = [];
+    this._document = document;
+  }
+  ngOnDestroy() {
+    this.detach();
+  }
+  /** Add a new overlay to the list of attached overlay refs. */
+  add(overlayRef) {
+    this.remove(overlayRef);
+    this._attachedOverlays.push(overlayRef);
+  }
+  /** Remove an overlay from the list of attached overlay refs. */
+  remove(overlayRef) {
+    const index = this._attachedOverlays.indexOf(overlayRef);
+    if (index > -1) {
+      this._attachedOverlays.splice(index, 1);
+    }
+    if (this._attachedOverlays.length === 0) {
+      this.detach();
+    }
+  }
+};
+_BaseOverlayDispatcher.ɵfac = function BaseOverlayDispatcher_Factory(t) {
+  return new (t || _BaseOverlayDispatcher)(ɵɵinject(DOCUMENT));
+};
+_BaseOverlayDispatcher.ɵprov = ɵɵdefineInjectable({
+  token: _BaseOverlayDispatcher,
+  factory: _BaseOverlayDispatcher.ɵfac,
+  providedIn: "root"
+});
+var BaseOverlayDispatcher = _BaseOverlayDispatcher;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(BaseOverlayDispatcher, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }], null);
+})();
+var _OverlayKeyboardDispatcher = class _OverlayKeyboardDispatcher extends BaseOverlayDispatcher {
+  constructor(document, _ngZone) {
+    super(document);
+    this._ngZone = _ngZone;
+    this._keydownListener = (event) => {
+      const overlays = this._attachedOverlays;
+      for (let i = overlays.length - 1; i > -1; i--) {
+        if (overlays[i]._keydownEvents.observers.length > 0) {
+          const keydownEvents = overlays[i]._keydownEvents;
+          if (this._ngZone) {
+            this._ngZone.run(() => keydownEvents.next(event));
+          } else {
+            keydownEvents.next(event);
+          }
+          break;
+        }
+      }
+    };
+  }
+  /** Add a new overlay to the list of attached overlay refs. */
+  add(overlayRef) {
+    super.add(overlayRef);
+    if (!this._isAttached) {
+      if (this._ngZone) {
+        this._ngZone.runOutsideAngular(() => this._document.body.addEventListener("keydown", this._keydownListener));
+      } else {
+        this._document.body.addEventListener("keydown", this._keydownListener);
+      }
+      this._isAttached = true;
+    }
+  }
+  /** Detaches the global keyboard event listener. */
+  detach() {
+    if (this._isAttached) {
+      this._document.body.removeEventListener("keydown", this._keydownListener);
+      this._isAttached = false;
+    }
+  }
+};
+_OverlayKeyboardDispatcher.ɵfac = function OverlayKeyboardDispatcher_Factory(t) {
+  return new (t || _OverlayKeyboardDispatcher)(ɵɵinject(DOCUMENT), ɵɵinject(NgZone, 8));
+};
+_OverlayKeyboardDispatcher.ɵprov = ɵɵdefineInjectable({
+  token: _OverlayKeyboardDispatcher,
+  factory: _OverlayKeyboardDispatcher.ɵfac,
+  providedIn: "root"
+});
+var OverlayKeyboardDispatcher = _OverlayKeyboardDispatcher;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(OverlayKeyboardDispatcher, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: NgZone,
+    decorators: [{
+      type: Optional
+    }]
+  }], null);
+})();
+var _OverlayOutsideClickDispatcher = class _OverlayOutsideClickDispatcher extends BaseOverlayDispatcher {
+  constructor(document, _platform, _ngZone) {
+    super(document);
+    this._platform = _platform;
+    this._ngZone = _ngZone;
+    this._cursorStyleIsSet = false;
+    this._pointerDownListener = (event) => {
+      this._pointerDownEventTarget = _getEventTarget(event);
+    };
+    this._clickListener = (event) => {
+      const target = _getEventTarget(event);
+      const origin = event.type === "click" && this._pointerDownEventTarget ? this._pointerDownEventTarget : target;
+      this._pointerDownEventTarget = null;
+      const overlays = this._attachedOverlays.slice();
+      for (let i = overlays.length - 1; i > -1; i--) {
+        const overlayRef = overlays[i];
+        if (overlayRef._outsidePointerEvents.observers.length < 1 || !overlayRef.hasAttached()) {
+          continue;
+        }
+        if (overlayRef.overlayElement.contains(target) || overlayRef.overlayElement.contains(origin)) {
+          break;
+        }
+        const outsidePointerEvents = overlayRef._outsidePointerEvents;
+        if (this._ngZone) {
+          this._ngZone.run(() => outsidePointerEvents.next(event));
+        } else {
+          outsidePointerEvents.next(event);
+        }
+      }
+    };
+  }
+  /** Add a new overlay to the list of attached overlay refs. */
+  add(overlayRef) {
+    super.add(overlayRef);
+    if (!this._isAttached) {
+      const body = this._document.body;
+      if (this._ngZone) {
+        this._ngZone.runOutsideAngular(() => this._addEventListeners(body));
+      } else {
+        this._addEventListeners(body);
+      }
+      if (this._platform.IOS && !this._cursorStyleIsSet) {
+        this._cursorOriginalValue = body.style.cursor;
+        body.style.cursor = "pointer";
+        this._cursorStyleIsSet = true;
+      }
+      this._isAttached = true;
+    }
+  }
+  /** Detaches the global keyboard event listener. */
+  detach() {
+    if (this._isAttached) {
+      const body = this._document.body;
+      body.removeEventListener("pointerdown", this._pointerDownListener, true);
+      body.removeEventListener("click", this._clickListener, true);
+      body.removeEventListener("auxclick", this._clickListener, true);
+      body.removeEventListener("contextmenu", this._clickListener, true);
+      if (this._platform.IOS && this._cursorStyleIsSet) {
+        body.style.cursor = this._cursorOriginalValue;
+        this._cursorStyleIsSet = false;
+      }
+      this._isAttached = false;
+    }
+  }
+  _addEventListeners(body) {
+    body.addEventListener("pointerdown", this._pointerDownListener, true);
+    body.addEventListener("click", this._clickListener, true);
+    body.addEventListener("auxclick", this._clickListener, true);
+    body.addEventListener("contextmenu", this._clickListener, true);
+  }
+};
+_OverlayOutsideClickDispatcher.ɵfac = function OverlayOutsideClickDispatcher_Factory(t) {
+  return new (t || _OverlayOutsideClickDispatcher)(ɵɵinject(DOCUMENT), ɵɵinject(Platform), ɵɵinject(NgZone, 8));
+};
+_OverlayOutsideClickDispatcher.ɵprov = ɵɵdefineInjectable({
+  token: _OverlayOutsideClickDispatcher,
+  factory: _OverlayOutsideClickDispatcher.ɵfac,
+  providedIn: "root"
+});
+var OverlayOutsideClickDispatcher = _OverlayOutsideClickDispatcher;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(OverlayOutsideClickDispatcher, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: Platform
+  }, {
+    type: NgZone,
+    decorators: [{
+      type: Optional
+    }]
+  }], null);
+})();
+var _OverlayContainer = class _OverlayContainer {
+  constructor(document, _platform) {
+    this._platform = _platform;
+    this._document = document;
+  }
+  ngOnDestroy() {
+    this._containerElement?.remove();
+  }
+  /**
+   * This method returns the overlay container element. It will lazily
+   * create the element the first time it is called to facilitate using
+   * the container in non-browser environments.
+   * @returns the container element
+   */
+  getContainerElement() {
+    if (!this._containerElement) {
+      this._createContainer();
+    }
+    return this._containerElement;
+  }
+  /**
+   * Create the overlay container element, which is simply a div
+   * with the 'cdk-overlay-container' class on the document body.
+   */
+  _createContainer() {
+    const containerClass = "cdk-overlay-container";
+    if (this._platform.isBrowser || _isTestEnvironment()) {
+      const oppositePlatformContainers = this._document.querySelectorAll(`.${containerClass}[platform="server"], .${containerClass}[platform="test"]`);
+      for (let i = 0; i < oppositePlatformContainers.length; i++) {
+        oppositePlatformContainers[i].remove();
+      }
+    }
+    const container = this._document.createElement("div");
+    container.classList.add(containerClass);
+    if (_isTestEnvironment()) {
+      container.setAttribute("platform", "test");
+    } else if (!this._platform.isBrowser) {
+      container.setAttribute("platform", "server");
+    }
+    this._document.body.appendChild(container);
+    this._containerElement = container;
+  }
+};
+_OverlayContainer.ɵfac = function OverlayContainer_Factory(t) {
+  return new (t || _OverlayContainer)(ɵɵinject(DOCUMENT), ɵɵinject(Platform));
+};
+_OverlayContainer.ɵprov = ɵɵdefineInjectable({
+  token: _OverlayContainer,
+  factory: _OverlayContainer.ɵfac,
+  providedIn: "root"
+});
+var OverlayContainer = _OverlayContainer;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(OverlayContainer, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: Platform
+  }], null);
+})();
+var OverlayRef = class {
+  constructor(_portalOutlet, _host, _pane, _config, _ngZone, _keyboardDispatcher, _document, _location, _outsideClickDispatcher, _animationsDisabled = false) {
+    this._portalOutlet = _portalOutlet;
+    this._host = _host;
+    this._pane = _pane;
+    this._config = _config;
+    this._ngZone = _ngZone;
+    this._keyboardDispatcher = _keyboardDispatcher;
+    this._document = _document;
+    this._location = _location;
+    this._outsideClickDispatcher = _outsideClickDispatcher;
+    this._animationsDisabled = _animationsDisabled;
+    this._backdropElement = null;
+    this._backdropClick = new Subject();
+    this._attachments = new Subject();
+    this._detachments = new Subject();
+    this._locationChanges = Subscription.EMPTY;
+    this._backdropClickHandler = (event) => this._backdropClick.next(event);
+    this._backdropTransitionendHandler = (event) => {
+      this._disposeBackdrop(event.target);
+    };
+    this._keydownEvents = new Subject();
+    this._outsidePointerEvents = new Subject();
+    if (_config.scrollStrategy) {
+      this._scrollStrategy = _config.scrollStrategy;
+      this._scrollStrategy.attach(this);
+    }
+    this._positionStrategy = _config.positionStrategy;
+  }
+  /** The overlay's HTML element */
+  get overlayElement() {
+    return this._pane;
+  }
+  /** The overlay's backdrop HTML element. */
+  get backdropElement() {
+    return this._backdropElement;
+  }
+  /**
+   * Wrapper around the panel element. Can be used for advanced
+   * positioning where a wrapper with specific styling is
+   * required around the overlay pane.
+   */
+  get hostElement() {
+    return this._host;
+  }
+  /**
+   * Attaches content, given via a Portal, to the overlay.
+   * If the overlay is configured to have a backdrop, it will be created.
+   *
+   * @param portal Portal instance to which to attach the overlay.
+   * @returns The portal attachment result.
+   */
+  attach(portal) {
+    if (!this._host.parentElement && this._previousHostParent) {
+      this._previousHostParent.appendChild(this._host);
+    }
+    const attachResult = this._portalOutlet.attach(portal);
+    if (this._positionStrategy) {
+      this._positionStrategy.attach(this);
+    }
+    this._updateStackingOrder();
+    this._updateElementSize();
+    this._updateElementDirection();
+    if (this._scrollStrategy) {
+      this._scrollStrategy.enable();
+    }
+    this._ngZone.onStable.pipe(take(1)).subscribe(() => {
+      if (this.hasAttached()) {
+        this.updatePosition();
+      }
+    });
+    this._togglePointerEvents(true);
+    if (this._config.hasBackdrop) {
+      this._attachBackdrop();
+    }
+    if (this._config.panelClass) {
+      this._toggleClasses(this._pane, this._config.panelClass, true);
+    }
+    this._attachments.next();
+    this._keyboardDispatcher.add(this);
+    if (this._config.disposeOnNavigation) {
+      this._locationChanges = this._location.subscribe(() => this.dispose());
+    }
+    this._outsideClickDispatcher.add(this);
+    if (typeof attachResult?.onDestroy === "function") {
+      attachResult.onDestroy(() => {
+        if (this.hasAttached()) {
+          this._ngZone.runOutsideAngular(() => Promise.resolve().then(() => this.detach()));
+        }
+      });
+    }
+    return attachResult;
+  }
+  /**
+   * Detaches an overlay from a portal.
+   * @returns The portal detachment result.
+   */
+  detach() {
+    if (!this.hasAttached()) {
+      return;
+    }
+    this.detachBackdrop();
+    this._togglePointerEvents(false);
+    if (this._positionStrategy && this._positionStrategy.detach) {
+      this._positionStrategy.detach();
+    }
+    if (this._scrollStrategy) {
+      this._scrollStrategy.disable();
+    }
+    const detachmentResult = this._portalOutlet.detach();
+    this._detachments.next();
+    this._keyboardDispatcher.remove(this);
+    this._detachContentWhenStable();
+    this._locationChanges.unsubscribe();
+    this._outsideClickDispatcher.remove(this);
+    return detachmentResult;
+  }
+  /** Cleans up the overlay from the DOM. */
+  dispose() {
+    const isAttached = this.hasAttached();
+    if (this._positionStrategy) {
+      this._positionStrategy.dispose();
+    }
+    this._disposeScrollStrategy();
+    this._disposeBackdrop(this._backdropElement);
+    this._locationChanges.unsubscribe();
+    this._keyboardDispatcher.remove(this);
+    this._portalOutlet.dispose();
+    this._attachments.complete();
+    this._backdropClick.complete();
+    this._keydownEvents.complete();
+    this._outsidePointerEvents.complete();
+    this._outsideClickDispatcher.remove(this);
+    this._host?.remove();
+    this._previousHostParent = this._pane = this._host = null;
+    if (isAttached) {
+      this._detachments.next();
+    }
+    this._detachments.complete();
+  }
+  /** Whether the overlay has attached content. */
+  hasAttached() {
+    return this._portalOutlet.hasAttached();
+  }
+  /** Gets an observable that emits when the backdrop has been clicked. */
+  backdropClick() {
+    return this._backdropClick;
+  }
+  /** Gets an observable that emits when the overlay has been attached. */
+  attachments() {
+    return this._attachments;
+  }
+  /** Gets an observable that emits when the overlay has been detached. */
+  detachments() {
+    return this._detachments;
+  }
+  /** Gets an observable of keydown events targeted to this overlay. */
+  keydownEvents() {
+    return this._keydownEvents;
+  }
+  /** Gets an observable of pointer events targeted outside this overlay. */
+  outsidePointerEvents() {
+    return this._outsidePointerEvents;
+  }
+  /** Gets the current overlay configuration, which is immutable. */
+  getConfig() {
+    return this._config;
+  }
+  /** Updates the position of the overlay based on the position strategy. */
+  updatePosition() {
+    if (this._positionStrategy) {
+      this._positionStrategy.apply();
+    }
+  }
+  /** Switches to a new position strategy and updates the overlay position. */
+  updatePositionStrategy(strategy) {
+    if (strategy === this._positionStrategy) {
+      return;
+    }
+    if (this._positionStrategy) {
+      this._positionStrategy.dispose();
+    }
+    this._positionStrategy = strategy;
+    if (this.hasAttached()) {
+      strategy.attach(this);
+      this.updatePosition();
+    }
+  }
+  /** Update the size properties of the overlay. */
+  updateSize(sizeConfig) {
+    this._config = __spreadValues(__spreadValues({}, this._config), sizeConfig);
+    this._updateElementSize();
+  }
+  /** Sets the LTR/RTL direction for the overlay. */
+  setDirection(dir) {
+    this._config = __spreadProps(__spreadValues({}, this._config), {
+      direction: dir
+    });
+    this._updateElementDirection();
+  }
+  /** Add a CSS class or an array of classes to the overlay pane. */
+  addPanelClass(classes) {
+    if (this._pane) {
+      this._toggleClasses(this._pane, classes, true);
+    }
+  }
+  /** Remove a CSS class or an array of classes from the overlay pane. */
+  removePanelClass(classes) {
+    if (this._pane) {
+      this._toggleClasses(this._pane, classes, false);
+    }
+  }
+  /**
+   * Returns the layout direction of the overlay panel.
+   */
+  getDirection() {
+    const direction = this._config.direction;
+    if (!direction) {
+      return "ltr";
+    }
+    return typeof direction === "string" ? direction : direction.value;
+  }
+  /** Switches to a new scroll strategy. */
+  updateScrollStrategy(strategy) {
+    if (strategy === this._scrollStrategy) {
+      return;
+    }
+    this._disposeScrollStrategy();
+    this._scrollStrategy = strategy;
+    if (this.hasAttached()) {
+      strategy.attach(this);
+      strategy.enable();
+    }
+  }
+  /** Updates the text direction of the overlay panel. */
+  _updateElementDirection() {
+    this._host.setAttribute("dir", this.getDirection());
+  }
+  /** Updates the size of the overlay element based on the overlay config. */
+  _updateElementSize() {
+    if (!this._pane) {
+      return;
+    }
+    const style = this._pane.style;
+    style.width = coerceCssPixelValue(this._config.width);
+    style.height = coerceCssPixelValue(this._config.height);
+    style.minWidth = coerceCssPixelValue(this._config.minWidth);
+    style.minHeight = coerceCssPixelValue(this._config.minHeight);
+    style.maxWidth = coerceCssPixelValue(this._config.maxWidth);
+    style.maxHeight = coerceCssPixelValue(this._config.maxHeight);
+  }
+  /** Toggles the pointer events for the overlay pane element. */
+  _togglePointerEvents(enablePointer) {
+    this._pane.style.pointerEvents = enablePointer ? "" : "none";
+  }
+  /** Attaches a backdrop for this overlay. */
+  _attachBackdrop() {
+    const showingClass = "cdk-overlay-backdrop-showing";
+    this._backdropElement = this._document.createElement("div");
+    this._backdropElement.classList.add("cdk-overlay-backdrop");
+    if (this._animationsDisabled) {
+      this._backdropElement.classList.add("cdk-overlay-backdrop-noop-animation");
+    }
+    if (this._config.backdropClass) {
+      this._toggleClasses(this._backdropElement, this._config.backdropClass, true);
+    }
+    this._host.parentElement.insertBefore(this._backdropElement, this._host);
+    this._backdropElement.addEventListener("click", this._backdropClickHandler);
+    if (!this._animationsDisabled && typeof requestAnimationFrame !== "undefined") {
+      this._ngZone.runOutsideAngular(() => {
+        requestAnimationFrame(() => {
+          if (this._backdropElement) {
+            this._backdropElement.classList.add(showingClass);
+          }
+        });
+      });
+    } else {
+      this._backdropElement.classList.add(showingClass);
+    }
+  }
+  /**
+   * Updates the stacking order of the element, moving it to the top if necessary.
+   * This is required in cases where one overlay was detached, while another one,
+   * that should be behind it, was destroyed. The next time both of them are opened,
+   * the stacking will be wrong, because the detached element's pane will still be
+   * in its original DOM position.
+   */
+  _updateStackingOrder() {
+    if (this._host.nextSibling) {
+      this._host.parentNode.appendChild(this._host);
+    }
+  }
+  /** Detaches the backdrop (if any) associated with the overlay. */
+  detachBackdrop() {
+    const backdropToDetach = this._backdropElement;
+    if (!backdropToDetach) {
+      return;
+    }
+    if (this._animationsDisabled) {
+      this._disposeBackdrop(backdropToDetach);
+      return;
+    }
+    backdropToDetach.classList.remove("cdk-overlay-backdrop-showing");
+    this._ngZone.runOutsideAngular(() => {
+      backdropToDetach.addEventListener("transitionend", this._backdropTransitionendHandler);
+    });
+    backdropToDetach.style.pointerEvents = "none";
+    this._backdropTimeout = this._ngZone.runOutsideAngular(() => setTimeout(() => {
+      this._disposeBackdrop(backdropToDetach);
+    }, 500));
+  }
+  /** Toggles a single CSS class or an array of classes on an element. */
+  _toggleClasses(element, cssClasses, isAdd) {
+    const classes = coerceArray(cssClasses || []).filter((c) => !!c);
+    if (classes.length) {
+      isAdd ? element.classList.add(...classes) : element.classList.remove(...classes);
+    }
+  }
+  /** Detaches the overlay content next time the zone stabilizes. */
+  _detachContentWhenStable() {
+    this._ngZone.runOutsideAngular(() => {
+      const subscription = this._ngZone.onStable.pipe(takeUntil(merge(this._attachments, this._detachments))).subscribe(() => {
+        if (!this._pane || !this._host || this._pane.children.length === 0) {
+          if (this._pane && this._config.panelClass) {
+            this._toggleClasses(this._pane, this._config.panelClass, false);
+          }
+          if (this._host && this._host.parentElement) {
+            this._previousHostParent = this._host.parentElement;
+            this._host.remove();
+          }
+          subscription.unsubscribe();
+        }
+      });
+    });
+  }
+  /** Disposes of a scroll strategy. */
+  _disposeScrollStrategy() {
+    const scrollStrategy = this._scrollStrategy;
+    if (scrollStrategy) {
+      scrollStrategy.disable();
+      if (scrollStrategy.detach) {
+        scrollStrategy.detach();
+      }
+    }
+  }
+  /** Removes a backdrop element from the DOM. */
+  _disposeBackdrop(backdrop) {
+    if (backdrop) {
+      backdrop.removeEventListener("click", this._backdropClickHandler);
+      backdrop.removeEventListener("transitionend", this._backdropTransitionendHandler);
+      backdrop.remove();
+      if (this._backdropElement === backdrop) {
+        this._backdropElement = null;
+      }
+    }
+    if (this._backdropTimeout) {
+      clearTimeout(this._backdropTimeout);
+      this._backdropTimeout = void 0;
+    }
+  }
+};
+var boundingBoxClass = "cdk-overlay-connected-position-bounding-box";
+var cssUnitPattern = /([A-Za-z%]+)$/;
+var FlexibleConnectedPositionStrategy = class {
+  /** Ordered list of preferred positions, from most to least desirable. */
+  get positions() {
+    return this._preferredPositions;
+  }
+  constructor(connectedTo, _viewportRuler, _document, _platform, _overlayContainer) {
+    this._viewportRuler = _viewportRuler;
+    this._document = _document;
+    this._platform = _platform;
+    this._overlayContainer = _overlayContainer;
+    this._lastBoundingBoxSize = {
+      width: 0,
+      height: 0
+    };
+    this._isPushed = false;
+    this._canPush = true;
+    this._growAfterOpen = false;
+    this._hasFlexibleDimensions = true;
+    this._positionLocked = false;
+    this._viewportMargin = 0;
+    this._scrollables = [];
+    this._preferredPositions = [];
+    this._positionChanges = new Subject();
+    this._resizeSubscription = Subscription.EMPTY;
+    this._offsetX = 0;
+    this._offsetY = 0;
+    this._appliedPanelClasses = [];
+    this.positionChanges = this._positionChanges;
+    this.setOrigin(connectedTo);
+  }
+  /** Attaches this position strategy to an overlay. */
+  attach(overlayRef) {
+    if (this._overlayRef && overlayRef !== this._overlayRef && (typeof ngDevMode === "undefined" || ngDevMode)) {
+      throw Error("This position strategy is already attached to an overlay");
+    }
+    this._validatePositions();
+    overlayRef.hostElement.classList.add(boundingBoxClass);
+    this._overlayRef = overlayRef;
+    this._boundingBox = overlayRef.hostElement;
+    this._pane = overlayRef.overlayElement;
+    this._isDisposed = false;
+    this._isInitialRender = true;
+    this._lastPosition = null;
+    this._resizeSubscription.unsubscribe();
+    this._resizeSubscription = this._viewportRuler.change().subscribe(() => {
+      this._isInitialRender = true;
+      this.apply();
+    });
+  }
+  /**
+   * Updates the position of the overlay element, using whichever preferred position relative
+   * to the origin best fits on-screen.
+   *
+   * The selection of a position goes as follows:
+   *  - If any positions fit completely within the viewport as-is,
+   *      choose the first position that does so.
+   *  - If flexible dimensions are enabled and at least one satisfies the given minimum width/height,
+   *      choose the position with the greatest available size modified by the positions' weight.
+   *  - If pushing is enabled, take the position that went off-screen the least and push it
+   *      on-screen.
+   *  - If none of the previous criteria were met, use the position that goes off-screen the least.
+   * @docs-private
+   */
+  apply() {
+    if (this._isDisposed || !this._platform.isBrowser) {
+      return;
+    }
+    if (!this._isInitialRender && this._positionLocked && this._lastPosition) {
+      this.reapplyLastPosition();
+      return;
+    }
+    this._clearPanelClasses();
+    this._resetOverlayElementStyles();
+    this._resetBoundingBoxStyles();
+    this._viewportRect = this._getNarrowedViewportRect();
+    this._originRect = this._getOriginRect();
+    this._overlayRect = this._pane.getBoundingClientRect();
+    this._containerRect = this._overlayContainer.getContainerElement().getBoundingClientRect();
+    const originRect = this._originRect;
+    const overlayRect = this._overlayRect;
+    const viewportRect = this._viewportRect;
+    const containerRect = this._containerRect;
+    const flexibleFits = [];
+    let fallback;
+    for (let pos of this._preferredPositions) {
+      let originPoint = this._getOriginPoint(originRect, containerRect, pos);
+      let overlayPoint = this._getOverlayPoint(originPoint, overlayRect, pos);
+      let overlayFit = this._getOverlayFit(overlayPoint, overlayRect, viewportRect, pos);
+      if (overlayFit.isCompletelyWithinViewport) {
+        this._isPushed = false;
+        this._applyPosition(pos, originPoint);
+        return;
+      }
+      if (this._canFitWithFlexibleDimensions(overlayFit, overlayPoint, viewportRect)) {
+        flexibleFits.push({
+          position: pos,
+          origin: originPoint,
+          overlayRect,
+          boundingBoxRect: this._calculateBoundingBoxRect(originPoint, pos)
+        });
+        continue;
+      }
+      if (!fallback || fallback.overlayFit.visibleArea < overlayFit.visibleArea) {
+        fallback = {
+          overlayFit,
+          overlayPoint,
+          originPoint,
+          position: pos,
+          overlayRect
+        };
+      }
+    }
+    if (flexibleFits.length) {
+      let bestFit = null;
+      let bestScore = -1;
+      for (const fit of flexibleFits) {
+        const score = fit.boundingBoxRect.width * fit.boundingBoxRect.height * (fit.position.weight || 1);
+        if (score > bestScore) {
+          bestScore = score;
+          bestFit = fit;
+        }
+      }
+      this._isPushed = false;
+      this._applyPosition(bestFit.position, bestFit.origin);
+      return;
+    }
+    if (this._canPush) {
+      this._isPushed = true;
+      this._applyPosition(fallback.position, fallback.originPoint);
+      return;
+    }
+    this._applyPosition(fallback.position, fallback.originPoint);
+  }
+  detach() {
+    this._clearPanelClasses();
+    this._lastPosition = null;
+    this._previousPushAmount = null;
+    this._resizeSubscription.unsubscribe();
+  }
+  /** Cleanup after the element gets destroyed. */
+  dispose() {
+    if (this._isDisposed) {
+      return;
+    }
+    if (this._boundingBox) {
+      extendStyles(this._boundingBox.style, {
+        top: "",
+        left: "",
+        right: "",
+        bottom: "",
+        height: "",
+        width: "",
+        alignItems: "",
+        justifyContent: ""
+      });
+    }
+    if (this._pane) {
+      this._resetOverlayElementStyles();
+    }
+    if (this._overlayRef) {
+      this._overlayRef.hostElement.classList.remove(boundingBoxClass);
+    }
+    this.detach();
+    this._positionChanges.complete();
+    this._overlayRef = this._boundingBox = null;
+    this._isDisposed = true;
+  }
+  /**
+   * This re-aligns the overlay element with the trigger in its last calculated position,
+   * even if a position higher in the "preferred positions" list would now fit. This
+   * allows one to re-align the panel without changing the orientation of the panel.
+   */
+  reapplyLastPosition() {
+    if (this._isDisposed || !this._platform.isBrowser) {
+      return;
+    }
+    const lastPosition = this._lastPosition;
+    if (lastPosition) {
+      this._originRect = this._getOriginRect();
+      this._overlayRect = this._pane.getBoundingClientRect();
+      this._viewportRect = this._getNarrowedViewportRect();
+      this._containerRect = this._overlayContainer.getContainerElement().getBoundingClientRect();
+      const originPoint = this._getOriginPoint(this._originRect, this._containerRect, lastPosition);
+      this._applyPosition(lastPosition, originPoint);
+    } else {
+      this.apply();
+    }
+  }
+  /**
+   * Sets the list of Scrollable containers that host the origin element so that
+   * on reposition we can evaluate if it or the overlay has been clipped or outside view. Every
+   * Scrollable must be an ancestor element of the strategy's origin element.
+   */
+  withScrollableContainers(scrollables) {
+    this._scrollables = scrollables;
+    return this;
+  }
+  /**
+   * Adds new preferred positions.
+   * @param positions List of positions options for this overlay.
+   */
+  withPositions(positions) {
+    this._preferredPositions = positions;
+    if (positions.indexOf(this._lastPosition) === -1) {
+      this._lastPosition = null;
+    }
+    this._validatePositions();
+    return this;
+  }
+  /**
+   * Sets a minimum distance the overlay may be positioned to the edge of the viewport.
+   * @param margin Required margin between the overlay and the viewport edge in pixels.
+   */
+  withViewportMargin(margin) {
+    this._viewportMargin = margin;
+    return this;
+  }
+  /** Sets whether the overlay's width and height can be constrained to fit within the viewport. */
+  withFlexibleDimensions(flexibleDimensions = true) {
+    this._hasFlexibleDimensions = flexibleDimensions;
+    return this;
+  }
+  /** Sets whether the overlay can grow after the initial open via flexible width/height. */
+  withGrowAfterOpen(growAfterOpen = true) {
+    this._growAfterOpen = growAfterOpen;
+    return this;
+  }
+  /** Sets whether the overlay can be pushed on-screen if none of the provided positions fit. */
+  withPush(canPush = true) {
+    this._canPush = canPush;
+    return this;
+  }
+  /**
+   * Sets whether the overlay's position should be locked in after it is positioned
+   * initially. When an overlay is locked in, it won't attempt to reposition itself
+   * when the position is re-applied (e.g. when the user scrolls away).
+   * @param isLocked Whether the overlay should locked in.
+   */
+  withLockedPosition(isLocked = true) {
+    this._positionLocked = isLocked;
+    return this;
+  }
+  /**
+   * Sets the origin, relative to which to position the overlay.
+   * Using an element origin is useful for building components that need to be positioned
+   * relatively to a trigger (e.g. dropdown menus or tooltips), whereas using a point can be
+   * used for cases like contextual menus which open relative to the user's pointer.
+   * @param origin Reference to the new origin.
+   */
+  setOrigin(origin) {
+    this._origin = origin;
+    return this;
+  }
+  /**
+   * Sets the default offset for the overlay's connection point on the x-axis.
+   * @param offset New offset in the X axis.
+   */
+  withDefaultOffsetX(offset) {
+    this._offsetX = offset;
+    return this;
+  }
+  /**
+   * Sets the default offset for the overlay's connection point on the y-axis.
+   * @param offset New offset in the Y axis.
+   */
+  withDefaultOffsetY(offset) {
+    this._offsetY = offset;
+    return this;
+  }
+  /**
+   * Configures that the position strategy should set a `transform-origin` on some elements
+   * inside the overlay, depending on the current position that is being applied. This is
+   * useful for the cases where the origin of an animation can change depending on the
+   * alignment of the overlay.
+   * @param selector CSS selector that will be used to find the target
+   *    elements onto which to set the transform origin.
+   */
+  withTransformOriginOn(selector) {
+    this._transformOriginSelector = selector;
+    return this;
+  }
+  /**
+   * Gets the (x, y) coordinate of a connection point on the origin based on a relative position.
+   */
+  _getOriginPoint(originRect, containerRect, pos) {
+    let x;
+    if (pos.originX == "center") {
+      x = originRect.left + originRect.width / 2;
+    } else {
+      const startX = this._isRtl() ? originRect.right : originRect.left;
+      const endX = this._isRtl() ? originRect.left : originRect.right;
+      x = pos.originX == "start" ? startX : endX;
+    }
+    if (containerRect.left < 0) {
+      x -= containerRect.left;
+    }
+    let y;
+    if (pos.originY == "center") {
+      y = originRect.top + originRect.height / 2;
+    } else {
+      y = pos.originY == "top" ? originRect.top : originRect.bottom;
+    }
+    if (containerRect.top < 0) {
+      y -= containerRect.top;
+    }
+    return {
+      x,
+      y
+    };
+  }
+  /**
+   * Gets the (x, y) coordinate of the top-left corner of the overlay given a given position and
+   * origin point to which the overlay should be connected.
+   */
+  _getOverlayPoint(originPoint, overlayRect, pos) {
+    let overlayStartX;
+    if (pos.overlayX == "center") {
+      overlayStartX = -overlayRect.width / 2;
+    } else if (pos.overlayX === "start") {
+      overlayStartX = this._isRtl() ? -overlayRect.width : 0;
+    } else {
+      overlayStartX = this._isRtl() ? 0 : -overlayRect.width;
+    }
+    let overlayStartY;
+    if (pos.overlayY == "center") {
+      overlayStartY = -overlayRect.height / 2;
+    } else {
+      overlayStartY = pos.overlayY == "top" ? 0 : -overlayRect.height;
+    }
+    return {
+      x: originPoint.x + overlayStartX,
+      y: originPoint.y + overlayStartY
+    };
+  }
+  /** Gets how well an overlay at the given point will fit within the viewport. */
+  _getOverlayFit(point, rawOverlayRect, viewport, position) {
+    const overlay = getRoundedBoundingClientRect(rawOverlayRect);
+    let {
+      x,
+      y
+    } = point;
+    let offsetX = this._getOffset(position, "x");
+    let offsetY = this._getOffset(position, "y");
+    if (offsetX) {
+      x += offsetX;
+    }
+    if (offsetY) {
+      y += offsetY;
+    }
+    let leftOverflow = 0 - x;
+    let rightOverflow = x + overlay.width - viewport.width;
+    let topOverflow = 0 - y;
+    let bottomOverflow = y + overlay.height - viewport.height;
+    let visibleWidth = this._subtractOverflows(overlay.width, leftOverflow, rightOverflow);
+    let visibleHeight = this._subtractOverflows(overlay.height, topOverflow, bottomOverflow);
+    let visibleArea = visibleWidth * visibleHeight;
+    return {
+      visibleArea,
+      isCompletelyWithinViewport: overlay.width * overlay.height === visibleArea,
+      fitsInViewportVertically: visibleHeight === overlay.height,
+      fitsInViewportHorizontally: visibleWidth == overlay.width
+    };
+  }
+  /**
+   * Whether the overlay can fit within the viewport when it may resize either its width or height.
+   * @param fit How well the overlay fits in the viewport at some position.
+   * @param point The (x, y) coordinates of the overlay at some position.
+   * @param viewport The geometry of the viewport.
+   */
+  _canFitWithFlexibleDimensions(fit, point, viewport) {
+    if (this._hasFlexibleDimensions) {
+      const availableHeight = viewport.bottom - point.y;
+      const availableWidth = viewport.right - point.x;
+      const minHeight = getPixelValue(this._overlayRef.getConfig().minHeight);
+      const minWidth = getPixelValue(this._overlayRef.getConfig().minWidth);
+      const verticalFit = fit.fitsInViewportVertically || minHeight != null && minHeight <= availableHeight;
+      const horizontalFit = fit.fitsInViewportHorizontally || minWidth != null && minWidth <= availableWidth;
+      return verticalFit && horizontalFit;
+    }
+    return false;
+  }
+  /**
+   * Gets the point at which the overlay can be "pushed" on-screen. If the overlay is larger than
+   * the viewport, the top-left corner will be pushed on-screen (with overflow occurring on the
+   * right and bottom).
+   *
+   * @param start Starting point from which the overlay is pushed.
+   * @param rawOverlayRect Dimensions of the overlay.
+   * @param scrollPosition Current viewport scroll position.
+   * @returns The point at which to position the overlay after pushing. This is effectively a new
+   *     originPoint.
+   */
+  _pushOverlayOnScreen(start, rawOverlayRect, scrollPosition) {
+    if (this._previousPushAmount && this._positionLocked) {
+      return {
+        x: start.x + this._previousPushAmount.x,
+        y: start.y + this._previousPushAmount.y
+      };
+    }
+    const overlay = getRoundedBoundingClientRect(rawOverlayRect);
+    const viewport = this._viewportRect;
+    const overflowRight = Math.max(start.x + overlay.width - viewport.width, 0);
+    const overflowBottom = Math.max(start.y + overlay.height - viewport.height, 0);
+    const overflowTop = Math.max(viewport.top - scrollPosition.top - start.y, 0);
+    const overflowLeft = Math.max(viewport.left - scrollPosition.left - start.x, 0);
+    let pushX = 0;
+    let pushY = 0;
+    if (overlay.width <= viewport.width) {
+      pushX = overflowLeft || -overflowRight;
+    } else {
+      pushX = start.x < this._viewportMargin ? viewport.left - scrollPosition.left - start.x : 0;
+    }
+    if (overlay.height <= viewport.height) {
+      pushY = overflowTop || -overflowBottom;
+    } else {
+      pushY = start.y < this._viewportMargin ? viewport.top - scrollPosition.top - start.y : 0;
+    }
+    this._previousPushAmount = {
+      x: pushX,
+      y: pushY
+    };
+    return {
+      x: start.x + pushX,
+      y: start.y + pushY
+    };
+  }
+  /**
+   * Applies a computed position to the overlay and emits a position change.
+   * @param position The position preference
+   * @param originPoint The point on the origin element where the overlay is connected.
+   */
+  _applyPosition(position, originPoint) {
+    this._setTransformOrigin(position);
+    this._setOverlayElementStyles(originPoint, position);
+    this._setBoundingBoxStyles(originPoint, position);
+    if (position.panelClass) {
+      this._addPanelClasses(position.panelClass);
+    }
+    if (this._positionChanges.observers.length) {
+      const scrollVisibility = this._getScrollVisibility();
+      if (position !== this._lastPosition || !this._lastScrollVisibility || !compareScrollVisibility(this._lastScrollVisibility, scrollVisibility)) {
+        const changeEvent = new ConnectedOverlayPositionChange(position, scrollVisibility);
+        this._positionChanges.next(changeEvent);
+      }
+      this._lastScrollVisibility = scrollVisibility;
+    }
+    this._lastPosition = position;
+    this._isInitialRender = false;
+  }
+  /** Sets the transform origin based on the configured selector and the passed-in position.  */
+  _setTransformOrigin(position) {
+    if (!this._transformOriginSelector) {
+      return;
+    }
+    const elements = this._boundingBox.querySelectorAll(this._transformOriginSelector);
+    let xOrigin;
+    let yOrigin = position.overlayY;
+    if (position.overlayX === "center") {
+      xOrigin = "center";
+    } else if (this._isRtl()) {
+      xOrigin = position.overlayX === "start" ? "right" : "left";
+    } else {
+      xOrigin = position.overlayX === "start" ? "left" : "right";
+    }
+    for (let i = 0; i < elements.length; i++) {
+      elements[i].style.transformOrigin = `${xOrigin} ${yOrigin}`;
+    }
+  }
+  /**
+   * Gets the position and size of the overlay's sizing container.
+   *
+   * This method does no measuring and applies no styles so that we can cheaply compute the
+   * bounds for all positions and choose the best fit based on these results.
+   */
+  _calculateBoundingBoxRect(origin, position) {
+    const viewport = this._viewportRect;
+    const isRtl = this._isRtl();
+    let height, top, bottom;
+    if (position.overlayY === "top") {
+      top = origin.y;
+      height = viewport.height - top + this._viewportMargin;
+    } else if (position.overlayY === "bottom") {
+      bottom = viewport.height - origin.y + this._viewportMargin * 2;
+      height = viewport.height - bottom + this._viewportMargin;
+    } else {
+      const smallestDistanceToViewportEdge = Math.min(viewport.bottom - origin.y + viewport.top, origin.y);
+      const previousHeight = this._lastBoundingBoxSize.height;
+      height = smallestDistanceToViewportEdge * 2;
+      top = origin.y - smallestDistanceToViewportEdge;
+      if (height > previousHeight && !this._isInitialRender && !this._growAfterOpen) {
+        top = origin.y - previousHeight / 2;
+      }
+    }
+    const isBoundedByRightViewportEdge = position.overlayX === "start" && !isRtl || position.overlayX === "end" && isRtl;
+    const isBoundedByLeftViewportEdge = position.overlayX === "end" && !isRtl || position.overlayX === "start" && isRtl;
+    let width, left, right;
+    if (isBoundedByLeftViewportEdge) {
+      right = viewport.width - origin.x + this._viewportMargin * 2;
+      width = origin.x - this._viewportMargin;
+    } else if (isBoundedByRightViewportEdge) {
+      left = origin.x;
+      width = viewport.right - origin.x;
+    } else {
+      const smallestDistanceToViewportEdge = Math.min(viewport.right - origin.x + viewport.left, origin.x);
+      const previousWidth = this._lastBoundingBoxSize.width;
+      width = smallestDistanceToViewportEdge * 2;
+      left = origin.x - smallestDistanceToViewportEdge;
+      if (width > previousWidth && !this._isInitialRender && !this._growAfterOpen) {
+        left = origin.x - previousWidth / 2;
+      }
+    }
+    return {
+      top,
+      left,
+      bottom,
+      right,
+      width,
+      height
+    };
+  }
+  /**
+   * Sets the position and size of the overlay's sizing wrapper. The wrapper is positioned on the
+   * origin's connection point and stretches to the bounds of the viewport.
+   *
+   * @param origin The point on the origin element where the overlay is connected.
+   * @param position The position preference
+   */
+  _setBoundingBoxStyles(origin, position) {
+    const boundingBoxRect = this._calculateBoundingBoxRect(origin, position);
+    if (!this._isInitialRender && !this._growAfterOpen) {
+      boundingBoxRect.height = Math.min(boundingBoxRect.height, this._lastBoundingBoxSize.height);
+      boundingBoxRect.width = Math.min(boundingBoxRect.width, this._lastBoundingBoxSize.width);
+    }
+    const styles = {};
+    if (this._hasExactPosition()) {
+      styles.top = styles.left = "0";
+      styles.bottom = styles.right = styles.maxHeight = styles.maxWidth = "";
+      styles.width = styles.height = "100%";
+    } else {
+      const maxHeight = this._overlayRef.getConfig().maxHeight;
+      const maxWidth = this._overlayRef.getConfig().maxWidth;
+      styles.height = coerceCssPixelValue(boundingBoxRect.height);
+      styles.top = coerceCssPixelValue(boundingBoxRect.top);
+      styles.bottom = coerceCssPixelValue(boundingBoxRect.bottom);
+      styles.width = coerceCssPixelValue(boundingBoxRect.width);
+      styles.left = coerceCssPixelValue(boundingBoxRect.left);
+      styles.right = coerceCssPixelValue(boundingBoxRect.right);
+      if (position.overlayX === "center") {
+        styles.alignItems = "center";
+      } else {
+        styles.alignItems = position.overlayX === "end" ? "flex-end" : "flex-start";
+      }
+      if (position.overlayY === "center") {
+        styles.justifyContent = "center";
+      } else {
+        styles.justifyContent = position.overlayY === "bottom" ? "flex-end" : "flex-start";
+      }
+      if (maxHeight) {
+        styles.maxHeight = coerceCssPixelValue(maxHeight);
+      }
+      if (maxWidth) {
+        styles.maxWidth = coerceCssPixelValue(maxWidth);
+      }
+    }
+    this._lastBoundingBoxSize = boundingBoxRect;
+    extendStyles(this._boundingBox.style, styles);
+  }
+  /** Resets the styles for the bounding box so that a new positioning can be computed. */
+  _resetBoundingBoxStyles() {
+    extendStyles(this._boundingBox.style, {
+      top: "0",
+      left: "0",
+      right: "0",
+      bottom: "0",
+      height: "",
+      width: "",
+      alignItems: "",
+      justifyContent: ""
+    });
+  }
+  /** Resets the styles for the overlay pane so that a new positioning can be computed. */
+  _resetOverlayElementStyles() {
+    extendStyles(this._pane.style, {
+      top: "",
+      left: "",
+      bottom: "",
+      right: "",
+      position: "",
+      transform: ""
+    });
+  }
+  /** Sets positioning styles to the overlay element. */
+  _setOverlayElementStyles(originPoint, position) {
+    const styles = {};
+    const hasExactPosition = this._hasExactPosition();
+    const hasFlexibleDimensions = this._hasFlexibleDimensions;
+    const config = this._overlayRef.getConfig();
+    if (hasExactPosition) {
+      const scrollPosition = this._viewportRuler.getViewportScrollPosition();
+      extendStyles(styles, this._getExactOverlayY(position, originPoint, scrollPosition));
+      extendStyles(styles, this._getExactOverlayX(position, originPoint, scrollPosition));
+    } else {
+      styles.position = "static";
+    }
+    let transformString = "";
+    let offsetX = this._getOffset(position, "x");
+    let offsetY = this._getOffset(position, "y");
+    if (offsetX) {
+      transformString += `translateX(${offsetX}px) `;
+    }
+    if (offsetY) {
+      transformString += `translateY(${offsetY}px)`;
+    }
+    styles.transform = transformString.trim();
+    if (config.maxHeight) {
+      if (hasExactPosition) {
+        styles.maxHeight = coerceCssPixelValue(config.maxHeight);
+      } else if (hasFlexibleDimensions) {
+        styles.maxHeight = "";
+      }
+    }
+    if (config.maxWidth) {
+      if (hasExactPosition) {
+        styles.maxWidth = coerceCssPixelValue(config.maxWidth);
+      } else if (hasFlexibleDimensions) {
+        styles.maxWidth = "";
+      }
+    }
+    extendStyles(this._pane.style, styles);
+  }
+  /** Gets the exact top/bottom for the overlay when not using flexible sizing or when pushing. */
+  _getExactOverlayY(position, originPoint, scrollPosition) {
+    let styles = {
+      top: "",
+      bottom: ""
+    };
+    let overlayPoint = this._getOverlayPoint(originPoint, this._overlayRect, position);
+    if (this._isPushed) {
+      overlayPoint = this._pushOverlayOnScreen(overlayPoint, this._overlayRect, scrollPosition);
+    }
+    if (position.overlayY === "bottom") {
+      const documentHeight = this._document.documentElement.clientHeight;
+      styles.bottom = `${documentHeight - (overlayPoint.y + this._overlayRect.height)}px`;
+    } else {
+      styles.top = coerceCssPixelValue(overlayPoint.y);
+    }
+    return styles;
+  }
+  /** Gets the exact left/right for the overlay when not using flexible sizing or when pushing. */
+  _getExactOverlayX(position, originPoint, scrollPosition) {
+    let styles = {
+      left: "",
+      right: ""
+    };
+    let overlayPoint = this._getOverlayPoint(originPoint, this._overlayRect, position);
+    if (this._isPushed) {
+      overlayPoint = this._pushOverlayOnScreen(overlayPoint, this._overlayRect, scrollPosition);
+    }
+    let horizontalStyleProperty;
+    if (this._isRtl()) {
+      horizontalStyleProperty = position.overlayX === "end" ? "left" : "right";
+    } else {
+      horizontalStyleProperty = position.overlayX === "end" ? "right" : "left";
+    }
+    if (horizontalStyleProperty === "right") {
+      const documentWidth = this._document.documentElement.clientWidth;
+      styles.right = `${documentWidth - (overlayPoint.x + this._overlayRect.width)}px`;
+    } else {
+      styles.left = coerceCssPixelValue(overlayPoint.x);
+    }
+    return styles;
+  }
+  /**
+   * Gets the view properties of the trigger and overlay, including whether they are clipped
+   * or completely outside the view of any of the strategy's scrollables.
+   */
+  _getScrollVisibility() {
+    const originBounds = this._getOriginRect();
+    const overlayBounds = this._pane.getBoundingClientRect();
+    const scrollContainerBounds = this._scrollables.map((scrollable) => {
+      return scrollable.getElementRef().nativeElement.getBoundingClientRect();
+    });
+    return {
+      isOriginClipped: isElementClippedByScrolling(originBounds, scrollContainerBounds),
+      isOriginOutsideView: isElementScrolledOutsideView(originBounds, scrollContainerBounds),
+      isOverlayClipped: isElementClippedByScrolling(overlayBounds, scrollContainerBounds),
+      isOverlayOutsideView: isElementScrolledOutsideView(overlayBounds, scrollContainerBounds)
+    };
+  }
+  /** Subtracts the amount that an element is overflowing on an axis from its length. */
+  _subtractOverflows(length, ...overflows) {
+    return overflows.reduce((currentValue, currentOverflow) => {
+      return currentValue - Math.max(currentOverflow, 0);
+    }, length);
+  }
+  /** Narrows the given viewport rect by the current _viewportMargin. */
+  _getNarrowedViewportRect() {
+    const width = this._document.documentElement.clientWidth;
+    const height = this._document.documentElement.clientHeight;
+    const scrollPosition = this._viewportRuler.getViewportScrollPosition();
+    return {
+      top: scrollPosition.top + this._viewportMargin,
+      left: scrollPosition.left + this._viewportMargin,
+      right: scrollPosition.left + width - this._viewportMargin,
+      bottom: scrollPosition.top + height - this._viewportMargin,
+      width: width - 2 * this._viewportMargin,
+      height: height - 2 * this._viewportMargin
+    };
+  }
+  /** Whether the we're dealing with an RTL context */
+  _isRtl() {
+    return this._overlayRef.getDirection() === "rtl";
+  }
+  /** Determines whether the overlay uses exact or flexible positioning. */
+  _hasExactPosition() {
+    return !this._hasFlexibleDimensions || this._isPushed;
+  }
+  /** Retrieves the offset of a position along the x or y axis. */
+  _getOffset(position, axis) {
+    if (axis === "x") {
+      return position.offsetX == null ? this._offsetX : position.offsetX;
+    }
+    return position.offsetY == null ? this._offsetY : position.offsetY;
+  }
+  /** Validates that the current position match the expected values. */
+  _validatePositions() {
+    if (typeof ngDevMode === "undefined" || ngDevMode) {
+      if (!this._preferredPositions.length) {
+        throw Error("FlexibleConnectedPositionStrategy: At least one position is required.");
+      }
+      this._preferredPositions.forEach((pair) => {
+        validateHorizontalPosition("originX", pair.originX);
+        validateVerticalPosition("originY", pair.originY);
+        validateHorizontalPosition("overlayX", pair.overlayX);
+        validateVerticalPosition("overlayY", pair.overlayY);
+      });
+    }
+  }
+  /** Adds a single CSS class or an array of classes on the overlay panel. */
+  _addPanelClasses(cssClasses) {
+    if (this._pane) {
+      coerceArray(cssClasses).forEach((cssClass) => {
+        if (cssClass !== "" && this._appliedPanelClasses.indexOf(cssClass) === -1) {
+          this._appliedPanelClasses.push(cssClass);
+          this._pane.classList.add(cssClass);
+        }
+      });
+    }
+  }
+  /** Clears the classes that the position strategy has applied from the overlay panel. */
+  _clearPanelClasses() {
+    if (this._pane) {
+      this._appliedPanelClasses.forEach((cssClass) => {
+        this._pane.classList.remove(cssClass);
+      });
+      this._appliedPanelClasses = [];
+    }
+  }
+  /** Returns the DOMRect of the current origin. */
+  _getOriginRect() {
+    const origin = this._origin;
+    if (origin instanceof ElementRef) {
+      return origin.nativeElement.getBoundingClientRect();
+    }
+    if (origin instanceof Element) {
+      return origin.getBoundingClientRect();
+    }
+    const width = origin.width || 0;
+    const height = origin.height || 0;
+    return {
+      top: origin.y,
+      bottom: origin.y + height,
+      left: origin.x,
+      right: origin.x + width,
+      height,
+      width
+    };
+  }
+};
+function extendStyles(destination, source) {
+  for (let key in source) {
+    if (source.hasOwnProperty(key)) {
+      destination[key] = source[key];
+    }
+  }
+  return destination;
+}
+function getPixelValue(input) {
+  if (typeof input !== "number" && input != null) {
+    const [value, units] = input.split(cssUnitPattern);
+    return !units || units === "px" ? parseFloat(value) : null;
+  }
+  return input || null;
+}
+function getRoundedBoundingClientRect(clientRect) {
+  return {
+    top: Math.floor(clientRect.top),
+    right: Math.floor(clientRect.right),
+    bottom: Math.floor(clientRect.bottom),
+    left: Math.floor(clientRect.left),
+    width: Math.floor(clientRect.width),
+    height: Math.floor(clientRect.height)
+  };
+}
+function compareScrollVisibility(a, b) {
+  if (a === b) {
+    return true;
+  }
+  return a.isOriginClipped === b.isOriginClipped && a.isOriginOutsideView === b.isOriginOutsideView && a.isOverlayClipped === b.isOverlayClipped && a.isOverlayOutsideView === b.isOverlayOutsideView;
+}
+var wrapperClass = "cdk-global-overlay-wrapper";
+var GlobalPositionStrategy = class {
+  constructor() {
+    this._cssPosition = "static";
+    this._topOffset = "";
+    this._bottomOffset = "";
+    this._alignItems = "";
+    this._xPosition = "";
+    this._xOffset = "";
+    this._width = "";
+    this._height = "";
+    this._isDisposed = false;
+  }
+  attach(overlayRef) {
+    const config = overlayRef.getConfig();
+    this._overlayRef = overlayRef;
+    if (this._width && !config.width) {
+      overlayRef.updateSize({
+        width: this._width
+      });
+    }
+    if (this._height && !config.height) {
+      overlayRef.updateSize({
+        height: this._height
+      });
+    }
+    overlayRef.hostElement.classList.add(wrapperClass);
+    this._isDisposed = false;
+  }
+  /**
+   * Sets the top position of the overlay. Clears any previously set vertical position.
+   * @param value New top offset.
+   */
+  top(value = "") {
+    this._bottomOffset = "";
+    this._topOffset = value;
+    this._alignItems = "flex-start";
+    return this;
+  }
+  /**
+   * Sets the left position of the overlay. Clears any previously set horizontal position.
+   * @param value New left offset.
+   */
+  left(value = "") {
+    this._xOffset = value;
+    this._xPosition = "left";
+    return this;
+  }
+  /**
+   * Sets the bottom position of the overlay. Clears any previously set vertical position.
+   * @param value New bottom offset.
+   */
+  bottom(value = "") {
+    this._topOffset = "";
+    this._bottomOffset = value;
+    this._alignItems = "flex-end";
+    return this;
+  }
+  /**
+   * Sets the right position of the overlay. Clears any previously set horizontal position.
+   * @param value New right offset.
+   */
+  right(value = "") {
+    this._xOffset = value;
+    this._xPosition = "right";
+    return this;
+  }
+  /**
+   * Sets the overlay to the start of the viewport, depending on the overlay direction.
+   * This will be to the left in LTR layouts and to the right in RTL.
+   * @param offset Offset from the edge of the screen.
+   */
+  start(value = "") {
+    this._xOffset = value;
+    this._xPosition = "start";
+    return this;
+  }
+  /**
+   * Sets the overlay to the end of the viewport, depending on the overlay direction.
+   * This will be to the right in LTR layouts and to the left in RTL.
+   * @param offset Offset from the edge of the screen.
+   */
+  end(value = "") {
+    this._xOffset = value;
+    this._xPosition = "end";
+    return this;
+  }
+  /**
+   * Sets the overlay width and clears any previously set width.
+   * @param value New width for the overlay
+   * @deprecated Pass the `width` through the `OverlayConfig`.
+   * @breaking-change 8.0.0
+   */
+  width(value = "") {
+    if (this._overlayRef) {
+      this._overlayRef.updateSize({
+        width: value
+      });
+    } else {
+      this._width = value;
+    }
+    return this;
+  }
+  /**
+   * Sets the overlay height and clears any previously set height.
+   * @param value New height for the overlay
+   * @deprecated Pass the `height` through the `OverlayConfig`.
+   * @breaking-change 8.0.0
+   */
+  height(value = "") {
+    if (this._overlayRef) {
+      this._overlayRef.updateSize({
+        height: value
+      });
+    } else {
+      this._height = value;
+    }
+    return this;
+  }
+  /**
+   * Centers the overlay horizontally with an optional offset.
+   * Clears any previously set horizontal position.
+   *
+   * @param offset Overlay offset from the horizontal center.
+   */
+  centerHorizontally(offset = "") {
+    this.left(offset);
+    this._xPosition = "center";
+    return this;
+  }
+  /**
+   * Centers the overlay vertically with an optional offset.
+   * Clears any previously set vertical position.
+   *
+   * @param offset Overlay offset from the vertical center.
+   */
+  centerVertically(offset = "") {
+    this.top(offset);
+    this._alignItems = "center";
+    return this;
+  }
+  /**
+   * Apply the position to the element.
+   * @docs-private
+   */
+  apply() {
+    if (!this._overlayRef || !this._overlayRef.hasAttached()) {
+      return;
+    }
+    const styles = this._overlayRef.overlayElement.style;
+    const parentStyles = this._overlayRef.hostElement.style;
+    const config = this._overlayRef.getConfig();
+    const {
+      width,
+      height,
+      maxWidth,
+      maxHeight
+    } = config;
+    const shouldBeFlushHorizontally = (width === "100%" || width === "100vw") && (!maxWidth || maxWidth === "100%" || maxWidth === "100vw");
+    const shouldBeFlushVertically = (height === "100%" || height === "100vh") && (!maxHeight || maxHeight === "100%" || maxHeight === "100vh");
+    const xPosition = this._xPosition;
+    const xOffset = this._xOffset;
+    const isRtl = this._overlayRef.getConfig().direction === "rtl";
+    let marginLeft = "";
+    let marginRight = "";
+    let justifyContent = "";
+    if (shouldBeFlushHorizontally) {
+      justifyContent = "flex-start";
+    } else if (xPosition === "center") {
+      justifyContent = "center";
+      if (isRtl) {
+        marginRight = xOffset;
+      } else {
+        marginLeft = xOffset;
+      }
+    } else if (isRtl) {
+      if (xPosition === "left" || xPosition === "end") {
+        justifyContent = "flex-end";
+        marginLeft = xOffset;
+      } else if (xPosition === "right" || xPosition === "start") {
+        justifyContent = "flex-start";
+        marginRight = xOffset;
+      }
+    } else if (xPosition === "left" || xPosition === "start") {
+      justifyContent = "flex-start";
+      marginLeft = xOffset;
+    } else if (xPosition === "right" || xPosition === "end") {
+      justifyContent = "flex-end";
+      marginRight = xOffset;
+    }
+    styles.position = this._cssPosition;
+    styles.marginLeft = shouldBeFlushHorizontally ? "0" : marginLeft;
+    styles.marginTop = shouldBeFlushVertically ? "0" : this._topOffset;
+    styles.marginBottom = this._bottomOffset;
+    styles.marginRight = shouldBeFlushHorizontally ? "0" : marginRight;
+    parentStyles.justifyContent = justifyContent;
+    parentStyles.alignItems = shouldBeFlushVertically ? "flex-start" : this._alignItems;
+  }
+  /**
+   * Cleans up the DOM changes from the position strategy.
+   * @docs-private
+   */
+  dispose() {
+    if (this._isDisposed || !this._overlayRef) {
+      return;
+    }
+    const styles = this._overlayRef.overlayElement.style;
+    const parent = this._overlayRef.hostElement;
+    const parentStyles = parent.style;
+    parent.classList.remove(wrapperClass);
+    parentStyles.justifyContent = parentStyles.alignItems = styles.marginTop = styles.marginBottom = styles.marginLeft = styles.marginRight = styles.position = "";
+    this._overlayRef = null;
+    this._isDisposed = true;
+  }
+};
+var _OverlayPositionBuilder = class _OverlayPositionBuilder {
+  constructor(_viewportRuler, _document, _platform, _overlayContainer) {
+    this._viewportRuler = _viewportRuler;
+    this._document = _document;
+    this._platform = _platform;
+    this._overlayContainer = _overlayContainer;
+  }
+  /**
+   * Creates a global position strategy.
+   */
+  global() {
+    return new GlobalPositionStrategy();
+  }
+  /**
+   * Creates a flexible position strategy.
+   * @param origin Origin relative to which to position the overlay.
+   */
+  flexibleConnectedTo(origin) {
+    return new FlexibleConnectedPositionStrategy(origin, this._viewportRuler, this._document, this._platform, this._overlayContainer);
+  }
+};
+_OverlayPositionBuilder.ɵfac = function OverlayPositionBuilder_Factory(t) {
+  return new (t || _OverlayPositionBuilder)(ɵɵinject(ViewportRuler), ɵɵinject(DOCUMENT), ɵɵinject(Platform), ɵɵinject(OverlayContainer));
+};
+_OverlayPositionBuilder.ɵprov = ɵɵdefineInjectable({
+  token: _OverlayPositionBuilder,
+  factory: _OverlayPositionBuilder.ɵfac,
+  providedIn: "root"
+});
+var OverlayPositionBuilder = _OverlayPositionBuilder;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(OverlayPositionBuilder, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: ViewportRuler
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: Platform
+  }, {
+    type: OverlayContainer
+  }], null);
+})();
+var nextUniqueId = 0;
+var _Overlay = class _Overlay {
+  constructor(scrollStrategies, _overlayContainer, _componentFactoryResolver, _positionBuilder, _keyboardDispatcher, _injector, _ngZone, _document, _directionality, _location, _outsideClickDispatcher, _animationsModuleType) {
+    this.scrollStrategies = scrollStrategies;
+    this._overlayContainer = _overlayContainer;
+    this._componentFactoryResolver = _componentFactoryResolver;
+    this._positionBuilder = _positionBuilder;
+    this._keyboardDispatcher = _keyboardDispatcher;
+    this._injector = _injector;
+    this._ngZone = _ngZone;
+    this._document = _document;
+    this._directionality = _directionality;
+    this._location = _location;
+    this._outsideClickDispatcher = _outsideClickDispatcher;
+    this._animationsModuleType = _animationsModuleType;
+  }
+  /**
+   * Creates an overlay.
+   * @param config Configuration applied to the overlay.
+   * @returns Reference to the created overlay.
+   */
+  create(config) {
+    const host = this._createHostElement();
+    const pane = this._createPaneElement(host);
+    const portalOutlet = this._createPortalOutlet(pane);
+    const overlayConfig = new OverlayConfig(config);
+    overlayConfig.direction = overlayConfig.direction || this._directionality.value;
+    return new OverlayRef(portalOutlet, host, pane, overlayConfig, this._ngZone, this._keyboardDispatcher, this._document, this._location, this._outsideClickDispatcher, this._animationsModuleType === "NoopAnimations");
+  }
+  /**
+   * Gets a position builder that can be used, via fluent API,
+   * to construct and configure a position strategy.
+   * @returns An overlay position builder.
+   */
+  position() {
+    return this._positionBuilder;
+  }
+  /**
+   * Creates the DOM element for an overlay and appends it to the overlay container.
+   * @returns Newly-created pane element
+   */
+  _createPaneElement(host) {
+    const pane = this._document.createElement("div");
+    pane.id = `cdk-overlay-${nextUniqueId++}`;
+    pane.classList.add("cdk-overlay-pane");
+    host.appendChild(pane);
+    return pane;
+  }
+  /**
+   * Creates the host element that wraps around an overlay
+   * and can be used for advanced positioning.
+   * @returns Newly-create host element.
+   */
+  _createHostElement() {
+    const host = this._document.createElement("div");
+    this._overlayContainer.getContainerElement().appendChild(host);
+    return host;
+  }
+  /**
+   * Create a DomPortalOutlet into which the overlay content can be loaded.
+   * @param pane The DOM element to turn into a portal outlet.
+   * @returns A portal outlet for the given DOM element.
+   */
+  _createPortalOutlet(pane) {
+    if (!this._appRef) {
+      this._appRef = this._injector.get(ApplicationRef);
+    }
+    return new DomPortalOutlet(pane, this._componentFactoryResolver, this._appRef, this._injector, this._document);
+  }
+};
+_Overlay.ɵfac = function Overlay_Factory(t) {
+  return new (t || _Overlay)(ɵɵinject(ScrollStrategyOptions), ɵɵinject(OverlayContainer), ɵɵinject(ComponentFactoryResolver$1), ɵɵinject(OverlayPositionBuilder), ɵɵinject(OverlayKeyboardDispatcher), ɵɵinject(Injector), ɵɵinject(NgZone), ɵɵinject(DOCUMENT), ɵɵinject(Directionality), ɵɵinject(Location), ɵɵinject(OverlayOutsideClickDispatcher), ɵɵinject(ANIMATION_MODULE_TYPE, 8));
+};
+_Overlay.ɵprov = ɵɵdefineInjectable({
+  token: _Overlay,
+  factory: _Overlay.ɵfac,
+  providedIn: "root"
+});
+var Overlay = _Overlay;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(Overlay, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: ScrollStrategyOptions
+  }, {
+    type: OverlayContainer
+  }, {
+    type: ComponentFactoryResolver$1
+  }, {
+    type: OverlayPositionBuilder
+  }, {
+    type: OverlayKeyboardDispatcher
+  }, {
+    type: Injector
+  }, {
+    type: NgZone
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: Directionality
+  }, {
+    type: Location
+  }, {
+    type: OverlayOutsideClickDispatcher
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [ANIMATION_MODULE_TYPE]
+    }, {
+      type: Optional
+    }]
+  }], null);
+})();
+var defaultPositionList = [{
+  originX: "start",
+  originY: "bottom",
+  overlayX: "start",
+  overlayY: "top"
+}, {
+  originX: "start",
+  originY: "top",
+  overlayX: "start",
+  overlayY: "bottom"
+}, {
+  originX: "end",
+  originY: "top",
+  overlayX: "end",
+  overlayY: "bottom"
+}, {
+  originX: "end",
+  originY: "bottom",
+  overlayX: "end",
+  overlayY: "top"
+}];
+var CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY = new InjectionToken("cdk-connected-overlay-scroll-strategy", {
+  providedIn: "root",
+  factory: () => {
+    const overlay = inject(Overlay);
+    return () => overlay.scrollStrategies.reposition();
+  }
+});
+var _CdkOverlayOrigin = class _CdkOverlayOrigin {
+  constructor(elementRef) {
+    this.elementRef = elementRef;
+  }
+};
+_CdkOverlayOrigin.ɵfac = function CdkOverlayOrigin_Factory(t) {
+  return new (t || _CdkOverlayOrigin)(ɵɵdirectiveInject(ElementRef));
+};
+_CdkOverlayOrigin.ɵdir = ɵɵdefineDirective({
+  type: _CdkOverlayOrigin,
+  selectors: [["", "cdk-overlay-origin", ""], ["", "overlay-origin", ""], ["", "cdkOverlayOrigin", ""]],
+  exportAs: ["cdkOverlayOrigin"],
+  standalone: true
+});
+var CdkOverlayOrigin = _CdkOverlayOrigin;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(CdkOverlayOrigin, [{
+    type: Directive,
+    args: [{
+      selector: "[cdk-overlay-origin], [overlay-origin], [cdkOverlayOrigin]",
+      exportAs: "cdkOverlayOrigin",
+      standalone: true
+    }]
+  }], () => [{
+    type: ElementRef
+  }], null);
+})();
+var _CdkConnectedOverlay = class _CdkConnectedOverlay {
+  /** The offset in pixels for the overlay connection point on the x-axis */
+  get offsetX() {
+    return this._offsetX;
+  }
+  set offsetX(offsetX) {
+    this._offsetX = offsetX;
+    if (this._position) {
+      this._updatePositionStrategy(this._position);
+    }
+  }
+  /** The offset in pixels for the overlay connection point on the y-axis */
+  get offsetY() {
+    return this._offsetY;
+  }
+  set offsetY(offsetY) {
+    this._offsetY = offsetY;
+    if (this._position) {
+      this._updatePositionStrategy(this._position);
+    }
+  }
+  /** Whether the overlay should be disposed of when the user goes backwards/forwards in history. */
+  get disposeOnNavigation() {
+    return this._disposeOnNavigation;
+  }
+  set disposeOnNavigation(value) {
+    this._disposeOnNavigation = value;
+  }
+  // TODO(jelbourn): inputs for size, scroll behavior, animation, etc.
+  constructor(_overlay, templateRef, viewContainerRef, scrollStrategyFactory, _dir) {
+    this._overlay = _overlay;
+    this._dir = _dir;
+    this._backdropSubscription = Subscription.EMPTY;
+    this._attachSubscription = Subscription.EMPTY;
+    this._detachSubscription = Subscription.EMPTY;
+    this._positionSubscription = Subscription.EMPTY;
+    this._disposeOnNavigation = false;
+    this._ngZone = inject(NgZone);
+    this.viewportMargin = 0;
+    this.open = false;
+    this.disableClose = false;
+    this.hasBackdrop = false;
+    this.lockPosition = false;
+    this.flexibleDimensions = false;
+    this.growAfterOpen = false;
+    this.push = false;
+    this.backdropClick = new EventEmitter();
+    this.positionChange = new EventEmitter();
+    this.attach = new EventEmitter();
+    this.detach = new EventEmitter();
+    this.overlayKeydown = new EventEmitter();
+    this.overlayOutsideClick = new EventEmitter();
+    this._templatePortal = new TemplatePortal(templateRef, viewContainerRef);
+    this._scrollStrategyFactory = scrollStrategyFactory;
+    this.scrollStrategy = this._scrollStrategyFactory();
+  }
+  /** The associated overlay reference. */
+  get overlayRef() {
+    return this._overlayRef;
+  }
+  /** The element's layout direction. */
+  get dir() {
+    return this._dir ? this._dir.value : "ltr";
+  }
+  ngOnDestroy() {
+    this._attachSubscription.unsubscribe();
+    this._detachSubscription.unsubscribe();
+    this._backdropSubscription.unsubscribe();
+    this._positionSubscription.unsubscribe();
+    if (this._overlayRef) {
+      this._overlayRef.dispose();
+    }
+  }
+  ngOnChanges(changes) {
+    if (this._position) {
+      this._updatePositionStrategy(this._position);
+      this._overlayRef.updateSize({
+        width: this.width,
+        minWidth: this.minWidth,
+        height: this.height,
+        minHeight: this.minHeight
+      });
+      if (changes["origin"] && this.open) {
+        this._position.apply();
+      }
+    }
+    if (changes["open"]) {
+      this.open ? this._attachOverlay() : this._detachOverlay();
+    }
+  }
+  /** Creates an overlay */
+  _createOverlay() {
+    if (!this.positions || !this.positions.length) {
+      this.positions = defaultPositionList;
+    }
+    const overlayRef = this._overlayRef = this._overlay.create(this._buildConfig());
+    this._attachSubscription = overlayRef.attachments().subscribe(() => this.attach.emit());
+    this._detachSubscription = overlayRef.detachments().subscribe(() => this.detach.emit());
+    overlayRef.keydownEvents().subscribe((event) => {
+      this.overlayKeydown.next(event);
+      if (event.keyCode === ESCAPE && !this.disableClose && !hasModifierKey(event)) {
+        event.preventDefault();
+        this._detachOverlay();
+      }
+    });
+    this._overlayRef.outsidePointerEvents().subscribe((event) => {
+      const origin = this._getOriginElement();
+      const target = _getEventTarget(event);
+      if (!origin || origin !== target && !origin.contains(target)) {
+        this.overlayOutsideClick.next(event);
+      }
+    });
+  }
+  /** Builds the overlay config based on the directive's inputs */
+  _buildConfig() {
+    const positionStrategy = this._position = this.positionStrategy || this._createPositionStrategy();
+    const overlayConfig = new OverlayConfig({
+      direction: this._dir,
+      positionStrategy,
+      scrollStrategy: this.scrollStrategy,
+      hasBackdrop: this.hasBackdrop,
+      disposeOnNavigation: this.disposeOnNavigation
+    });
+    if (this.width || this.width === 0) {
+      overlayConfig.width = this.width;
+    }
+    if (this.height || this.height === 0) {
+      overlayConfig.height = this.height;
+    }
+    if (this.minWidth || this.minWidth === 0) {
+      overlayConfig.minWidth = this.minWidth;
+    }
+    if (this.minHeight || this.minHeight === 0) {
+      overlayConfig.minHeight = this.minHeight;
+    }
+    if (this.backdropClass) {
+      overlayConfig.backdropClass = this.backdropClass;
+    }
+    if (this.panelClass) {
+      overlayConfig.panelClass = this.panelClass;
+    }
+    return overlayConfig;
+  }
+  /** Updates the state of a position strategy, based on the values of the directive inputs. */
+  _updatePositionStrategy(positionStrategy) {
+    const positions = this.positions.map((currentPosition) => ({
+      originX: currentPosition.originX,
+      originY: currentPosition.originY,
+      overlayX: currentPosition.overlayX,
+      overlayY: currentPosition.overlayY,
+      offsetX: currentPosition.offsetX || this.offsetX,
+      offsetY: currentPosition.offsetY || this.offsetY,
+      panelClass: currentPosition.panelClass || void 0
+    }));
+    return positionStrategy.setOrigin(this._getOrigin()).withPositions(positions).withFlexibleDimensions(this.flexibleDimensions).withPush(this.push).withGrowAfterOpen(this.growAfterOpen).withViewportMargin(this.viewportMargin).withLockedPosition(this.lockPosition).withTransformOriginOn(this.transformOriginSelector);
+  }
+  /** Returns the position strategy of the overlay to be set on the overlay config */
+  _createPositionStrategy() {
+    const strategy = this._overlay.position().flexibleConnectedTo(this._getOrigin());
+    this._updatePositionStrategy(strategy);
+    return strategy;
+  }
+  _getOrigin() {
+    if (this.origin instanceof CdkOverlayOrigin) {
+      return this.origin.elementRef;
+    } else {
+      return this.origin;
+    }
+  }
+  _getOriginElement() {
+    if (this.origin instanceof CdkOverlayOrigin) {
+      return this.origin.elementRef.nativeElement;
+    }
+    if (this.origin instanceof ElementRef) {
+      return this.origin.nativeElement;
+    }
+    if (typeof Element !== "undefined" && this.origin instanceof Element) {
+      return this.origin;
+    }
+    return null;
+  }
+  /** Attaches the overlay and subscribes to backdrop clicks if backdrop exists */
+  _attachOverlay() {
+    if (!this._overlayRef) {
+      this._createOverlay();
+    } else {
+      this._overlayRef.getConfig().hasBackdrop = this.hasBackdrop;
+    }
+    if (!this._overlayRef.hasAttached()) {
+      this._overlayRef.attach(this._templatePortal);
+    }
+    if (this.hasBackdrop) {
+      this._backdropSubscription = this._overlayRef.backdropClick().subscribe((event) => {
+        this.backdropClick.emit(event);
+      });
+    } else {
+      this._backdropSubscription.unsubscribe();
+    }
+    this._positionSubscription.unsubscribe();
+    if (this.positionChange.observers.length > 0) {
+      this._positionSubscription = this._position.positionChanges.pipe(takeWhile(() => this.positionChange.observers.length > 0)).subscribe((position) => {
+        this._ngZone.run(() => this.positionChange.emit(position));
+        if (this.positionChange.observers.length === 0) {
+          this._positionSubscription.unsubscribe();
+        }
+      });
+    }
+  }
+  /** Detaches the overlay and unsubscribes to backdrop clicks if backdrop exists */
+  _detachOverlay() {
+    if (this._overlayRef) {
+      this._overlayRef.detach();
+    }
+    this._backdropSubscription.unsubscribe();
+    this._positionSubscription.unsubscribe();
+  }
+};
+_CdkConnectedOverlay.ɵfac = function CdkConnectedOverlay_Factory(t) {
+  return new (t || _CdkConnectedOverlay)(ɵɵdirectiveInject(Overlay), ɵɵdirectiveInject(TemplateRef), ɵɵdirectiveInject(ViewContainerRef), ɵɵdirectiveInject(CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY), ɵɵdirectiveInject(Directionality, 8));
+};
+_CdkConnectedOverlay.ɵdir = ɵɵdefineDirective({
+  type: _CdkConnectedOverlay,
+  selectors: [["", "cdk-connected-overlay", ""], ["", "connected-overlay", ""], ["", "cdkConnectedOverlay", ""]],
+  inputs: {
+    origin: [InputFlags.None, "cdkConnectedOverlayOrigin", "origin"],
+    positions: [InputFlags.None, "cdkConnectedOverlayPositions", "positions"],
+    positionStrategy: [InputFlags.None, "cdkConnectedOverlayPositionStrategy", "positionStrategy"],
+    offsetX: [InputFlags.None, "cdkConnectedOverlayOffsetX", "offsetX"],
+    offsetY: [InputFlags.None, "cdkConnectedOverlayOffsetY", "offsetY"],
+    width: [InputFlags.None, "cdkConnectedOverlayWidth", "width"],
+    height: [InputFlags.None, "cdkConnectedOverlayHeight", "height"],
+    minWidth: [InputFlags.None, "cdkConnectedOverlayMinWidth", "minWidth"],
+    minHeight: [InputFlags.None, "cdkConnectedOverlayMinHeight", "minHeight"],
+    backdropClass: [InputFlags.None, "cdkConnectedOverlayBackdropClass", "backdropClass"],
+    panelClass: [InputFlags.None, "cdkConnectedOverlayPanelClass", "panelClass"],
+    viewportMargin: [InputFlags.None, "cdkConnectedOverlayViewportMargin", "viewportMargin"],
+    scrollStrategy: [InputFlags.None, "cdkConnectedOverlayScrollStrategy", "scrollStrategy"],
+    open: [InputFlags.None, "cdkConnectedOverlayOpen", "open"],
+    disableClose: [InputFlags.None, "cdkConnectedOverlayDisableClose", "disableClose"],
+    transformOriginSelector: [InputFlags.None, "cdkConnectedOverlayTransformOriginOn", "transformOriginSelector"],
+    hasBackdrop: [InputFlags.HasDecoratorInputTransform, "cdkConnectedOverlayHasBackdrop", "hasBackdrop", booleanAttribute],
+    lockPosition: [InputFlags.HasDecoratorInputTransform, "cdkConnectedOverlayLockPosition", "lockPosition", booleanAttribute],
+    flexibleDimensions: [InputFlags.HasDecoratorInputTransform, "cdkConnectedOverlayFlexibleDimensions", "flexibleDimensions", booleanAttribute],
+    growAfterOpen: [InputFlags.HasDecoratorInputTransform, "cdkConnectedOverlayGrowAfterOpen", "growAfterOpen", booleanAttribute],
+    push: [InputFlags.HasDecoratorInputTransform, "cdkConnectedOverlayPush", "push", booleanAttribute],
+    disposeOnNavigation: [InputFlags.HasDecoratorInputTransform, "cdkConnectedOverlayDisposeOnNavigation", "disposeOnNavigation", booleanAttribute]
+  },
+  outputs: {
+    backdropClick: "backdropClick",
+    positionChange: "positionChange",
+    attach: "attach",
+    detach: "detach",
+    overlayKeydown: "overlayKeydown",
+    overlayOutsideClick: "overlayOutsideClick"
+  },
+  exportAs: ["cdkConnectedOverlay"],
+  standalone: true,
+  features: [ɵɵInputTransformsFeature, ɵɵNgOnChangesFeature]
+});
+var CdkConnectedOverlay = _CdkConnectedOverlay;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(CdkConnectedOverlay, [{
+    type: Directive,
+    args: [{
+      selector: "[cdk-connected-overlay], [connected-overlay], [cdkConnectedOverlay]",
+      exportAs: "cdkConnectedOverlay",
+      standalone: true
+    }]
+  }], () => [{
+    type: Overlay
+  }, {
+    type: TemplateRef
+  }, {
+    type: ViewContainerRef
+  }, {
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY]
+    }]
+  }, {
+    type: Directionality,
+    decorators: [{
+      type: Optional
+    }]
+  }], {
+    origin: [{
+      type: Input,
+      args: ["cdkConnectedOverlayOrigin"]
+    }],
+    positions: [{
+      type: Input,
+      args: ["cdkConnectedOverlayPositions"]
+    }],
+    positionStrategy: [{
+      type: Input,
+      args: ["cdkConnectedOverlayPositionStrategy"]
+    }],
+    offsetX: [{
+      type: Input,
+      args: ["cdkConnectedOverlayOffsetX"]
+    }],
+    offsetY: [{
+      type: Input,
+      args: ["cdkConnectedOverlayOffsetY"]
+    }],
+    width: [{
+      type: Input,
+      args: ["cdkConnectedOverlayWidth"]
+    }],
+    height: [{
+      type: Input,
+      args: ["cdkConnectedOverlayHeight"]
+    }],
+    minWidth: [{
+      type: Input,
+      args: ["cdkConnectedOverlayMinWidth"]
+    }],
+    minHeight: [{
+      type: Input,
+      args: ["cdkConnectedOverlayMinHeight"]
+    }],
+    backdropClass: [{
+      type: Input,
+      args: ["cdkConnectedOverlayBackdropClass"]
+    }],
+    panelClass: [{
+      type: Input,
+      args: ["cdkConnectedOverlayPanelClass"]
+    }],
+    viewportMargin: [{
+      type: Input,
+      args: ["cdkConnectedOverlayViewportMargin"]
+    }],
+    scrollStrategy: [{
+      type: Input,
+      args: ["cdkConnectedOverlayScrollStrategy"]
+    }],
+    open: [{
+      type: Input,
+      args: ["cdkConnectedOverlayOpen"]
+    }],
+    disableClose: [{
+      type: Input,
+      args: ["cdkConnectedOverlayDisableClose"]
+    }],
+    transformOriginSelector: [{
+      type: Input,
+      args: ["cdkConnectedOverlayTransformOriginOn"]
+    }],
+    hasBackdrop: [{
+      type: Input,
+      args: [{
+        alias: "cdkConnectedOverlayHasBackdrop",
+        transform: booleanAttribute
+      }]
+    }],
+    lockPosition: [{
+      type: Input,
+      args: [{
+        alias: "cdkConnectedOverlayLockPosition",
+        transform: booleanAttribute
+      }]
+    }],
+    flexibleDimensions: [{
+      type: Input,
+      args: [{
+        alias: "cdkConnectedOverlayFlexibleDimensions",
+        transform: booleanAttribute
+      }]
+    }],
+    growAfterOpen: [{
+      type: Input,
+      args: [{
+        alias: "cdkConnectedOverlayGrowAfterOpen",
+        transform: booleanAttribute
+      }]
+    }],
+    push: [{
+      type: Input,
+      args: [{
+        alias: "cdkConnectedOverlayPush",
+        transform: booleanAttribute
+      }]
+    }],
+    disposeOnNavigation: [{
+      type: Input,
+      args: [{
+        alias: "cdkConnectedOverlayDisposeOnNavigation",
+        transform: booleanAttribute
+      }]
+    }],
+    backdropClick: [{
+      type: Output
+    }],
+    positionChange: [{
+      type: Output
+    }],
+    attach: [{
+      type: Output
+    }],
+    detach: [{
+      type: Output
+    }],
+    overlayKeydown: [{
+      type: Output
+    }],
+    overlayOutsideClick: [{
+      type: Output
+    }]
+  });
+})();
+function CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY_PROVIDER_FACTORY(overlay) {
+  return () => overlay.scrollStrategies.reposition();
+}
+var CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY_PROVIDER = {
+  provide: CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY,
+  deps: [Overlay],
+  useFactory: CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY_PROVIDER_FACTORY
+};
+var _OverlayModule = class _OverlayModule {
+};
+_OverlayModule.ɵfac = function OverlayModule_Factory(t) {
+  return new (t || _OverlayModule)();
+};
+_OverlayModule.ɵmod = ɵɵdefineNgModule({
+  type: _OverlayModule,
+  imports: [BidiModule, PortalModule, ScrollingModule, CdkConnectedOverlay, CdkOverlayOrigin],
+  exports: [CdkConnectedOverlay, CdkOverlayOrigin, ScrollingModule]
+});
+_OverlayModule.ɵinj = ɵɵdefineInjector({
+  providers: [Overlay, CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY_PROVIDER],
+  imports: [BidiModule, PortalModule, ScrollingModule, ScrollingModule]
+});
+var OverlayModule = _OverlayModule;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(OverlayModule, [{
+    type: NgModule,
+    args: [{
+      imports: [BidiModule, PortalModule, ScrollingModule, CdkConnectedOverlay, CdkOverlayOrigin],
+      exports: [CdkConnectedOverlay, CdkOverlayOrigin, ScrollingModule],
+      providers: [Overlay, CDK_CONNECTED_OVERLAY_SCROLL_STRATEGY_PROVIDER]
+    }]
+  }], null, null);
+})();
+var _FullscreenOverlayContainer = class _FullscreenOverlayContainer extends OverlayContainer {
+  constructor(_document, platform) {
+    super(_document, platform);
+  }
+  ngOnDestroy() {
+    super.ngOnDestroy();
+    if (this._fullScreenEventName && this._fullScreenListener) {
+      this._document.removeEventListener(this._fullScreenEventName, this._fullScreenListener);
+    }
+  }
+  _createContainer() {
+    super._createContainer();
+    this._adjustParentForFullscreenChange();
+    this._addFullscreenChangeListener(() => this._adjustParentForFullscreenChange());
+  }
+  _adjustParentForFullscreenChange() {
+    if (!this._containerElement) {
+      return;
+    }
+    const fullscreenElement = this.getFullscreenElement();
+    const parent = fullscreenElement || this._document.body;
+    parent.appendChild(this._containerElement);
+  }
+  _addFullscreenChangeListener(fn) {
+    const eventName = this._getEventName();
+    if (eventName) {
+      if (this._fullScreenListener) {
+        this._document.removeEventListener(eventName, this._fullScreenListener);
+      }
+      this._document.addEventListener(eventName, fn);
+      this._fullScreenListener = fn;
+    }
+  }
+  _getEventName() {
+    if (!this._fullScreenEventName) {
+      const _document = this._document;
+      if (_document.fullscreenEnabled) {
+        this._fullScreenEventName = "fullscreenchange";
+      } else if (_document.webkitFullscreenEnabled) {
+        this._fullScreenEventName = "webkitfullscreenchange";
+      } else if (_document.mozFullScreenEnabled) {
+        this._fullScreenEventName = "mozfullscreenchange";
+      } else if (_document.msFullscreenEnabled) {
+        this._fullScreenEventName = "MSFullscreenChange";
+      }
+    }
+    return this._fullScreenEventName;
+  }
+  /**
+   * When the page is put into fullscreen mode, a specific element is specified.
+   * Only that element and its children are visible when in fullscreen mode.
+   */
+  getFullscreenElement() {
+    const _document = this._document;
+    return _document.fullscreenElement || _document.webkitFullscreenElement || _document.mozFullScreenElement || _document.msFullscreenElement || null;
+  }
+};
+_FullscreenOverlayContainer.ɵfac = function FullscreenOverlayContainer_Factory(t) {
+  return new (t || _FullscreenOverlayContainer)(ɵɵinject(DOCUMENT), ɵɵinject(Platform));
+};
+_FullscreenOverlayContainer.ɵprov = ɵɵdefineInjectable({
+  token: _FullscreenOverlayContainer,
+  factory: _FullscreenOverlayContainer.ɵfac,
+  providedIn: "root"
+});
+var FullscreenOverlayContainer = _FullscreenOverlayContainer;
+(() => {
+  (typeof ngDevMode === "undefined" || ngDevMode) && setClassMetadata(FullscreenOverlayContainer, [{
+    type: Injectable,
+    args: [{
+      providedIn: "root"
+    }]
+  }], () => [{
+    type: void 0,
+    decorators: [{
+      type: Inject,
+      args: [DOCUMENT]
+    }]
+  }, {
+    type: Platform
+  }], null);
+})();
+
+export {
+  ComponentPortal,
+  TemplatePortal,
+  BasePortalOutlet,
+  CdkPortalOutlet,
+  PortalModule,
+  OverlayConfig,
+  OverlayContainer,
+  OverlayRef,
+  Overlay,
+  CdkOverlayOrigin,
+  CdkConnectedOverlay,
+  OverlayModule
+};
+//# sourceMappingURL=chunk-RM7V6FV7.js.map
