+++ Aulas-Angular/app-modules/.angular/cache/17.3.8/vite/deps/chunk-SXOQ3UHC.js
@@ -4,14 +4,14 @@ import {
   style,
   transition,
   trigger
-} from "./chunk-2AXA5NS2.js";
+} from "./chunk-VNW2MBA7.js";
 import {
   Directionality,
   MatCommonModule,
   ObserversModule,
   Platform,
   coerceBooleanProperty
-} from "./chunk-FLGVYJXI.js";
+} from "./chunk-7ZFYYFG2.js";
 import {
   CommonModule,
   DOCUMENT,
@@ -1586,4 +1586,4 @@ export {
   MatFormField,
   MatFormFieldModule
 };
-//# sourceMappingURL=chunk-C2Z32BJ6.js.map
+//# sourceMappingURL=chunk-SXOQ3UHC.js.map
