+++ Aulas-Angular/app-modules/src/app/modules/private/user-area/user-area.component.html
@@ -1,57 +1,41 @@
 <div class="cont">
-<div class="novo">
-  <button>
- <img src="../../../../assets/add-36.png" alt=""> Novo projeto 
-</button>
-</div>
+  <div class="novo">
+    <button>
+      <img src="../../../../assets/add-36.png" alt="" /> Novo projeto
+    </button>
+  </div>
 
-<table mat-table [dataSource]="projects" class="mat-elevation-z8">
-  
-  <ng-container class="posi-cont" matColumnDef="position">
-    <th  class="posi" *matHeaderCellDef>N°</th>
-    <td class="row" *matCellDef="let element">{{ element.position }}</td>
-  </ng-container>
+  <table mat-table [dataSource]="projects" class="mat-elevation-z8">
+    <ng-container class="posi-cont" matColumnDef="position">
+      <th class="posi" *matHeaderCellDef>N°</th>
+      <td *matCellDef="let element">{{ element.position }}</td>
+    </ng-container>
 
-  <ng-container  matColumnDef="project">
-    <th class="proj" *matHeaderCellDef>Projetos</th>
-    <td  *matCellDef="let element">{{ element.title }}</td>
-  </ng-container>
+    <ng-container matColumnDef="project">
+      <th class="proj" *matHeaderCellDef>Projetos</th>
+      <td *matCellDef="let element">{{ element.title }}</td>
+    </ng-container>
 
-  <ng-container  matColumnDef="task">
-    <th class="tsk" *matHeaderCellDef>Tasks</th>
-    <td class="row" *matCellDef="let element">{{ element.completedTasks }}/{{element.tasks.length}}</td>
-  </ng-container>
+    <ng-container matColumnDef="task">
+      <th class="tsk" *matHeaderCellDef>Tasks</th>
+      <td *matCellDef="let element">
+        {{ element.completedTasks }}/{{ element.tasks.length }}
+      </td>
+    </ng-container>
 
-  <ng-container  matColumnDef="action">
-    <th class="acao" *matHeaderCellDef>Ações</th>
-    <td class="row" *matCellDef="let element">
-      <button routerLink="/project-area/{{element._id}}"><img src="../../../../assets/visibility-24.png" alt=""></button>
-      <button (click)="openModal(element._id)"><img src="../../../../assets/delete-24.png" alt=""></button>
-    </td>
-  </ng-container>
+    <ng-container matColumnDef="action">
+      <th class="acao" *matHeaderCellDef>Ações</th>
+      <td *matCellDef="let element">
+        <button routerLink="/project-area/{{ element._id }}">
+          <img src="../../../../assets/visibility-24.png" alt="" />
+        </button>
+        <button (click)="openModal(element._id)">
+          <img src="../../../../assets/delete-24.png" alt="" />
+        </button>
+      </td>
+    </ng-container>
 
-  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-  <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-</table>
-</div>
-<div class="modal" *ngIf="modalAberto" >
-<div class="modal-content">
-  <span class="close" (click)="fechaModal()">&times;</span>
-
-  <h1>
-    Excluir projeto?
-  </h1>
-
-  <p>Atenção! Está ação e definitiva!</p>
-  
-  <section *matHeaderCellDef >
-    <div class="btn" *matCellDef="let element">
-      <button class="cancelar" (click)="fechaModal()">Cancelar</button>
-      <button class="delete" (click)="deleteProject(element._id)" >Excluir</button> 
-    </div>
-  </section>
-
- 
-  
-</div>
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+    <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+  </table>
 </div>
