+++ Aulas-Angular/app-modules/src/app/modules/shared/components/modal-delete/modal-delete.component.html
@@ -1,1 +1,21 @@
-<p>modal-delete works!</p>
+<body>
+  <h2>{{ title }}</h2>
+
+  <div class="content">
+    <p [innerHTML]="message"></p>
+  </div>
+
+  <mat-dialog-actions>
+    <button
+      mat-button
+      class="outline-light"
+      (click)="close()"
+    >Cancelar
+    </button>
+    <button
+      mat-button
+      (click)="onDelete()"
+      >Excluir
+    </button>
+  </mat-dialog-actions>
+</body>
