+++ Aulas-Angular/app-modules/src/app/modules/shared/components/modal-delete/modal-delete.component.ts
@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import { Component, Inject } from '@angular/core'
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog'
 
 @Component({
   selector: 'app-modal-delete',
@@ -6,5 +7,22 @@ import { Component } from '@angular/core';
   styleUrl: './modal-delete.component.scss'
 })
 export class ModalDeleteComponent {
+  title: string;
+  message: string;
+  
+  constructor(
+    private dialogRef: MatDialogRef<ModalDeleteComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: any
+  ) {
+    this.title = data.title;
+    this.message = data.message;
+  }
 
+  onDelete() {
+    this.dialogRef.close(true);
+  }
+
+  close() {
+    this.dialogRef.close();
+  }
 }
